### 2022-08-03 #4639 @editorialbot
[PRE REVIEW]: PyAutoCTI: Open-Source Charge Transfer Inefficiency Calibration
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** https://github.com/Jammy2211/PyAutoCTI
**Branch with paper.md** (empty if default branch):
**Version:** 2022.05.02.1
