### 2022-08-05 #4639 @arfon
Also, is this submission associated with a paper in a AAS journal?
