### 2022-08-26 #4639 @adonath
Thanks @Jammy2211 for the submission! I recently joined the JOSS editor team and happy to take over the editing role for this submission.
