### 2022-09-15 #4639 @adonath
What is the current status of the API documentation and license file @Jammy2211 ?
