### 2022-09-21 #4639 @Jammy2211
Deadlines are finally over so sorting this out this week.

Its a bit less trivial than it should be, as we have made a key dependency pip installable so I'll update installation docs at the same time, which should make the life of the reviewers easier.
