### 2022-09-26 4214316
genralize docstring
