### 2022-09-26 113cf03
split extract object modules
