+++ autocti/extract/two_d/parallel/overscan.py
@@ -2,34 +2,24 @@ from typing import List, Tuple
 
 import autoarray as aa
 
-from autocti.extract.two_d.abstract import Extract2D
+from autocti.extract.two_d.parallel.abstract import Extract2DParallel
 
 from autocti.extract.two_d import extract_2d_util
 
 
-class Extract2DSerialOverscan(Extract2D):
-    @property
-    def binning_axis(self) -> int:
-        """
-        The axis over which binning is performed to turn a 2D serial overscan into a 1D array (which likely
-        contains the EPER of the main data).
-
-        For a serial extract `axis=1` such that binning is performed over the rows containing the FPR.
-        """
-        return 0
-
+class Extract2DParallelOverscan(Extract2DParallel):
     def region_list_from(self, pixels: Tuple[int, int]) -> List[aa.Region2D]:
         """
-        Returns a list containing the 2D serial verscan region, which is simply the parallel overscan input to the
+        Returns a list containing the 2D parallel overscan region, which is simply the parallel overscan input to the
         object, extracted between two input `pixels` indexes (this is somewhat redundant information, but mimicks
         the `Extract` object API across all other `Extract` objects).
 
         (top-row, bottom-row, left-column, right-column) = (y0, y1, x0, x1)
 
-        The serial overscan spans all columns of the image, thus the coordinates x0 and x1 do not change. y0 and y1
+        The parallel overscan spans all columns of the image, thus the coordinates x0 and x1 do not change. y0 and y1
         are updated based on the `pixels` input.
 
-        Negative pixel values are supported to the `pixels` tuple, whereby columns in front of the serial overscan are
+        Negative pixel values are supported to the `pixels` tuple, whereby rows in front of the parallel overscan are
         also extracted.
 
         The diagram below illustrates the extraction for `pixels=(0, 1)`:
@@ -37,10 +27,10 @@ class Extract2DSerialOverscan(Extract2D):
         [] = read-out electronics
         [==========] = read-out register
         [..........] = serial prescan
-        [pppppppppp] = serial overscan
+        [pppppppppp] = parallel overscan
         [ssssssssss] = serial overscan
         [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
-        [tttttttttt] = serial / serial charge injection region trail
+        [tttttttttt] = parallel / serial charge injection region trail
 
                [ppppppppppppppppppppp]
                [ppppppppppppppppppppp]
@@ -55,29 +45,29 @@ class Extract2DSerialOverscan(Extract2D):
         []     [=====================]
                <---------Ser--------
 
-        The extracted regions correspond to the serial overscan [sss] regions.
+        The extracted regions correspond to the parallel overscan [ppppppppppppppp] regions.
 
         For `pixels=(0,1)` the extracted arrays returned via the `array_2d_list_from()` function keep the first
-        serial pixels across the columns of the overscan:
+        parallel pixels across the rows of the overscan:
 
-        array_2d_list[0] = [s]
+        array_2d_list[0] = [ppppppppppppppppppppp]
 
         Parameters
         ----------
         pixels
-            The column pixel index which determines the region of the overscan (e.g. `pixels=(0, 3)` will compute the
-            region corresponding to the 1st, 2nd and 3rd overscan columns).
+            The row pixel index which determines the region of the overscan (e.g. `pixels=(0, 3)` will compute the
+            region corresponding to the 1st, 2nd and 3rd overscan rows).
         """
-        serial_overscan_extract = aa.Region2D(
+        parallel_overscan_extract = aa.Region2D(
             (
-                self.serial_overscan.y0,
-                self.serial_overscan.y1,
[…]
