### 2022-09-26 cd3064c
added method to SerialExtract which takes median over ros
