+++ autocti/extract/two_d/parallel/abstract.py
@@ -3,6 +3,7 @@ from typing import List, Tuple
 
 import autoarray as aa
 
+
 from autocti.extract.two_d.abstract import Extract2D
 
 
@@ -16,15 +17,76 @@ class Extract2DParallel(Extract2D):
         """
         return 1
 
-    def median_lists_of_individual_regions_from(
+    def median_list_from(
+        self, array: aa.Array2D, pixels: Tuple[int, int]
+    ) -> List[float]:
+        """
+        Returns the median values of the `Extract2D` object's corresponding region, where the median is taken over
+        all columns of the region(s).
+
+        To describe this function we will use the example of estimating the median of the charge lines in charge
+        injection data, by taking the median over every individual column of charge injection (e.g. aligned with the
+        FPR).
+
+        The inner regions of the FPR of each charge injection line informs us of the injected level of charge for
+        that injection.
+
+        By taking the median of values of the charge injection regions (after accounting for those which have had
+        electrons captured and relocated due to CTI), we can therefore estimate the input charge injection
+        normalization.
+
+        This function does this for every column of every charge injection. If multiple charge injections are performed
+        per column, the median of all charge regions is taken.
+
+        For example, if there are 3 charge injection regions, this function returns a list where each value
+        estimates the charge injection normalization of a given charge injection region. The size of this list
+        is therefore the number of charge injection columns.
+
+        The function `median_list_of_lists_from` performs the median over each individual charge injection
+        region in each column. It therefore estimates multiple injection normalizations per column. for each individual
+        charge region. Which function one uses depends on the properties of the charge injection on the instrumentation.
+
+        Parameters
+        ----------
+        array
+            The charge injection image from which the charge injection normalizations are estimated.
+        pixels
+            The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
+            FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
+            as `pixels=(10, 30)` would be used.
+        """
+        median_list = []
+
+        arr_list = [
+            np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
+            for region in self.region_list_from(pixels=pixels)
+        ]
+
+        for column_index in range(arr_list[0].shape[1]):
+
+            for i, array_2d in enumerate(arr_list):
+
+                if i == 0:
+                    arr = array_2d[:, column_index]
+                else:
+                    arr = np.concatenate((arr[:], array_2d[:, column_index]))
+
+            median_list.append(float(np.ma.median(arr)))
+
+        return median_list
+
+    def median_list_of_lists_from(
         self, array: aa.Array2D, pixels: Tuple[int, int]
     ) -> List[List]:
         """
-        Returns the median values of the `Extract2D` object's corresponding region, for every individual region it
-        contains.
+        Returns the median values of the `Extract2D` object's corresponding region, where the median is taken over
+        all column(s) of every individual region.
 
         To describe this function we will use the example of estimating the median of the charge lines in charge
-        injection data. The inner regions of the FPR of each charge injection line informs us of the injected level of
+        injection data, by taking the median over every individual column of charge injection (e.g. aligned with the
+        FPR).
+
+        The inner regions of the FPR of each charge injection line informs us of the injected level of
         charge for that injection.
 
         By taking the median of values of the charge injection regions (after accounting for those which have had
@@ -58,69 +120,14 @@ class Extract2DParallel(Extract2D):
 
         for array_2d in arr_list:
 
-            injection_of_array_list = []
+            value_list = []
 
             for column_index in range(array_2d.shape[1]):
 
-                injection_norm = float(np.ma.median(array_2d[:, column_index]))
[…]
