### 2022-09-27 ef6c432
updated config for new arctic
