### 2022-09-27 f090730
remove flip for euclid hack
