### 2022-09-28 cee714b
update tests for new arctic master
