### 2022-09-30 e41e7f4
parse pixel line collection
