+++ test_autocti/clocker/test_dict.py
@@ -10,7 +10,7 @@ def make_clocker_dict():
         "type": "autocti.clocker.two_d.Clocker2D",
         "iterations": 5,
         "parallel_roe": {
-            "type": "arcticpy.roe.ROE",
+            "type": "arcticpytest.roe.ROE",
             "dwell_times": {
                 "type": "numpy.ndarray",
                 "array": [1.0],
@@ -28,7 +28,7 @@ def make_clocker_dict():
         "parallel_poisson_traps": False,
         "parallel_fast_mode": False,
         "serial_roe": {
-            "type": "arcticpy.roe.ROE",
+            "type": "arcticpytest.roe.ROE",
             "dwell_times": {
                 "type": "numpy.ndarray",
                 "array": [1.0],
