+++ test_autocti/clocker/test_one_d.py
@@ -3,7 +3,7 @@ import os
 import pytest
 import numpy as np
 
-from arcticpy import add_cti
+from arcticpytest import add_cti
 
 import autocti as ac
 
