+++ .github/workflows/main.yml
@@ -62,7 +62,7 @@ jobs:
         pip3 install setuptools
         pip3 install wheel
         pip3 install pytest
-        python3 -m pip install --index-url https://test.pypi.org/simple --extra-index-url https://pypi.org/simple/ arcticpy
+        python3 -m pip install --index-url https://test.pypi.org/simple --extra-index-url https://pypi.org/simple/ arcticpytest
         pip3 install -r PyAutoConf/requirements.txt
         pip3 install -r PyAutoFit/requirements.txt
         pip3 install -r PyAutoArray/requirements.txt
