+++ .github/workflows/main.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]  # 3.9 not supported on numba as of 2020-02-11
+        python-version: [3.8, 3.9, 3.10]
     steps:
     - name: Checkout PyAutoConf
       uses: actions/checkout@v2
