+++ autocti/clocker/abstract.py
@@ -1,8 +1,8 @@
 from typing import List, Optional
 
-from arcticpytest import CCD
-from arcticpytest import CCDPhase
-from arcticpytest import TrapInstantCapture
+from arcticpy import CCD
+from arcticpy import CCDPhase
+from arcticpy import TrapInstantCapture
 
 from autoconf.dictable import Dictable
 
