+++ requirements.txt
@@ -1,0 +1,1 @@
+arcticpytest==1.2
