+++ readthedocs.yaml
@@ -1,7 +1,7 @@
 # .readthedocs.yml
 
 python:
-  version: 3.7
+  version: 3.10
   setup_py_install: true
 
 # Build documentation in the docs/ directory with Sphinx
