+++ readthedocs.yaml
@@ -1,7 +1,7 @@
 # .readthedocs.yml
 
 python:
-  version: 3.9
+  version: 3.8
   setup_py_install: true
 
 # Build documentation in the docs/ directory with Sphinx
