### 2022-10-05 d3197a3
remove bad dependencies readthedocs reque
