### 2022-10-10 55e8843
updated to use extracted pixel line
