+++ docs/requirements.txt
@@ -1,7 +1,3 @@
-setuptools_scm==7.0.5
-setuptools >=62
-numpy==1.21
-cython==0.29
 arcticpytest==1.2
 autoconf==2022.07.11.1
 autofit==2022.07.11.1
