+++ readthedocs.yaml
@@ -1,19 +1,18 @@
 # .readthedocs.yml
 
-python:
-  version: 3.8
-  setup_py_install: true
-
-# Build documentation in the docs/ directory with Sphinx
-sphinx:
-  configuration: docs/conf.py
+version: 2
 
 build:
-   image: latest
    apt_packages:
       - libgsl-dev
 
+sphinx:
+  configuration: docs/conf.py
+
+python:
+  version: 3.8
+  install:
+  - requirements: docs/requirements.txt
+
 formats:
   []
-
-requirements_file: docs/requirements.txt
