+++ readthedocs.yaml
@@ -1,18 +1,15 @@
-# .readthedocs.yml
-
 version: 2
 
 build:
-   apt_packages:
-      - libgsl-dev
-
-sphinx:
-  configuration: docs/conf.py
+  os: ubuntu-20.04
+  tools:
+    python: "3.10"
+ apt_packages:
+    - libgsl-dev
 
 python:
-  version: 3.8
-  install:
-  - requirements: docs/requirements.txt
+   install:
+   - requirements: docs/requirements.txt
 
-formats:
-  []
+sphinx:
+   configuration: docs/conf.py
