+++ readthedocs.yaml
@@ -4,12 +4,12 @@ build:
   os: ubuntu-20.04
   tools:
     python: "3.10"
- apt_packages:
+  apt_packages:
     - libgsl-dev
 
 python:
-   install:
-   - requirements: docs/requirements.txt
+  install:
+  - requirements: docs/requirements.txt
 
 sphinx:
-   configuration: docs/conf.py
+  configuration: docs/conf.py
