+++ docs/conf.py
@@ -19,7 +19,7 @@ workspace_path = str(here())
 
 import os
 import sys
-import autocti
+# import autocti
 
 sys.path.insert(0, os.path.abspath("."))
 
