+++ docs/conf.py
@@ -19,7 +19,6 @@ workspace_path = str(here())
 
 import os
 import sys
-# import autocti
 
 sys.path.insert(0, os.path.abspath("."))
 
@@ -31,7 +30,7 @@ copyright = "2022, James Nightingale, Richard Hayes"
 author = "James Nightingale, Richard Hayes"
 
 # The full version, including alpha/beta/rc tags
-release = "2022.05.02.1"
+release = "2022.07.11.1"
 master_doc = "index"
 
 
@@ -92,8 +91,8 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 # -- Options for HTML output -------------------------------------------------
 
 html_theme = "furo"
-html_title = "PyAutoCTI"
-html_short_title = "PyAutoCTI"
+html_title = "PyAutoLens"
+html_short_title = "PyAutoLens"
 html_permalinks_icon = "<span>#</span>"
 html_last_updated_fmt = "%b %d, %Y"
 
@@ -101,21 +100,35 @@ html_show_sourcelink = False
 html_show_sphinx = True
 html_show_copyright = True
 
-# pygments_style = "sphinx"
-# pygments_dark_style = "monokai"
+pygments_style = "sphinx"
+pygments_dark_style = "monokai"
 add_function_parentheses = False
 
 html_context = {
     "menu_links_name": "Repository",
     "doc_path": "docs",
-    "github_project": "pyautocti",
-    "github_repo": "pyautocti",
+    "github_project": "pyautolens",
+    "github_repo": "pyautolens",
     "github_version": "master",
 }
 language = "en"
 
 html_static_path = ["_static"]
 html_css_files = ["pied-piper-admonition.css"]
+
+html_theme_options = {
+    "light_css_variables": {
+        "color-brand-primary": "#7C4DFF",
+        "color-brand-content": "#7C4DFF",
+    }
+}
+
+if "READTHEDOCS" in os.environ:
+    html_theme_options["announcement"] = (
+        "This documentation is hosted on Read the Docs only for testing. Please use "
+        "<a href='https://pradyunsg.me/furo/'>the main documentation</a> instead."
+    )
+
 
 from sphinx.builders.html import StandaloneHTMLBuilder
 
