+++ docs/requirements.txt
@@ -1,7 +1,5 @@
 arcticpytest==1.2
 autoconf==2022.07.11.1
-autofit==2022.07.11.1
-autoarray==2022.07.11.1
 numpydoc>=1.0.0
 pyprojroot==0.2.0
 sphinx==4.0.0
