+++ docs/requirements.txt
@@ -2,7 +2,7 @@ arcticpytest==1.2
 autoconf==2022.07.11.1
 numpydoc>=1.0.0
 pyprojroot==0.2.0
-sphinx==4.0.0
+sphinx==5.2.3
 furo
 myst-parser
 sphinx_copybutton
