### 2022-10-10 8d466d9
added autofit plottrs to plot
