+++ autocti/plot/__init__.py
@@ -1,6 +1,9 @@
-from pkgutil import extend_path
-
-__path__ = extend_path(__path__, __name__)
+from autofit.plot.samples_plotters import SamplesPlotter
+from autofit.non_linear.nest.dynesty.plotter import DynestyPlotter
+from autofit.non_linear.nest.ultranest.plotter import UltraNestPlotter
+from autofit.non_linear.mcmc.emcee.plotter import EmceePlotter
+from autofit.non_linear.mcmc.zeus.plotter import ZeusPlotter
+from autofit.non_linear.optimize.pyswarms.plotter import PySwarmsPlotter
 
 from autoarray.plot.wrap.wrap_base import Axis
 from autoarray.plot.wrap.wrap_base import Units
