### 2022-10-10 #4639 @Jammy2211
Okay, things are finally up:

https://pyautocti.readthedocs.io/en/latest/api/api.html

https://github.com/Jammy2211/PyAutoCTI/blob/master/LICENSE

Sorry that took so long!
