+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-arcticpytest==1.2
+pyarctic==1.0
 autoconf==2022.07.11.1
 numpydoc>=1.0.0
 pyprojroot==0.2.0
