### 2022-10-11 e3852cf
median list uses stack
