+++ autocti/extract/two_d/parallel/calibration.py
@@ -36,48 +36,48 @@ class Extract2DParallelCalibration:
 
     def extraction_region_from(self, columns: Tuple[int, int]) -> aa.type.Region2DLike:
         """
-        Extract a parallel calibration array from an input array, where this array contains a sub-set of the input
-        array which is specifically used for only parallel CTI calibration. This array is simply a specified number
-        of columns that are closest to the read-out electronics.
+         Extract a parallel calibration array from an input array, where this array contains a sub-set of the input
+         array which is specifically used for only parallel CTI calibration. This array is simply a specified number
+         of columns that are closest to the read-out electronics.
 
-        The diagram below illustrates the arrays that is extracted from a array with columns=(0, 3):
+         The diagram below illustrates the arrays that is extracted from a array with columns=(0, 3):
 
-        [] = read-out electronics
-        [==========] = read-out register
-        [..........] = serial prescan
-        [pppppppppp] = parallel overscan
-        [ssssssssss] = serial overscan
-        [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
-        [tttttttttt] = parallel / serial charge injection region trail
+         [] = read-out electronics
+         [==========] = read-out register
+         [..........] = serial prescan
+         [pppppppppp] = parallel overscan
+         [ssssssssss] = serial overscan
+         [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
+         [tttttttttt] = parallel / serial charge injection region trail
 
-               [ptptptptptptptptptptp]
-               [tptptptptptptptptptpt]
-          [...][xxxxxxxxxxxxxxxxxxxxx][sss]
-          [...][ccccccccccccccccccccc][sss]
-        | [...][ccccccccccccccccccccc][sts]    |
-        | [...][xxxxxxxxxxxxxxxxxxxxx][sss]    | Direction
-        P [...][xxxxxxxxxxxxxxxxxxxxx][sss]    | of
-        | [...][ccccccccccccccccccccc][sss]    | clocking
-        \/[...][ccccccccccccccccccccc][sss]   \/
+                [ptptptptptptptptptptp]
+                [tptptptptptptptptptpt]
+           [...][xxxxxxxxxxxxxxxxxxxxx][sss]
+           [...][ccccccccccccccccccccc][sss]
+         | [...][ccccccccccccccccccccc][sts]    |
+         | [...][xxxxxxxxxxxxxxxxxxxxx][sss]    | Direction
+         P [...][xxxxxxxxxxxxxxxxxxxxx][sss]    | of
+         | [...][ccccccccccccccccccccc][sss]    | clocking
+         \/[...][ccccccccccccccccccccc][sss]   \/
 
-        []     [=====================]
-               <--------Ser---------
+         []     [=====================]
+                <--------Ser---------
 
-        The extracted array keeps just the trails following all charge injection scans and replaces all other
-        values with 0s:
+         The extracted array keeps just the trails following all charge injection scans and replaces all other
+         values with 0s:
 
-               [ptp]
-               [tpt]
-               [xxx]
-               [ccc]
-        |      [ccc]                           |
-        |      [xxx]                           | Direction
-       Par     [xxx]                           | of
-        |      [ccc]                           | clocking
-               [ccc]                           |
+                [ptp]
+                [tpt]
+                [xxx]
+                [ccc]
+         |      [ccc]                           |
+         |      [xxx]                           | Direction
+        Par     [xxx]                           | of
+         |      [ccc]                           | clocking
+                [ccc]                           |
 
-        []     [=====================]
-               <--------Ser---------
+         []     [=====================]
+                <--------Ser---------
         """
[…]
