+++ autocti/extract/two_d/serial/eper.py
@@ -10,50 +10,50 @@ from autocti.extract.two_d import extract_2d_util
 class Extract2DSerialEPER(Extract2DSerial):
     def region_list_from(self, pixels: Tuple[int, int]):
         """
-        Returns a list of the 2D serial EPER regions from the `region_list` containing signal  (e.g. the charge
-        injection regions of charge injection data), extracted between two input `pixels` indexes.
+         Returns a list of the 2D serial EPER regions from the `region_list` containing signal  (e.g. the charge
+         injection regions of charge injection data), extracted between two input `pixels` indexes.
 
-        Negative pixel values are supported to the `pixels` tuple, whereby columns in front of the serial EPERs
-        (e.g. the FPRs) are also extracted.
+         Negative pixel values are supported to the `pixels` tuple, whereby columns in front of the serial EPERs
+         (e.g. the FPRs) are also extracted.
 
-        The diagram below illustrates the extraction for `pixels=(0, 1)`:
+         The diagram below illustrates the extraction for `pixels=(0, 1)`:
 
-        [] = read-out electronics
-        [==========] = read-out register
-        [..........] = serial prescan
-        [pppppppppp] = parallel overscan
-        [ssssssssss] = serial overscan
-        [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
-        [tttttttttt] = parallel / serial charge injection region trail
+         [] = read-out electronics
+         [==========] = read-out register
+         [..........] = serial prescan
+         [pppppppppp] = parallel overscan
+         [ssssssssss] = serial overscan
+         [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
+         [tttttttttt] = parallel / serial charge injection region trail
 
-               [ppppppppppppppppppppp]
-               [ppppppppppppppppppppp]
-           [...][ttttttttttttttttttttt][sss]
-           [...][c1c1cc1c1cc1cc1ccc1cc][st1]
-        |  [...][1c1c1cc1c1cc1ccc1cc1c][ts0]    |
-        |  [...][ttttttttttttttttttttt][sss]    | Direction
-       Par [...][ttttttttttttttttttttt][sss]    | of
-        |  [...][0ccc0cccc0cccc0cccc0c][st1]    | clocking
-       \/  [...][cc0ccc0cccc0cccc0cccc][ts0]   \/
+                [ppppppppppppppppppppp]
+                [ppppppppppppppppppppp]
+            [...][ttttttttttttttttttttt][sss]
+            [...][c1c1cc1c1cc1cc1ccc1cc][st1]
+         |  [...][1c1c1cc1c1cc1ccc1cc1c][ts0]    |
+         |  [...][ttttttttttttttttttttt][sss]    | Direction
+        Par [...][ttttttttttttttttttttt][sss]    | of
+         |  [...][0ccc0cccc0cccc0cccc0c][st1]    | clocking
+        \/  [...][cc0ccc0cccc0cccc0cccc][ts0]   \/
 
-        []     [=====================]
-               <---------Ser--------
+         []     [=====================]
+                <---------Ser--------
 
-        The extracted regions correspond to the first serial EPER all charge injection regions:
+         The extracted regions correspond to the first serial EPER all charge injection regions:
 
-        region_list[0] = [0, 2, 22, 225 (serial prescan is 3 pixels)
-        region_list[1] = [4, 6, 22, 25] (serial prescan is 3 pixels)
+         region_list[0] = [0, 2, 22, 225 (serial prescan is 3 pixels)
+         region_list[1] = [4, 6, 22, 25] (serial prescan is 3 pixels)
 
-        For `pixels=(0,1)` the extracted arrays returned via the `array_2d_list_from()` function keep the first
[…]
