### 2022-10-18 bc32ed8
reduce individual noise scaling makes in ResuiltImagingCI
