+++ autocti/charge_injection/model/result.py
@@ -16,35 +16,11 @@ class ResultImagingCI(ResultDataset):
         )
 
     @property
-    def noise_scaling_map_of_regions_ci(self):
-        return (
-            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_regions_ci
-        )
-
-    @property
-    def noise_scaling_map_of_parallel_epers(self):
-        return (
-            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_parallel_epers
-        )
-
-    @property
-    def noise_scaling_map_of_serial_epers(self):
-        return (
-            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_serial_epers
-        )
-
-    @property
-    def noise_scaling_map_of_serial_overscan_no_trails(self):
-        return (
-            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_serial_overscan_no_trails
-        )
-
-    @property
     def noise_scaling_map_list(self):
 
         return [
-            self.noise_scaling_map_of_regions_ci,
-            self.noise_scaling_map_of_parallel_epers,
-            self.noise_scaling_map_of_serial_epers,
-            self.noise_scaling_map_of_serial_overscan_no_trails,
+            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_regions_ci,
+            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_parallel_epers,
+            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_serial_epers,
+            self.max_log_likelihood_full_fit_no_hyper_scaling.chi_squared_map_of_serial_overscan_no_trails,
         ]
