### 2022-10-18 9019876
simplify hyper nosie list in imagingci
