### 2022-10-18 96f958c
nosie scaling map list now dict
