+++ test_autocti/charge_injection/model/test_analysis_ci.py
@@ -198,12 +198,12 @@ def test__extracted_fits_from_instance_and_imaging_ci__include_noise_scaling(
         ),
     )
 
-    noise_scaling_map_list_list_of_regions_ci = [
+    noise_scaling_map_dict_list_of_regions_ci = [
         ac.Array2D.ones(shape_native=(7, 7), pixel_scales=1.0)
     ]
 
-    imaging_ci_7x7.noise_scaling_map_list = [
-        noise_scaling_map_list_list_of_regions_ci[0]
+    imaging_ci_7x7.noise_scaling_map_dict = [
+        noise_scaling_map_dict_list_of_regions_ci[0]
     ]
 
     masked_imaging_ci = imaging_ci_7x7.apply_mask(mask=mask_2d_7x7_unmasked)
