+++ autocti/charge_injection/fit.py
@@ -76,7 +76,7 @@ class FitImagingCI(aa.FitImaging):
         return self.layout.extract.serial_eper.array_2d_from(array=self.chi_squared_map)
 
     @property
-    def chi_squared_map_of_serial_overscan_no_trails(self):
+    def chi_squared_map_of_serial_overscan_no_epers(self):
         return self.layout.extract.serial_overscan_above_epers_array_2d_from(
             array=self.chi_squared_map
         )
