+++ autocti/charge_injection/hyper.py
@@ -24,10 +24,10 @@ class HyperCINoiseCollection:
         regions_ci: Optional[HyperCINoiseScalar] = None,
         parallel_epers: Optional[HyperCINoiseScalar] = None,
         serial_eper: Optional[HyperCINoiseScalar] = None,
-        serial_overscan_no_trails: Optional[HyperCINoiseScalar] = None,
+        serial_overscan_no_epers: Optional[HyperCINoiseScalar] = None,
     ):
 
         self.regions_ci = regions_ci
         self.parallel_epers = parallel_epers
         self.serial_eper = serial_eper
-        self.serial_overscan_no_trails = serial_overscan_no_trails
+        self.serial_overscan_no_epers = serial_overscan_no_epers
