### 2022-10-18 036972c
as_dict method added to hyper collection
