+++ autocti/charge_injection/hyper.py
@@ -31,3 +31,17 @@ class HyperCINoiseCollection:
         self.parallel_epers = parallel_epers
         self.serial_eper = serial_eper
         self.serial_overscan_no_epers = serial_overscan_no_epers
+
+    @property
+    def as_dict(self):
+
+        hyper_dict = {
+            "regions_ci" : self.regions_ci,
+            "parallel_epers" : self.parallel_epers,
+            "serial_eper" : self.serial_eper,
+            "serial_overscan_no_epers" : self.serial_overscan_no_epers
+        }
+
+        return {
+            key: value for key, value in hyper_dict.items() if value is not None
+        }
