+++ autocti/charge_injection/model/result.py
@@ -22,7 +22,7 @@ class ResultImagingCI(ResultDataset):
 
         return {
             "regions_ci": fit.chi_squared_map_of_regions_ci,
-            "parallel_epers": fit.chi_squared_map_of_parallel_epers,
-            "serial_epers": fit.chi_squared_map_of_serial_epers,
-            "serial_overscan_no_epers": fit.chi_squared_map_of_serial_overscan_no_epers,
+            "parallel_eper": fit.chi_squared_map_of_parallel_eper,
+            "serial_eper": fit.chi_squared_map_of_serial_eper,
+            "serial_overscan_no_eper": fit.chi_squared_map_of_serial_overscan_no_eper,
         }
