### 2022-10-18 ca38901
charge injection fiut now uses dict
