+++ test_autocti/charge_injection/test_fit_ci.py
@@ -10,7 +10,7 @@ def test__fit_figure_of_merit(imaging_ci_7x7):
     fit = ac.FitImagingCI(
         dataset=imaging_ci_7x7,
         post_cti_data=imaging_ci_7x7.pre_cti_data,
-        hyper_noise_scalar_list=None,
+        hyper_noise_scalar_dict=None,
     )
 
     assert fit.log_likelihood == pytest.approx(-575.11719997, 1e-4)
@@ -18,10 +18,14 @@ def test__fit_figure_of_merit(imaging_ci_7x7):
     hyper_noise_scalar_0 = ac.HyperCINoiseScalar(scale_factor=1.0)
     hyper_noise_scalar_1 = ac.HyperCINoiseScalar(scale_factor=2.0)
 
+    hyper_collection = ac.HyperCINoiseCollection(
+        parallel_eper=hyper_noise_scalar_0, serial_eper=hyper_noise_scalar_1
+    )
+
     fit = ac.FitImagingCI(
         dataset=imaging_ci_7x7,
         post_cti_data=imaging_ci_7x7.pre_cti_data,
-        hyper_noise_scalar_list=[hyper_noise_scalar_0, hyper_noise_scalar_1],
+        hyper_noise_scalar_dict=hyper_collection.as_dict,
     )
 
     assert fit.log_likelihood == pytest.approx(-180.877585, 1.0e-4)
@@ -29,13 +33,16 @@ def test__fit_figure_of_merit(imaging_ci_7x7):
 
 def test__hyper_noise_map_from():
     noise_map = ac.Array2D.full(fill_value=2.0, shape_native=(2, 2), pixel_scales=1.0)
-    noise_scaling_map_dict = [
-        ac.Array2D.manual(array=[[0.0, 0.0], [0.0, 0.0]], pixel_scales=1.0)
-    ]
-    hyper_noise_scalar_list = [ac.HyperCINoiseScalar(scale_factor=1.0)]
+    noise_scaling_map_dict = {
+        "parallel_eper": ac.Array2D.manual(
+            array=[[0.0, 0.0], [0.0, 0.0]], pixel_scales=1.0
+        )
+    }
+
+    hyper_noise_scalar_dict = {"parallel_eper": ac.HyperCINoiseScalar(scale_factor=1.0)}
 
     noise_map = hyper_noise_map_from(
-        hyper_noise_scalar_list=hyper_noise_scalar_list,
+        hyper_noise_scalar_dict=hyper_noise_scalar_dict,
         noise_map=noise_map,
         noise_scaling_map_dict=noise_scaling_map_dict,
     )
@@ -43,13 +50,15 @@ def test__hyper_noise_map_from():
     assert (noise_map.native == (np.array([[2.0, 2.0], [2.0, 2.0]]))).all()
 
     noise_map = ac.Array2D.full(fill_value=2.0, shape_native=(2, 2), pixel_scales=1.0)
-    noise_scaling_map_dict = [
-        ac.Array2D.manual(array=[[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
-    ]
-    hyper_noise_scalar_list = [ac.HyperCINoiseScalar(scale_factor=0.0)]
+    noise_scaling_map_dict = {
+        "parallel_eper": ac.Array2D.manual(
+            array=[[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0
+        )
+    }
+    hyper_noise_scalar_dict = {"parallel_eper": ac.HyperCINoiseScalar(scale_factor=0.0)}
 
     noise_map = hyper_noise_map_from(
-        hyper_noise_scalar_list=hyper_noise_scalar_list,
[…]
