### 2022-10-18 849e082
hyper quantities now use dictionaries
