+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-pyarctic==1.0
+arcticpy==2.01.0
 autoconf==2022.07.11.1
 numpydoc>=1.0.0
 pyprojroot==0.2.0
