### 2022-10-19 2e2d480
updated charge injection calculations to work on TVAC data
