### 2022-10-20 acf5530
removed fit full instance method
