### 2022-10-20 3ed1e96
removed fit full from instance
