### 2022-10-20 151fdd8
moved instruments module from autoarray to autocit
