+++ autocti/instruments/euclid.py
@@ -1,0 +1,580 @@
+from typing import Dict, Tuple, Optional
+
+from autoarray.structures.header import Header
+from autoarray.structures.arrays.uniform_2d import Array2D
+from autoarray.layout.layout import Layout2D
+from autoarray.layout.region import Region2D
+
+from autoarray.layout import layout_util
+
+
+def roe_corner_from(ccd_id, quadrant_id):
+
+    row_index = ccd_id[-1]
+
+    if (row_index in "123") and (quadrant_id == "E"):
+        return (1, 0)
+    elif (row_index in "123") and (quadrant_id == "F"):
+        return (1, 1)
+    elif (row_index in "123") and (quadrant_id == "G"):
+        return (0, 1)
+    elif (row_index in "123") and (quadrant_id == "H"):
+        return (0, 0)
+    elif (row_index in "456") and (quadrant_id == "E"):
+        return (0, 1)
+    elif (row_index in "456") and (quadrant_id == "F"):
+        return (0, 0)
+    elif (row_index in "456") and (quadrant_id == "G"):
+        return (1, 0)
+    elif (row_index in "456") and (quadrant_id == "H"):
+        return (1, 1)
+
+
+class Array2DEuclid(Array2D):
+    """
+    In the Euclid FPA, the quadrant id ('E', 'F', 'G', 'H') depends on whether the CCD is located
+    on the left side (rows 1-3) or right side (rows 4-6) of the FPA:
+
+    LEFT SIDE ROWS 1-2-3
+    --------------------
+
+     <--------S-----------   ---------S----------->
+    [] [========= 2 =========] [========= 3 =========] []          |
+    /    [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  /          |
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  |         | Direction arctic
+    P   [xxxxxxxxx H xxxxxxxxx] [xxxxxxxxx G xxxxxxxxx]  P         | clocks an image
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  |         | without any rotation
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  |         | (e.g. towards row 0
+                                                                   | of the ndarrays)
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx] |          |
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx] |          |
+    P   [xxxxxxxxx E xxxxxxxxx] [xxxxxxxxx F xxxxxxxxx] P          |
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx] |          |
+        [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]            |
+
+    [] [========= 0 =========] [========= 1 =========] []
+        <---------S----------   ----------S----------->
+
+
+    RIGHT SIDE ROWS 4-5-6
+    ---------------------
+
+     <--------S-----------   ---------S----------->
+    [] [========= 2 =========] [========= 3 =========] []          |
[…]
