+++ autocti/instruments/__init__.py
@@ -1,0 +1,6 @@
+from pkgutil import extend_path
+
+__path__ = extend_path(__path__, __name__)
+
+from . import acs
+from . import euclid
