+++ autocti/instruments/acs.py
@@ -1,0 +1,757 @@
+from astropy.io import fits
+import copy
+import logging
+import numpy as np
+import os
+from os import path
+import shutil
+
+from autoarray.structures.header import Header
+from autoarray.structures.arrays.uniform_2d import Array2D
+from autoarray.layout.layout import Layout2D
+from autoarray.layout.region import Region2D
+
+from autoarray import exc
+from autoarray.structures.arrays import array_2d_util
+from autoarray.layout import layout_util
+
+logging.basicConfig()
+logger = logging.getLogger()
+logger.setLevel("INFO")
+
+
+def fits_hdu_via_quadrant_letter_from(quadrant_letter):
+
+    if quadrant_letter == "D" or quadrant_letter == "C":
+        return 1
+    elif quadrant_letter == "B" or quadrant_letter == "A":
+        return 4
+    else:
+        raise exc.ArrayException("Quadrant letter for FrameACS must be A, B, C or D.")
+
+
+def array_eps_to_counts(array_eps, bscale, bzero):
+
+    if bscale is None:
+        raise exc.ArrayException(
+            "Cannot convert a Frame2D to units COUNTS without a bscale attribute (bscale = None)."
+        )
+
+    return (array_eps - bzero) / bscale
+
+
+class Array2DACS(Array2D):
+    """
+    An ACS array consists of four quadrants ('A', 'B', 'C', 'D') which have the following layout (which are described
+    at the following STScI 
+    link https://github.com/spacetelescope/hstcal/blob/master/pkg/acs/calacs/acscte/dopcte-gen2.c#L418).
+
+       <--------S-----------   ---------S----------->
+    [] [========= 2 =========] [========= 3 =========] []          /\
+    /    [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  /        |
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  |         | Direction arctic
+    P   [xxxxxxxxx B/C xxxxxxx] [xxxxxxxxx A/D xxxxxxx]  P         | clocks an image
+    |   [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  |         | without any rotation
+    \/  [xxxxxxxxxxxxxxxxxxxxx] [xxxxxxxxxxxxxxxxxxxxx]  \/        | (e.g. towards row 0
+                                                                   | of the NumPy arrays)
+
+    For a ACS .fits file:
+
+    - The images contained in hdu 1 correspond to quadrants B (left) and A (right).
+    - The images contained in hdu 4 correspond to quadrants C (left) and D (right).
+    """
+
+    @classmethod
+    def from_fits(cls, file_path, quadrant_letter):
+        """
+        Use the input .fits file and quadrant letter to extract the quadrant from the full CCD, perform
+        the rotations required to give correct arctic clocking and convert the image from units of COUNTS / CPS to
+        ELECTRONS.
+
+        See the docstring of the `FrameACS` class for a complete description of the HST FPA, quadrants and
+        rotations.
+
+        Also see https://github.com/spacetelescope/hstcal/blob/master/pkg/acs/calacs/acscte/dopcte-gen2.c#L418
+        """
+
+        hdu = fits_hdu_via_quadrant_letter_from(quadrant_letter=quadrant_letter)
+
+        array = array_2d_util.numpy_array_2d_via_fits_from(file_path=file_path, hdu=hdu)
+
+        return cls.from_ccd(array_electrons=array, quadrant_letter=quadrant_letter)
+
+    @classmethod
+    def from_ccd(
+        cls,
+        array_electrons,
+        quadrant_letter,
+        header=None,
[…]

[…]
