### 2022-10-21 898bc00
documented quadrant loading for Array2DEuclid
