### 2022-10-24 1d9eb92
fixes to test_ci_util.py
