### 2022-10-24 b908606
tests fixed in instrument module
