+++ autocti/instruments/acs/__init__.py
@@ -2,3 +2,4 @@ from .array_2d import Array2DACS
 from .image import ImageACS
 from .layout import Layout2DACS
 from .header import HeaderACS
+from . import acs_util
