### 2022-10-24 69e7c5c
TVAC values for upper CCD test
