### 2022-11-02 #4904 @editorialbot
[REVIEW]: PyAutoCTI: Open-Source Charge Transfer Inefficiency Calibration
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** https://github.com/Jammy2211/PyAutoCTI
**Branch with paper.md** (empty if default branch):
**Version:** 2024.5.16.0
