+++ test_autocti/charge_injection/imaging/test_imaging_ci.py
@@ -16,11 +16,12 @@ def test__norm_columns_list():
     image = ac.Array2D.full(
         fill_value=1.0, shape_native=(5, 5), pixel_scales=(1.0, 1.0)
     )
+    noise_map = ac.Array2D.ones(shape_native=image.shape_native, pixel_scales=image.pixel_scales)
 
     layout = ac.Layout2DCI(shape_2d=image.shape_native, region_list=[(1, 4, 1, 4)])
 
     imaging = ac.ImagingCI(
-        image=image, noise_map=image, pre_cti_data=image, layout=layout
+        image=image, noise_map=noise_map, pre_cti_data=image, layout=layout
     )
 
     assert imaging.norm_columns_list == [1.0, 1.0, 1.0]
@@ -41,7 +42,7 @@ def test__norm_columns_list():
     layout = ac.Layout2DCI(shape_2d=image.shape_native, region_list=[(1, 4, 1, 4)])
 
     imaging = ac.ImagingCI(
-        image=image, noise_map=image, pre_cti_data=image, layout=layout
+        image=image, noise_map=noise_map, pre_cti_data=image, layout=layout
     )
 
     assert imaging.norm_columns_list == [2.0, 5.0, 8.0]
@@ -49,7 +50,7 @@ def test__norm_columns_list():
     layout = ac.Layout2DCI(shape_2d=image.shape_native, region_list=[(2, 4, 1, 4)])
 
     imaging = ac.ImagingCI(
-        image=image, noise_map=image, pre_cti_data=image, layout=layout
+        image=image,  noise_map=noise_map, pre_cti_data=image, layout=layout
     )
 
     assert imaging.norm_columns_list == [2.5, 5.5, 8.5]
@@ -70,7 +71,7 @@ def test__norm_columns_list():
     image = image.apply_mask(mask=mask)
 
     imaging = ac.ImagingCI(
-        image=image, noise_map=image, pre_cti_data=image, layout=layout
+        image=image, noise_map=noise_map, pre_cti_data=image, layout=layout
     )
 
     assert imaging.norm_columns_list == [2.0, 6.0, 8.5]
