+++ autocti/charge_injection/plot/imaging_ci_plotters.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 import autoarray.plot as aplt
 
-from autoarray.plot.mat_wrap.mat_plot import AutoLabels
+from autoarray.plot.auto_labels import AutoLabels
 from autoarray.dataset.plot.imaging_plotters import ImagingPlotterMeta
 
 from autocti.plot.abstract_plotters import Plotter
