+++ autocti/plot/__init__.py
@@ -5,44 +5,44 @@ from autofit.non_linear.mcmc.emcee.plotter import EmceePlotter
 from autofit.non_linear.mcmc.zeus.plotter import ZeusPlotter
 from autofit.non_linear.optimize.pyswarms.plotter import PySwarmsPlotter
 
-from autoarray.plot.wrap.wrap_base import Axis
-from autoarray.plot.wrap.wrap_base import Units
-from autoarray.plot.wrap.wrap_base import Figure
-from autoarray.plot.wrap.wrap_base import Cmap
-from autoarray.plot.wrap.wrap_base import Colorbar
-from autoarray.plot.wrap.wrap_base import ColorbarTickParams
-from autoarray.plot.wrap.wrap_base import TickParams
-from autoarray.plot.wrap.wrap_base import YTicks
-from autoarray.plot.wrap.wrap_base import XTicks
-from autoarray.plot.wrap.wrap_base import Title
-from autoarray.plot.wrap.wrap_base import YLabel
-from autoarray.plot.wrap.wrap_base import XLabel
-from autoarray.plot.wrap.wrap_base import Legend
-from autoarray.plot.wrap.wrap_base import Output
+from autoarray.plot.wrap.base import Axis
+from autoarray.plot.wrap.base import Units
+from autoarray.plot.wrap.base import Figure
+from autoarray.plot.wrap.base import Cmap
+from autoarray.plot.wrap.base import Colorbar
+from autoarray.plot.wrap.base import ColorbarTickParams
+from autoarray.plot.wrap.base import TickParams
+from autoarray.plot.wrap.base import YTicks
+from autoarray.plot.wrap.base import XTicks
+from autoarray.plot.wrap.base import Title
+from autoarray.plot.wrap.base import YLabel
+from autoarray.plot.wrap.base import XLabel
+from autoarray.plot.wrap.base import Legend
+from autoarray.plot.wrap.base import Output
 
-from autoarray.plot.wrap.wrap_1d import YXPlot
-from autoarray.plot.wrap.wrap_2d import ArrayOverlay
-from autoarray.plot.wrap.wrap_2d import GridScatter
-from autoarray.plot.wrap.wrap_2d import GridPlot
-from autoarray.plot.wrap.wrap_2d import VectorYXQuiver
-from autoarray.plot.wrap.wrap_2d import PatchOverlay
-from autoarray.plot.wrap.wrap_2d import VoronoiDrawer
-from autoarray.plot.wrap.wrap_2d import OriginScatter
-from autoarray.plot.wrap.wrap_2d import MaskScatter
-from autoarray.plot.wrap.wrap_2d import BorderScatter
-from autoarray.plot.wrap.wrap_2d import PositionsScatter
-from autoarray.plot.wrap.wrap_2d import IndexScatter
-from autoarray.plot.wrap.wrap_2d import MeshGridScatter
-from autoarray.plot.wrap.wrap_2d import ParallelOverscanPlot
-from autoarray.plot.wrap.wrap_2d import SerialPrescanPlot
-from autoarray.plot.wrap.wrap_2d import SerialOverscanPlot
+from autoarray.plot.wrap.one_d import YXPlot
+from autoarray.plot.wrap.two_d import ArrayOverlay
+from autoarray.plot.wrap.two_d import GridScatter
+from autoarray.plot.wrap.two_d import GridPlot
+from autoarray.plot.wrap.two_d import VectorYXQuiver
+from autoarray.plot.wrap.two_d import PatchOverlay
+from autoarray.plot.wrap.two_d import VoronoiDrawer
+from autoarray.plot.wrap.two_d import OriginScatter
+from autoarray.plot.wrap.two_d import MaskScatter
[…]
