### 2022-11-11 e9f99af
serial extract now includes mask
