+++ test_autocti/extract/two_d/parallel/test_parallel_calibration.py
@@ -1,4 +1,3 @@
-import copy
 import numpy as np
 import autocti as ac
 
