+++ test_autocti/config/grids/radial_minimum.ini
@@ -23,7 +23,7 @@ EllipticalBrokenPowerLaw=1e-8
 SphericalBrokenPowerLaw=1e-8
 EllipticalCoredPowerLaw=1e-8
 SphericalCoredPowerLaw=1e-8
-EllPowerLaw=1e-8
+PowerLaw=1e-8
 SphericalPowerLaw=1e-8
 EllipticalCoredIsothermal=1e-8
 SphericalCoredIsothermal=1e-8
