+++ autocti/dataset_1d/dataset_1d/settings.py
@@ -1,5 +1,6 @@
 import autoarray as aa
 
+
 class SettingsDataset1D(aa.AbstractSettingsDataset):
 
     pass
