+++ autocti/charge_injection/model/analysis.py
@@ -1,6 +1,6 @@
 from typing import Optional, List
 
-from autofit.non_linear.samples import PDFSamples
+from autofit.non_linear.samples import SamplesPDF
 from autofit.mapper.prior_model.collection import CollectionPriorModel
 from autofit.mapper.model import ModelInstance
 from autofit.non_linear.abstract_search import Analysis
@@ -169,7 +169,7 @@ class AnalysisImagingCI(Analysis):
 
     def make_result(
         self,
-        samples: PDFSamples,
+        samples: SamplesPDF,
         model: CollectionPriorModel,
         sigma=1.0,
         use_errors=True,
