+++ autocti/__init__.py
@@ -65,4 +65,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2022.07.11.1"
+__version__ = "2022.7.11.1"
