+++ docs/installation/pip.rst
@@ -25,13 +25,13 @@ the installation has clean dependencies):
 
 .. code-block:: bash
 
-    pip install autocti==2022.07.11.1
+    pip install autocti==2022.7.11.1
 
 You may get warnings which state something like:
 
 .. code-block:: bash
 
-    ERROR: autoarray 2022.07.11.1 has requirement numpy<=1.22.1, but you'll have numpy 1.22.2 which is incompatible.
+    ERROR: autoarray 2022.7.11.1 has requirement numpy<=1.22.1, but you'll have numpy 1.22.2 which is incompatible.
     ERROR: numba 0.53.1 has requirement llvmlite<0.37,>=0.36.0rc1, but you'll have llvmlite 0.38.0 which is incompatible.
 
 If you see these messages, they do not mean that the installation has failed and the instructions below will
