+++ optional_requirements.txt
@@ -1,5 +1,4 @@
 numba
-llvmlite
 pylops>=1.10.0,<=1.18.3
 pynufft==2020.2.7
 ultranest==3.2.0
