+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 arcticpy==2.01.0
-autoconf==2022.07.11.1
+autoconf==2022.7.11.1
 numpydoc>=1.0.0
 pyprojroot==0.2.0
 sphinx==5.2.3
