+++ test_autocti/config/non_linear/nest/DynestyDynamic.ini
@@ -37,4 +37,5 @@ use_widths=True
 
 [parallel]
 number_of_cores=1
+force_x1_cpu=False
 
