+++ .github/workflows/main.yml
@@ -52,6 +52,7 @@ jobs:
             echo "Branch $BRANCH did not exist in $PACKAGE"
           else
             echo "Branch $BRANCH did exist in $PACKAGE"
+            git fetch
             git checkout $BRANCH
           fi
           popd
