+++ autocti/config/visualize/general.ini
@@ -1,11 +1,0 @@
-[general]
-backend=default
-zoom_around_mask=False
-imshow_origin=lower
-
-[units]
-in_kpc=False
-use_scaled=False
-
-[inversion]
-reconstruction_vmax_factor=0.5
