+++ autocti/config/general.ini
@@ -1,23 +1,0 @@
-[output]
-log_to_file=False
-log_file=output.log
-log_level=INFO
-samples_to_csv=False
-model_results_decimal_places=3
-info_whitespace_length=80
-remove_files=False
-force_pickle_overwrite=False
-identifier_version=4
-
-[fits]
-flip_for_ds9=False
-
-[hpc]
-hpc_mode=False
-iterations_per_update=5000
-
-[model]
-ignore_prior_limits=False
-
-[structures]
-use_dataset_grids=False
