+++ autocti/config/notation/label_format.ini
@@ -1,10 +1,0 @@
-[format]
-density={:.2f}
-release_timescale={:.2f}
-release_timescale_sigma={:.2f}
-full_well_depth={:.2f}
-well_fill_alpha={:.2f}
-well_fill_power={:.2f}
-well_fill_gamma={:.2f}
-well_notch_depth={:.2f}
-scale_factor={:.2f}
