+++ autocti/config/visualize/plots.ini
@@ -1,28 +1,0 @@
-[dataset]
-subplot_dataset=True
-data=False
-noise_map=False
-inverse_noise_map=False
-signal_to_noise_map=False
-pre_cti_data=False
-cosmic_ray_map=False
-noise_scaling_map_dict=False
-absolute_signal_to_noise_map=False
-potential_chi_squared_map=False
-
-[fit]
-all_at_end_png=False
-all_at_end_fits=False
-subplot_fit=True
-subplot_residual_maps=True
-subplot_normalized_residual_maps=True
-subplot_chi_squared_maps=True
-data=False
-noise_map=False
-signal_to_noise_map=False
-pre_cti_data=False
-post_cti_data=False
-residual_map=False
-normalized_residual_map=False
-chi_squared_map=False
-noise_scaling_map_dict=False
