### 2022-12-01 Jammy2211/PyAutoCTI#31 @jryon
Problems installing arCTIc - JOSS review
Hi, I've started reviewing your JOSS submission. My attempts to install arCTIc have not been successful.

First, I'll note that the conda and pip installation guides on readthedocs both say "Once you have created your conda environment you must install arCTIc before installing PyAutoCTI. The installation guide is found at this link." but no link is present. I think you mean to link to [this page](https://pyautocti.readthedocs.io/en/latest/installation/arctic.html).

However, my first attempts to install arCTIc were by using the instructions listed in the arCTIc README [here](https://github.com/jkeger/arctic#installation). These instructions are a bit confusing, and the uncommented pip commands failed for me. I also do not have homebrew installed, so it's not clear to me how to install those libraries without it, or check that I already have them.

Then, I found the Installing arCTIc page on readthedocs. When running `make all` (I cannot use sudo on my machine) after cloning the arCTIc repository, I get the following error:

/System/Volumes/Data/user/ryon/src/arctic/src /System/Volumes/Data/user/ryon/src/arctic/build
/System/Volumes/Data/user/ryon/src/arctic/src/trap_managers.cpp /System/Volumes/Data/user/ryon/src/arctic/src/main.cpp /System/Volumes/Data/user/ryon/src/arctic/src/roe.cpp /System/Volumes/Data/user/ryon/src/arctic/src/cti.cpp /System/Volumes/Data/user/ryon/src/arctic/src/util.cpp /System/Volumes/Data/user/ryon/src/arctic/src/ccd.cpp /System/Volumes/Data/user/ryon/src/arctic/src/traps.cpp  /System/Volumes/Data/user/ryon/src/arctic/build/trap_managers.o  /System/Volumes/Data/user/ryon/src/arctic/build/main.o  /System/Volumes/Data/user/ryon/src/arctic/build/roe.o  /System/Volumes/Data/user/ryon/src/arctic/build/cti.o  /System/Volumes/Data/user/ryon/src/arctic/build/util.o  /System/Volumes/Data/user/ryon/src/arctic/build/ccd.o  /System/Volumes/Data/user/ryon/src/arctic/build/traps.o
Downloading and installing GSL 2.6 to /opt/local
+ cd /System/Volumes/Data/user/ryon/src/arctic
+ '[' -d gsl-2.6 ']'
+ '[' -f gsl-2.6.tar.gz ']'
+ wget ftp://ftp.gnu.org/gnu/gsl/gsl-2.6.tar.gz
./get_gsl.sh: line 18: wget: command not found
make: *** [gsl] Error 127

I will now go install `wget`, but I think more description of the requirements for installing arCTIc is needed in the installation guide.
