### 2022-12-01 Jammy2211/PyAutoCTI#31 @jryon
Quick update. I've now installed `wget` with conda, but appear to have run into a permissions problem, likely because I can't use sudo on my work machine. I'll try installing on my personal machine instead.

...
x gsl-2.6/doc/examples/statsort.txt
x gsl-2.6/doc/examples/stat.txt
x gsl-2.6/doc/examples/sum.txt
x gsl-2.6/doc/examples/vectorr.txt
x gsl-2.6/doc/examples/vectorview.txt
x gsl-2.6/doc/examples/nlfit4.c
+ mkdir -p /opt/local
mkdir: /opt/local: Permission denied
make: *** [gsl] Error 1
