### 2022-12-01 #4904 @jryon
Installation issues: https://github.com/Jammy2211/PyAutoCTI/issues/31
