+++ docs/general/credits.rst
@@ -1,0 +1,35 @@
+.. _credits:
+
+
+Credits
+-------
+
+**Developers**:
+
+`James Nightingale <https://github.com/Jammy2211>`_: Lead developer & PyAutoLens guru.
+
+`Richard Hayes <https://github.com/rhayes777>`_: Lead developer & `PyAutoFit <https://github.com/rhayes777/PyAutoFit>`_ guru.
+
+`Aristeidis Amvrosiadis <https://github.com/Sketos>`_: Interferometer Analysis.
+
+`Amy Etherington <https://github.com/amyetherington>`_ : Magnification, Critical Curves and Caustic Calculations.
+
+Qiuhan He: NFW Profile Lensing Calculations.
+
+`Johnathan Frawley <https://github.com/jonathanfrawley>`_: Profiling, optimization and build server tools.
+
+`Ashley Kelly <https://github.com/AshKelly>`_: Developer of `pyquad <https://github.com/AshKelly/pyquad>`_ for fast deflections computations.
+
+`Xiaoyue Cao <https://github.com/caoxiaoyue>`_: Analytic Ellipitcal Power-Law Deflection Angle Calculations.
+
+`Nan Li <https://github.com/linan7788626>`_: Docker integration & support.
+
+**Code Donors**:
+
+`Andrew Robertson <https://github.com/Andrew-Robertson>`_: Critical curve & caustic calculations.
+
+Mattia Negrello: Visibility models in the uv-plane via direct Fourier transforms.
+
+`Andrea Enia <https://github.com/AndreaEnia>`_: Voronoi source-plane plotting tools.
+
+Conor O'Riordan : Broken Power-Law mass profile.
