+++ CITATIONS.rst
@@ -1,0 +1,20 @@
+.. _references:
+
+Citations & References
+======================
+
+The bibtex entries for **PyAutoCTI** and its affiliated software packages can be found
+`here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.bib>`_, with example text for citing **PyAutoCTI**
+in `.tex format here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.tex>`_ format here and
+`.md format here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.md>`_.
+
+As shown in the examples, we would greatly appreciate it if you mention **PyAutoCTI** by name and include a link to
+our GitHub page!
+
+**PyAutoCTI** is published in the `Journal of Open Source Software <https://joss.theoj.org/papers/10.21105/joss.02825#>`_ and its
+entry in the above .bib file is under the citation key ``pyautocti``.
+
+You should also specify the non-linear search(es) you use in your analysis (e.g. Dynesty, Emcee, PySwarms, etc) in
+the main body of text, and delete as appropriate any packages your analysis did not use. The citations.bib file includes
+the citation key for all of these projects.
+
