+++ test_autocti/config/visualize/plots.ini
@@ -1,34 +1,0 @@
-[dataset]
-subplot_dataset=True
-data=True
-noise_map=False
-inverse_noise_map=False
-signal_to_noise_map=False
-pre_cti_data=True
-cosmic_ray_map=True
-noise_scaling_map_dict=False
-absolute_signal_to_noise_map=False
-potential_chi_squared_map=Fals
-
-[fit]
-all_at_end_png=True
-all_at_end_fits=False
-subplot_fit=True
-subplot_residual_maps=True
-subplot_normalized_residual_maps=True
-subplot_chi_squared_maps=False
-data=True
-noise_map=False
-signal_to_noise_map=False
-pre_cti_data=True
-post_cti_data=True
-residual_map=False
-normalized_residual_map=True
-chi_squared_map=True
-noise_scaling_map_dict=False
-
-[positions]
-image_with_positions=True
-
-[other]
-stochastic_histogram=False
