+++ test_autocti/config/non_linear/nest/DynestyStatic.ini
@@ -1,37 +1,0 @@
-[search]
-nlive=150
-bound=multi
-sample=auto
-bootstrap=0
-enlarge=-1
-update_interval=-1.0
-walks=25
-sampling_efficiency=0.5
-slices=5
-fmove=0.9
-max_move=100
-maxiter=None
-maxcall=None
-logl_max=inf
-n_effective=0
-
-[updates]
-iterations_per_update=2500
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=True
-
-[initialize]
-method=prior
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[parallel]
-number_of_cores=1

[…]
