+++ test_autocti/charge_injection/test_ou_sim_ci.py
@@ -12,7 +12,8 @@ def test__non_uniform_array_is_correct_with_rotation():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="E",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -29,7 +30,8 @@ def test__non_uniform_array_is_correct_with_rotation():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="H",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -45,7 +47,8 @@ def test__non_uniform_array_is_correct_with_rotation():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="F",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -62,7 +65,8 @@ def test__non_uniform_array_is_correct_with_rotation():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="G",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -93,7 +97,8 @@ def test__add_cti_to_pre_cti_data():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="E",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -123,7 +128,8 @@ def test__add_cti_to_pre_cti_data():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="H",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -153,7 +159,8 @@ def test__add_cti_to_pre_cti_data():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="F",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -183,7 +190,8 @@ def test__add_cti_to_pre_cti_data():
     array = ou_sim_ci.charge_injection_array_from(
         ccd_id="123",
         quadrant_id="G",
-        injection_total=5,
+        injection_start=0,
+        injection_end=2000,
         injection_on=200,
         injection_off=200,
         injection_norm=50000.0,
@@ -215,8 +223,8 @@ def test__tvac_values():
         #   iquad=0,
         ccd_id="123",
         quadrant_id="E",
-        injection_total=4,
         injection_start=16,
+        injection_end=2086,
         injection_on=420,
         injection_off=100,
         injection_norm=50000.0,
@@ -242,8 +250,8 @@ def test__tvac_values():
         #   iquad=1,
         ccd_id="123",
         quadrant_id="F",
-        injection_total=4,
         injection_start=16,
+        injection_end=2086,
         injection_on=420,
         injection_off=100,
         injection_norm=50000.0,
@@ -269,8 +277,8 @@ def test__tvac_values():
         #  iquad=2,
         ccd_id="123",
         quadrant_id="H",
-        injection_total=4,
         injection_start=16,
+        injection_end=2086,
[…]
