+++ files/citations.bib
@@ -33,38 +33,6 @@ Year = 2018,
 Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}
 }
 
-@article{PyLops,
-abstract = {Linear operators and optimisation are at the core of many algorithms used in signal and image processing, remote sensing, and inverse problems. For small to medium-scale problems, existing software packages (e.g., MATLAB, Python numpy and scipy) allow for explicitly building dense (or sparse) matrices and performing algebraic operations (e.g., computation of matrix-vector products and manipulation of matrices) with syntax that closely represents their corresponding analytical forms. However, many real application, large-scale operators do not lend themselves to explicit matrix representations, usually forcing practitioners to forego of the convenient linear-algebra syntax available for their explicit-matrix counterparts. PyLops is an open-source Python library providing a flexible and scalable framework for the creation and combination of so-called linear operators, class-based entities that represent matrices and inherit their associated syntax convenience, but do not rely on the creation of explicit matrices. We show that PyLops operators can dramatically reduce the memory load and CPU computations compared to explicit-matrix calculations, while still allowing users to seamlessly use their existing knowledge of compact matrix-based syntax that scales to any problem size because no explicit matrices are required.},
-archivePrefix = {arXiv},
-arxivId = {1907.12349},
-author = {Ravasi, Matteo and Vasconcelos, Ivan},
-eprint = {1907.12349},
-file = {:home/jammy/Documents/Papers/Software/PyLops.pdf:pdf},
-title = {{PyLops -- A Linear-Operator Python Library for large scale optimization}},
-url = {http://arxiv.org/abs/1907.12349},
-year = {2019}
-}
-
-@article{colossus,
-abstract = {This paper introduces Colossus, a public, open-source python package for calculations related to cosmology, the large-scale structure (LSS) of matter in the universe, and the properties of dark matter halos. The code is designed to be fast and easy to use, with a coherent, well-documented user interface. The cosmology module implements Friedman-Lemaitre-Robertson-Walker cosmologies including curvature, relativistic species, and different dark energy equations of state, and provides fast computations of the linear matter power spectrum, variance, and correlation function. The LSS module is concerned with the properties of peaks in Gaussian random fields and halos in a statistical sense, including their peak height, peak curvature, halo bias, and mass function. The halo module deals with spherical overdensity radii and masses, density profiles, concentration, and the splashback radius. To facilitate the rapid exploration of these quantities, Colossus implements more than 40 different fitting functions from the literature. I discuss the core routines in detail, with particular emphasis on their accuracy. Colossus is available at bitbucket.org/bdiemer/colossus.},
-archivePrefix = {arXiv},
-arxivId = {1712.04512},
-author = {Diemer, Benedikt},
-doi = {10.3847/1538-4365/aaee8c},
-eprint = {1712.04512},
-file = {:home/jammy/Documents/Papers/Software/Collosus2018.pdf:pdf},
-issn = {0067-0049},
-journal = {The Astrophysical Journal Supplement Series},
-keywords = {cosmology,cosmology: theory,methods: numerical,methods,numerical,theory},
-number = {2},
-pages = {35},
-publisher = {IOP Publishing},
-title = {{COLOSSUS: A Python Toolkit for Cosmology, Large-scale Structure, and Dark Matter Halos}},
-url = {http://dx.doi.org/10.3847/1538-4365/aaee8c},
[…]
