### 2022-12-02 2fd405c
comsic ray map in visualize config
