+++ autocti/config/priors/traps.yaml
@@ -1,14 +1,14 @@
 TrapInstantCapture:
   density:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 10.0
     width_modifier:
       type: Relative
       value: 0.5
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
   fractional_volume_full_exposed:
     type: Constant
     value: 0.0
@@ -16,43 +16,43 @@ TrapInstantCapture:
     type: Constant
     value: 0.0
   release_timescale:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 50.0
     width_modifier:
       type: Relative
       value: 0.5
-TrapInstantCaptureContinuum:
-  density:
     gaussian_limits:
       lower: 0.0
       upper: inf
-    lower_limit: 0.0
+TrapInstantCaptureContinuum:
+  density:
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 10.0
     width_modifier:
       type: Relative
       value: 0.5
-  release_timescale:
     gaussian_limits:
       lower: 0.0
       upper: inf
-    lower_limit: 0.0
+  release_timescale:
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 50.0
     width_modifier:
       type: Relative
       value: 0.5
-  release_timescale_sigma:
     gaussian_limits:
       lower: 0.0
       upper: inf
-    lower_limit: 0.0
+  release_timescale_sigma:
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 1.0
     width_modifier:
       type: Relative
       value: 0.5
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
