+++ autocti/config/priors/hyper.yaml
@@ -1,11 +1,11 @@
 HyperCINoiseScalar:
   scale_factor:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
     type: Uniform
+    lower_limit: 0.0
     upper_limit: 10.0
     width_modifier:
       type: Relative
       value: 0.5
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
