### 2022-12-06 9de30a5
docs on arcticpy installation
