### 2022-12-06 Jammy2211/PyAutoCTI#31 @Jammy2211
Can you check if you can install arcticpy via the command `pip install arcticpy==2.1`.

I don't think this will work with autocti, and I'll fix this with a release in the next week, but its good first making sure the pip build of arcticpy installs OK.

(The install instructions were outdated as we got a pip release finally working, I have updated readthedocs accordingly).
