### 2022-12-06 Jammy2211/PyAutoCTI#31 @jryon
This doesn't appear to be working for me. I did the following:
`conda create -n python=3.11 autocti astropy numpy scikit-image scikit-learn scipy`
`conda activate autocti`
`pip install --upgrade pip`
`pip install arcticpy==2.1`

Here's the output:

Collecting arcticpy==2.1
  Downloading arcticpy-2.1.tar.gz (285 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 285.3/285.3 kB 5.4 MB/s eta 0:00:00
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting autoconf
  Downloading autoconf-2022.11.26.11-py3-none-any.whl (14 kB)
Requirement already satisfied: numpy~=1.21 in /Users/ryon/miniconda3/envs/autocti/lib/python3.11/site-packages (from arcticpy==2.1) (1.23.5)
Collecting typing-inspect>=0.4.0
  Using cached typing_inspect-0.8.0-py3-none-any.whl (8.7 kB)
Collecting PyYAML==5.4.1
  Using cached PyYAML-5.4.1-cp311-cp311-macosx_10_9_x86_64.whl
Collecting mypy-extensions>=0.3.0
  Using cached mypy_extensions-0.4.3-py2.py3-none-any.whl (4.5 kB)
Collecting typing-extensions>=3.7.4
  Using cached typing_extensions-4.4.0-py3-none-any.whl (26 kB)
Building wheels for collected packages: arcticpy
  Building wheel for arcticpy (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for arcticpy (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [102 lines of output]
      rm python/arcticpy/wrapper.cpp
      Compiling python/arcticpy/wrapper.pyx because it changed.
      [1/1] Cythonizing python/arcticpy/wrapper.pyx
      /private/var/folders/nz/gtcj6sl91mb8v7ccdhlxvl2w0001tl/T/pip-build-env-jo6157ub/overlay/lib/python3.11/site-packages/setuptools/config/pyprojecttoml.py:108: _BetaConfiguration: Support for `[tool.setuptools]` in `pyproject.toml` is still *beta*.
        warnings.warn(msg, _BetaConfiguration)
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build/lib.macosx-10.9-x86_64-cpython-311
      creating build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/__init__.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/pixel_bounce.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/traps.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/cti.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/roe.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/ccd.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/read_noise.py -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      running egg_info
      writing python/arcticpy.egg-info/PKG-INFO
      writing dependency_links to python/arcticpy.egg-info/dependency_links.txt
      writing requirements to python/arcticpy.egg-info/requires.txt
      writing top-level names to python/arcticpy.egg-info/top_level.txt
      listing git files failed - pretending there aren't any
      reading manifest file 'python/arcticpy.egg-info/SOURCES.txt'
      reading manifest template 'MANIFEST.in'
      warning: no previously-included files found matching 'python/arcticpy/src/wrapper.cpp'
      writing manifest file 'python/arcticpy.egg-info/SOURCES.txt'
      /private/var/folders/nz/gtcj6sl91mb8v7ccdhlxvl2w0001tl/T/pip-build-env-jo6157ub/overlay/lib/python3.11/site-packages/setuptools/command/build_py.py:202: SetuptoolsDeprecationWarning:     Installing 'arcticpy.include' as data is deprecated, please list it in `packages`.
          !!

          ############################
          # Package would be ignored #
          ############################
          Python recognizes 'arcticpy.include' as an importable package,
          but it is not listed in the `packages` configuration of setuptools.

          'arcticpy.include' has been automatically added to the distribution only
          because it may contain data files, but this behavior is likely to change
          in future versions of setuptools (and therefore is considered deprecated).

          Please make sure that 'arcticpy.include' is included as a package by using
          the `packages` configuration field or the proper discovery methods
          (for example by using `find_namespace_packages(...)`/`find_namespace:`
          instead of `find_packages(...)`/`find:`).

          You can read more about "package discovery" and "data files" on setuptools
          documentation page.

      !!

        check.warn(importable)
      /private/var/folders/nz/gtcj6sl91mb8v7ccdhlxvl2w0001tl/T/pip-build-env-jo6157ub/overlay/lib/python3.11/site-packages/setuptools/command/build_py.py:202: SetuptoolsDeprecationWarning:     Installing 'arcticpy.src' as data is deprecated, please list it in `packages`.
          !!

          ############################
          # Package would be ignored #
          ############################
          Python recognizes 'arcticpy.src' as an importable package,
          but it is not listed in the `packages` configuration of setuptools.

          'arcticpy.src' has been automatically added to the distribution only
          because it may contain data files, but this behavior is likely to change
          in future versions of setuptools (and therefore is considered deprecated).

          Please make sure that 'arcticpy.src' is included as a package by using
          the `packages` configuration field or the proper discovery methods
          (for example by using `find_namespace_packages(...)`/`find_namespace:`
          instead of `find_packages(...)`/`find:`).

          You can read more about "package discovery" and "data files" on setuptools
          documentation page.

      !!

        check.warn(importable)
      copying python/arcticpy/wrapper.cpp -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      copying python/arcticpy/wrapper.pyx -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy
      creating build/lib.macosx-10.9-x86_64-cpython-311/arcticpy/include
      copying python/arcticpy/include/interface.hpp -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy/include
      creating build/lib.macosx-10.9-x86_64-cpython-311/arcticpy/src
      copying python/arcticpy/src/interface.cpp -> build/lib.macosx-10.9-x86_64-cpython-311/arcticpy/src
      running build_ext
      building 'arcticpy.wrapper' extension
      creating build/temp.macosx-10.9-x86_64-cpython-311
      creating build/temp.macosx-10.9-x86_64-cpython-311/python
      creating build/temp.macosx-10.9-x86_64-cpython-311/python/arcticpy
      creating build/temp.macosx-10.9-x86_64-cpython-311/python/arcticpy/src
      creating build/temp.macosx-10.9-x86_64-cpython-311/src
      g++ -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/ryon/miniconda3/envs/autocti/include -fPIC -O2 -isystem /Users/ryon/miniconda3/envs/autocti/include -DNPY_NO_DEPRECATED_API=0 -Ipython/arcticpy/include -Iinclude/ -I/private/var/folders/nz/gtcj6sl91mb8v7ccdhlxvl2w0001tl/T/pip-build-env-jo6157ub/overlay/lib/python3.11/site-packages/numpy/core/include -I/usr/local/include -I/Users/ryon/miniconda3/envs/autocti/include/python3.11 -c python/arcticpy/src/interface.cpp -o build/temp.macosx-10.9-x86_64-cpython-311/python/arcticpy/src/interface.o -std=c++17 -O3
      In file included from python/arcticpy/src/interface.cpp:2:
      In file included from python/arcticpy/include/interface.hpp:2:
      In file included from include/cti.hpp:7:
      include/traps.hpp:5:10: fatal error: 'gsl/gsl_integration.h' file not found
      #include <gsl/gsl_integration.h>
               ^~~~~~~~~~~~~~~~~~~~~~~
      1 error generated.
      error: command '/usr/bin/g++' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for arcticpy
Failed to build arcticpy
ERROR: Could not build wheels for arcticpy, which is required to install pyproject.toml-based projects
