### 2022-12-09 Jammy2211/PyAutoCTI#31 @jryon
GSL is installed on my personal machine, but homebrew puts libraries in /opt/homebrew on computers with Apple silicon. So it does appear GSL is not linked correctly for installing arctic.  I'm not familiar enough with brew, linking, or makefiles to make this work. Any suggestions?
