### 2022-12-09 #4904 @jryon
@adonath I started my review, but quickly ran into installation issues with arctic, a key requirement for PyAutoCTI. The authors are working on solutions, but it's possible I will not be able to install and run PyAutoCTI until after the 6 week review deadline. This brings up a couple of questions about how this review should work (I couldn't find the answers in the JOSS documentation). Is the review expected to be an iterative process with the authors until all checklist items are completed to our satisfaction? If so, I'm a bit confused by the need for a review deadline. Relatedly, what does a complete review look like? Do we simply complete the checklist? Or should we include a short writeup in a comment?
