### 2022-12-13 #4904 @adonath
Thanks @jryon for starting the review! Please find my answers below:

> Is the review expected to be an iterative process with the authors until all checklist items are completed to our satisfaction?

Yes, this is mostly it. The review is driven by the checklist, so the goal should be to check all of the points on the list off. If you have further viable points, feel free to bring those up to.

> If so, I'm a bit confused by the need for a review deadline. Relatedly, what does a complete review look like? Do we simply complete the checklist?

Yes, again this is mostly the goal. The review deadline is not strict, but usually JOSS aims for a fast publication, which is more fair to submitters. So ideally the review can happened within 2-4 weeks. If more time is need to resolve open issues with the authors, then that's it. In case I have the impression a reviewer is overly critical and tries intentionally (for whatever reasons) to delay the release I will intervene in the review process.

> Or should we include a short writeup in a comment?

This is not required, but I think authors are happy to read a short final summary!

As the whole review process is open, you can find many example reviews in the issue tracker of closed issues https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+is%3Aclosed+contains%3A%5BREVIEW%5D+label%3Areview
