+++ optional_requirements.txt
@@ -1,4 +1,5 @@
 numba
+lacosmic
 pylops>=1.10.0,<=1.18.3
 pynufft
 ultranest==3.2.0
