### 2022-12-16 Jammy2211/PyAutoCTI#31 @Jammy2211
Can you try a `pip install arcticpy_no_gsl` and let me know if you get an error? :)
