### 2022-12-17 0c1ff63
cr .fits files now python scripts
