+++ autocti/cosmics/cr_lengths.py
@@ -1,0 +1,471 @@
+import numpy as np
+
+lengths = np.array(
+    [
+        (0.0, 0.0),
+        (1.0, 0.00707294),
+        (2.0, 0.05645055),
+        (3.0, 0.13844089),
+        (4.0, 0.26378389),
+        (5.0, 0.3878169),
+        (6.0, 0.4998),
+        (7.0, 0.59645923),
+        (8.0, 0.67163863),
+        (9.0, 0.72822017),
+        (10.0, 0.77314481),
+        (11.0, 0.81086551),
+        (12.0, 0.84033528),
+        (13.0, 0.8623391),
+        (14.0, 0.87949696),
+        (15.0, 0.89757182),
+        (16.0, 0.90883573),
+        (17.0, 0.91931365),
+        (18.0, 0.92586259),
+        (19.0, 0.93411353),
+        (20.0, 0.94040048),
+        (21.0, 0.94485344),
+        (22.0, 0.9496994),
+        (23.0, 0.95493836),
+        (24.0, 0.95952232),
+        (25.0, 0.9629273),
+        (26.0, 0.96541628),
+        (27.0, 0.96751226),
+        (28.0, 0.96987024),
+        (29.0, 0.97183523),
+        (30.0, 0.97288322),
+        (31.0, 0.9744552),
+        (32.0, 0.97589619),
+        (33.0, 0.97773018),
+        (34.0, 0.97904017),
+        (35.0, 0.97995716),
+        (36.0, 0.98100515),
+        (37.0, 0.98218414),
+        (38.0, 0.98283914),
+        (39.0, 0.98349413),
+        (40.0, 0.98388713),
+        (41.0, 0.98506612),
+        (42.0, 0.98559012),
+        (43.0, 0.98650711),
+        (44.0, 0.98676911),
+        (45.0, 0.9874241),
+        (46.0, 0.9875551),
+        (47.0, 0.9880791),
+        (48.0, 0.98834109),
+        (49.0, 0.98886509),
+        (50.0, 0.98925809),
+        (51.0, 0.98978208),
+        (52.0, 0.99030608),
+        (53.0, 0.99056808),
+        (54.0, 0.99069907),
+        (55.0, 0.99069907),
+        (56.0, 0.99135407),
+        (57.0, 0.99135407),
+        (58.0, 0.99200906),
+        (59.0, 0.99227106),
+        (60.0, 0.99253306),
+        (61.0, 0.99253306),
+        (62.0, 0.99266406),
+        (63.0, 0.99305706),
+        (64.0, 0.99305706),
+        (65.0, 0.99318805),
+        (66.0, 0.99345005),
+        (67.0, 0.99397405),
+        (68.0, 0.99397405),
+        (69.0, 0.99397405),
+        (70.0, 0.99410505),
+        (71.0, 0.99436705),
+        (72.0, 0.99449804),
+        (73.0, 0.99449804),
+        (74.0, 0.99449804),
+        (75.0, 0.99476004),
+        (76.0, 0.99489104),
+        (77.0, 0.99489104),
+        (78.0, 0.99502204),
+        (79.0, 0.99502204),
+        (80.0, 0.99528404),
+        (81.0, 0.99554604),
+        (82.0, 0.99580803),
+        (83.0, 0.99593903),
+        (84.0, 0.99607003),
+        (85.0, 0.99607003),
+        (86.0, 0.99607003),
+        (87.0, 0.99607003),
+        (88.0, 0.99607003),
+        (89.0, 0.99620103),
+        (90.0, 0.99620103),
+        (91.0, 0.99633203),
+        (92.0, 0.99633203),
+        (93.0, 0.99633203),
+        (94.0, 0.99646303),
+        (95.0, 0.99659403),
+        (96.0, 0.99659403),
[…]
