+++ autocti/cosmics/cr_distances.py
@@ -1,0 +1,1007 @@
+import numpy as np
+
+distances = np.array(
+    [
+        (1.19748284e03, 0.0000000e00),
+        (1.20710112e03, 1.5214563e-05),
+        (1.21679666e03, 3.0769871e-05),
+        (1.22657008e03, 4.6678274e-05),
+        (1.23642198e03, 6.2952733e-05),
+        (1.24635304e03, 7.9606853e-05),
+        (1.25636386e03, 9.6654917e-05),
+        (1.26645508e03, 1.1411193e-04),
+        (1.27662734e03, 1.3199364e-04),
+        (1.28688132e03, 1.5031663e-04),
+        (1.29721766e03, 1.6909828e-04),
+        (1.30763704e03, 1.8835692e-04),
+        (1.31814008e03, 2.0811178e-04),
+        (1.32872750e03, 2.2838311e-04),
+        (1.33939996e03, 2.4919223e-04),
+        (1.35015812e03, 2.7056155e-04),
+        (1.36100272e03, 2.9251470e-04),
+        (1.37193440e03, 3.1507653e-04),
+        (1.38295390e03, 3.3827323e-04),
+        (1.39406190e03, 3.6213240e-04),
+        (1.40525912e03, 3.8668312e-04),
+        (1.41654628e03, 4.1195604e-04),
+        (1.42792410e03, 4.3798347e-04),
+        (1.43939332e03, 4.6479950e-04),
+        (1.45095464e03, 4.9244008e-04),
+        (1.46260884e03, 5.2094313e-04),
+        (1.47435664e03, 5.5034867e-04),
+        (1.48619880e03, 5.8069894e-04),
+        (1.49813608e03, 6.1203852e-04),
+        (1.51016922e03, 6.4441451e-04),
+        (1.52229904e03, 6.7787659e-04),
+        (1.53452628e03, 7.1247727e-04),
+        (1.54685172e03, 7.4827202e-04),
+        (1.55927616e03, 7.8531940e-04),
+        (1.57180040e03, 8.2368131e-04),
+        (1.58442524e03, 8.6342317e-04),
+        (1.59715148e03, 9.0461407e-04),
+        (1.60997994e03, 9.4732706e-04),
+        (1.62291144e03, 9.9163932e-04),
+        (1.63594680e03, 1.0376324e-03),
+        (1.64908686e03, 1.0853927e-03),
+        (1.66233246e03, 1.1350111e-03),
+        (1.67568446e03, 1.1865842e-03),
+        (1.68914370e03, 1.2402136e-03),
+        (1.70271104e03, 1.2960070e-03),
+        (1.71638736e03, 1.3540781e-03),
+        (1.73017354e03, 1.4145471e-03),
+        (1.74407044e03, 1.4775408e-03),
+        (1.75807896e03, 1.5431935e-03),
+        (1.77220000e03, 1.6116468e-03),
+        (1.78643446e03, 1.6830502e-03),
+        (1.80078326e03, 1.7575617e-03),
+        (1.81524730e03, 1.8353480e-03),
+        (1.82982752e03, 1.9165850e-03),
+        (1.84452486e03, 2.0014583e-03),
+        (1.85934024e03, 2.0901635e-03),
+        (1.87427462e03, 2.1829071e-03),
+        (1.88932896e03, 2.2799064e-03),
+        (1.90450420e03, 2.3813904e-03),
+        (1.91980134e03, 2.4876003e-03),
+        (1.93522136e03, 2.5987898e-03),
+        (1.95076522e03, 2.7152258e-03),
+        (1.96643394e03, 2.8371887e-03),
+        (1.98222850e03, 2.9649735e-03),
[…]
