### 2022-12-18 53420c8
move ci visualization to modify_before_fit
