### 2022-12-18 eeac7c7
auto preload oise normalizaiton for now, until Rich fixes has
