### 2022-12-18 2d5354c
comment out test which Rich will fix in autofit
