### 2022-12-18 4f05798
fast chi squared calculation
