### 2022-12-18 ddfc1fe
fast chi squared calculation now used
