### 2022-12-18 31a1bb7
fix chi squared fqast
