+++ docs/general/citations.rst
@@ -4,9 +4,9 @@ Citations & References
 ======================
 
 The bibtex entries for **PyAutoCTI** and its affiliated software packages can be found
-`here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.bib>`_, with example text for citing **PyAutoCTI**
-in `.tex format here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.tex>`_ format here and
-`.md format here <https://github.com/Jammy2211/PyAutoCTI/blob/master/files/citations.md>`_.
+`here <https://github.com/Jammy2211/PyAutoCTI/blob/main/files/citations.bib>`_, with example text for citing **PyAutoCTI**
+in `.tex format here <https://github.com/Jammy2211/PyAutoCTI/blob/main/files/citations.tex>`_ format here and
+`.md format here <https://github.com/Jammy2211/PyAutoCTI/blob/main/files/citations.md>`_.
 
 As shown in the examples, we would greatly appreciate it if you mention **PyAutoCTI** by name and include a link to
 our GitHub page!
