+++ README.rst
@@ -11,13 +11,19 @@ PyAutoCTI: Charge Transfer Inefficiency Modeling
     :target: https://pyautocti.readthedocs.io/en/latest/?badge=latest
     :alt: Documentation Status
 
+.. |Tests| image:: https://github.com/Jammy2211/PyAutoCTI/actions/workflows/main.yml/badge.svg
+   :target: https://github.com/Jammy2211/PyAutoCTI/actions
+
+.. |Build| image:: https://github.com/Jammy2211/PyAutoBuild/actions/workflows/release.yml/badge.svg
+   :target: https://github.com/Jammy2211/PyAutoBuild/actions
+
 .. |code-style| image:: https://img.shields.io/badge/code%20style-black-000000.svg
     :target: https://github.com/psf/black
 
 .. |arXiv| image:: https://img.shields.io/badge/arXiv-1708.07377-blue
     :target: https://arxiv.org/abs/0909.0507
 
-|binder| |RTD| |code-style| |arXiv|
+|binder| |RTD| |Tests| |Build| |code-style| |arXiv|
 
 `Installation Guide <https://pyautocti.readthedocs.io/en/latest/installation/overview.html>`_ |
 `readthedocs <https://pyautocti.readthedocs.io/en/latest/index.html>`_ |
