### 2022-12-20 Jammy2211/PyAutoCTI#31 @jryon
Yes, that worked. I then tried `pip install autocti` but that failed.
