### 2022-12-20 Jammy2211/PyAutoCTI#31 @Jammy2211
What was the error?

I dont think autocti should work yet (I need to do a release) but would be good to confirm the error makes sense...
