### 2022-12-20 Jammy2211/PyAutoCTI#31 @jryon
Looks like it errored on numba and llvmlite. The full error is quite long, so I've removed many of the lines related to copying and creating builds for numba:

Building wheels for collected packages: numba, llvmlite
  Building wheel for numba (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [795 lines of output]
      TBB not found
      Using OpenMP from: /Users/jryon/miniconda3/envs/autocti
      running bdist_wheel
      running build
      got version from file /private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/numba_c97e508dab0a4e65a3b9738df503a076/numba/_version.py {'version': '0.51.2', 'full': '9d570961590c09a1eba748c9c37e91d1224fc9ad'}
      running build_py
      creating build
      creating build/lib.macosx-11.1-arm64-cpython-310
      creating build/lib.macosx-11.1-arm64-cpython-310/numba
      copying numba/_version.py -> build/lib.macosx-11.1-arm64-cpython-310/numba
      copying numba/__init__.py -> build/lib.macosx-11.1-arm64-cpython-310/numba
      copying numba/__main__.py -> build/lib.macosx-11.1-arm64-cpython-310/numba
      .....
      copying numba/cuda/tests/cudadrv/data/jitlink.ptx -> build/lib.macosx-11.1-arm64-cpython-310/numba/cuda/tests/cudadrv/data
      running build_ext
      building 'numba._dynfunc' extension
      Warning: Can't read registry to find the necessary compiler setting
      Make sure that Python modules winreg, win32api or win32con are installed.
      INFO: C compiler: clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64 -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64

      creating build/temp.macosx-11.1-arm64-cpython-310
      creating build/temp.macosx-11.1-arm64-cpython-310/numba
      INFO: compile options: '-I/Users/jryon/miniconda3/envs/autocti/include/python3.10 -c'
      extra options: '-g'
      INFO: clang: numba/_dynfuncmod.c
      INFO: clang -bundle -undefined dynamic_lookup -Wl,-rpath,/Users/jryon/miniconda3/envs/autocti/lib -L/Users/jryon/miniconda3/envs/autocti/lib -Wl,-rpath,/Users/jryon/miniconda3/envs/autocti/lib -L/Users/jryon/miniconda3/envs/autocti/lib build/temp.macosx-11.1-arm64-cpython-310/numba/_dynfuncmod.o -o build/lib.macosx-11.1-arm64-cpython-310/numba/_dynfunc.cpython-310-darwin.so
      ld: warning: -undefined dynamic_lookup may not work with chained fixups
      building 'numba._dispatcher' extension
      INFO: C compiler: clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64 -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64

      creating build/temp.macosx-11.1-arm64-cpython-310/numba/core
      creating build/temp.macosx-11.1-arm64-cpython-310/numba/core/typeconv
      INFO: compile options: '-I/Users/jryon/miniconda3/envs/autocti/lib/python3.10/site-packages/numpy/core/include -I/Users/jryon/miniconda3/envs/autocti/include/python3.10 -c'
      INFO: clang: numba/_hashtable.c
      INFO: clang: numba/_typeof.c
      INFO: clang: numba/core/typeconv/typeconv.cpp
      INFO: clang: numba/_dispatcher.c
      INFO: clang: numba/_dispatcherimpl.cpp
      numba/_dispatcher.c:26:13: error: no member named 'use_tracing' in 'struct _ts'
          tstate->use_tracing = 0;
          ~~~~~~  ^
      numba/_dispatcher.c:28:13: error: no member named 'use_tracing' in 'struct _ts'
          tstate->use_tracing = ((tstate->c_tracefunc != NULL)
          ~~~~~~  ^
      numba/_dispatcher.c:303:17: error: no member named 'use_tracing' in 'struct _ts'
          if (tstate->use_tracing && tstate->c_profilefunc)
              ~~~~~~  ^
      numba/_dispatcher.c:512:13: error: no member named 'use_tracing' in 'struct _ts'
          if (ts->use_tracing && ts->c_profilefunc) {
              ~~  ^
      4 errors generated.
      error: Command "clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64 -fPIC -O2 -isystem /Users/jryon/miniconda3/envs/autocti/include -arch arm64 -I/Users/jryon/miniconda3/envs/autocti/lib/python3.10/site-packages/numpy/core/include -I/Users/jryon/miniconda3/envs/autocti/include/python3.10 -c numba/_dispatcher.c -o build/temp.macosx-11.1-arm64-cpython-310/numba/_dispatcher.o" failed with exit status 1
      [end of output]

  ERROR: Failed building wheel for numba
  Running setup.py clean for numba
  Building wheel for llvmlite (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [26 lines of output]
      running bdist_wheel
      /Users/jryon/miniconda3/envs/autocti/bin/python /private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py
      LLVM version... Traceback (most recent call last):
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 105, in main_posix
          out = subprocess.check_output([llvm_config, '--version'])
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 421, in check_output
          return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 503, in run
          with Popen(*popenargs, **kwargs) as process:
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 971, in __init__
          self._execute_child(args, executable, preexec_fn, close_fds,
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 1847, in _execute_child
          raise child_exception_type(errno_num, err_msg, err_filename)
      FileNotFoundError: [Errno 2] No such file or directory: 'llvm-config'

      Traceback (most recent call last):
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 191, in <module>
          main()
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 185, in main
          main_posix('osx', '.dylib')
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 107, in main_posix
          raise RuntimeError("%s failed executing, please point LLVM_CONFIG "
      RuntimeError: llvm-config failed executing, please point LLVM_CONFIG to the path for llvm-config
      error: command '/Users/jryon/miniconda3/envs/autocti/bin/python' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for llvmlite
  Running setup.py clean for llvmlite
Failed to build numba llvmlite
Installing collected packages: python-dateutil, llvmlite, kiwisolver, Jinja2, imagesize, idna, h5py, future, fonttools, exceptiongroup, emcee, docutils, dill, decorator, cycler, contourpy, charset-normalizer, babel, attrs, astunparse, requests, pytest, pandas, numba, matplotlib, autoconf, sphinx, seaborn, pyswarms, dynesty, corner, autoarray, zeus-mcmc, numpydoc, autofit, autocti
  Running setup.py install for llvmlite ... error
  error: subprocess-exited-with-error

  × Running setup.py install for llvmlite did not run successfully.
  │ exit code: 1
  ╰─> [31 lines of output]
      running install
      /Users/jryon/miniconda3/envs/autocti/lib/python3.10/site-packages/setuptools/command/install.py:34: SetuptoolsDeprecationWarning: setup.py install is deprecated. Use build and pip and other standards-based tools.
        warnings.warn(
      running build
      got version from file /private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/llvmlite/_version.py {'version': '0.34.0', 'full': 'c5889c9e98c6b19d5d85ebdd982d64a03931f8e2'}
      running build_ext
      /Users/jryon/miniconda3/envs/autocti/bin/python /private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py
      LLVM version... Traceback (most recent call last):
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 105, in main_posix
          out = subprocess.check_output([llvm_config, '--version'])
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 421, in check_output
          return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 503, in run
          with Popen(*popenargs, **kwargs) as process:
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 971, in __init__
          self._execute_child(args, executable, preexec_fn, close_fds,
        File "/Users/jryon/miniconda3/envs/autocti/lib/python3.10/subprocess.py", line 1847, in _execute_child
          raise child_exception_type(errno_num, err_msg, err_filename)
      FileNotFoundError: [Errno 2] No such file or directory: 'llvm-config'

      Traceback (most recent call last):
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 191, in <module>
          main()
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 185, in main
          main_posix('osx', '.dylib')
        File "/private/var/folders/5y/70tct30n5m92rcy0h5hqzg2m0000gn/T/pip-install-hke2ivpc/llvmlite_6f7e6b5d09bc4a30b4666d2f147a048e/ffi/build.py", line 107, in main_posix
          raise RuntimeError("%s failed executing, please point LLVM_CONFIG "
      RuntimeError: llvm-config failed executing, please point LLVM_CONFIG to the path for llvm-config
      error: command '/Users/jryon/miniconda3/envs/autocti/bin/python' failed with exit code 1
      [end of output]

error: legacy-install-failure

× Encountered error while trying to install package.
╰─> llvmlite

note: This is an issue with the package mentioned above, not pip.
hint: See above for output from the failure.
