### 2022-12-20 Jammy2211/PyAutoCTI#31 @Jammy2211
Ok cool, `numba` is optional (may not have been on the version you are installing) so should be simple enough to fix.

I'll sort a release and let you know when I think everything should work...
