### 2022-12-20 a6e88b5
data.rst now describe charge injection data
