+++ docs/api/clocking.rst
@@ -1,0 +1,31 @@
+========
+Clocking
+========
+
+--------
+Clockers
+--------
+
+Performs a full clocking and read out of a CCD, including CTI.
+
+This object effectiely acts as a wrapper to arCTIc.
+
+.. currentmodule:: autocti
+
+.. autosummary::
+   :toctree: generated/
+
+    Clocker1D
+    Clocker2D
+
+--------------------------
+Read Out Electronics (ROE)
+--------------------------
+
+The read-out electronics of the CCD, describing how electrons are clocked over the CCD.
+
+.. autosummary::
+   :toctree: generated/
+
+    ROE
+    ROEChargeInjection
