### 2022-12-20 c836bda
abstract clocker types
