### 2022-12-20 b8bc9d8
extract typing
