### 2022-12-29 c656987
remove Mask2D.manual / Mask1D.manual
