### 2023-01-09 #4904 @jryon
Thanks for the reminder, @adonath. I have been waiting for a new release from the authors that should fix my installation issues: https://github.com/Jammy2211/PyAutoCTI/issues/31
