+++ test_autocti/extract/two_d/test_master.py
@@ -6,8 +6,8 @@ def test__regions_array_2d_from():
 
     extract = ac.Extract2DMaster(region_list=[(0, 3, 0, 3)])
 
-    array = ac.Array2D.manual(
-        array=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
+    array = ac.Array2D.no_mask(
+        values=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
         pixel_scales=1.0,
     )
 
@@ -22,8 +22,8 @@ def test__regions_array_2d_from():
 
     extract = ac.Extract2DMaster(region_list=[(0, 1, 1, 2), (2, 3, 1, 3)])
 
-    array = ac.Array2D.manual(
-        array=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
+    array = ac.Array2D.no_mask(
+        values=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
         pixel_scales=1.0,
     )
 
@@ -41,8 +41,8 @@ def test__non_regions_array_2d_from():
 
     extract = ac.Extract2DMaster(region_list=[(0, 3, 0, 3)])
 
-    array = ac.Array2D.manual(
-        array=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
+    array = ac.Array2D.no_mask(
+        values=[[0.0, 1.0, 2.0], [3.0, 4.0, 5.0], [6.0, 7.0, 8.0], [9.0, 10.0, 11.0]],
         pixel_scales=1.0,
     )
 
@@ -57,8 +57,8 @@ def test__non_regions_array_2d_from():
 
     extract = ac.Extract2DMaster(region_list=[(0, 1, 0, 3), (3, 4, 0, 3)])
 
-    array = ac.Array2D.manual(
-        array=[
+    array = ac.Array2D.no_mask(
+        values=[
             [0.0, 1.0, 2.0],
             [3.0, 4.0, 5.0],
             [6.0, 7.0, 8.0],
@@ -86,8 +86,8 @@ def test__non_regions_array_2d_from():
 
 def test__parallel_fpr_and_eper_array_2d_from():
 
-    parallel_array = ac.Array2D.manual(
-        array=[
+    parallel_array = ac.Array2D.no_mask(
+        values=[
             [0.0, 0.0, 0.0],
             [1.0, 1.0, 1.0],  # <- Front edge .
             [2.0, 2.0, 2.0],  # <- Next front edge row.
@@ -155,8 +155,8 @@ def test__serial_fpr_and_eper_array_2d_from():
 
     extract = ac.Extract2DMaster(region_list=[(0, 3, 0, 3)])
 
-    array = ac.Array2D.manual(
-        array=[[0.0, 1.0, 2.0, 3.0], [4.0, 5.0, 6.0, 7.0], [8.0, 9.0, 10.0, 11.0]],
+    array = ac.Array2D.no_mask(
+        values=[[0.0, 1.0, 2.0, 3.0], [4.0, 5.0, 6.0, 7.0], [8.0, 9.0, 10.0, 11.0]],
         pixel_scales=1.0,
     )
 
@@ -202,8 +202,8 @@ def test__serial_fpr_and_eper_array_2d_from():
 
[…]
