### 2023-01-15 Jammy2211/PyAutoCTI#31 @Jammy2211
`pip install autocti-no-gsl`

Let me know if you have more issues :)
