### 2023-01-16 01b61ee
efficiency by not producing native array unecessarily
