+++ autocti/config/general.yaml
@@ -7,7 +7,6 @@ model:
   ignore_prior_limits: false
 output:
   force_pickle_overwrite: false
-  identifier_version: 4
   info_whitespace_length: 80
   log_file: output.log
   log_level: INFO
