### 2023-01-19 #4904 @adonath
@jryon Just wanted to bring to your attention https://github.com/Jammy2211/PyAutoCTI/issues/31#issuecomment-1383232019
