### 2023-01-20 Jammy2211/PyAutoCTI#31 @jryon
Sorry, spoke too soon. Installation worked, but I ran into an error with `import autocti`:

Python 3.10.8 | packaged by conda-forge | (main, Nov 22 2022, 08:27:35) [Clang 14.0.6 ] on darwin
>>> import autocti as ac
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/__init__.py", line 19, in <module>
    from .charge_injection.fit import FitImagingCI
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/charge_injection/fit.py", line 8, in <module>
    from autocti.charge_injection.imaging.imaging import ImagingCI
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/charge_injection/imaging/imaging.py", line 6, in <module>
    from autocti.charge_injection.layout import Layout2DCI
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/charge_injection/layout.py", line 14, in <module>
    from autocti.layout.two_d import Layout2D
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/layout/two_d.py", line 7, in <module>
    from autocti import exc
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autocti/exc.py", line 1, in <module>
    from autofit import exc
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/__init__.py", line 6, in <module>
    from .non_linear.grid.grid_search import GridSearch as SearchGridSearch
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/non_linear/grid/grid_search/__init__.py", line 10, in <module>
    from autofit.non_linear.abstract_search import NonLinearSearch
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/non_linear/abstract_search.py", line 18, in <module>
    from autofit.graphical import (
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/__init__.py", line 2, in <module>
    from .declarative.abstract import PriorFactor
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/declarative/abstract.py", line 7, in <module>
    from autofit.graphical.declarative.graph import DeclarativeFactorGraph
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/declarative/graph.py", line 5, in <module>
    from autofit.graphical.expectation_propagation.ep_mean_field import EPMeanField
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/expectation_propagation/__init__.py", line 3, in <module>
    from .optimiser import AbstractFactorOptimiser
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/expectation_propagation/optimiser.py", line 20, in <module>
    from .visualise import Visualise
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/autofit/graphical/expectation_propagation/visualise.py", line 5, in <module>
    import matplotlib.pyplot as plt
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/matplotlib/__init__.py", line 113, in <module>
    from . import _api, _version, cbook, _docstring, rcsetup
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/matplotlib/rcsetup.py", line 27, in <module>
    from matplotlib.colors import Colormap, is_color_like
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/matplotlib/colors.py", line 51, in <module>
    from PIL import Image
  File "/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/PIL/Image.py", line 100, in <module>
    from . import _imaging as core
ImportError: dlopen(/Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/PIL/_imaging.cpython-310-darwin.so, 2): Library not loaded: @rpath/libtiff.5.dylib
  Referenced from: /Users/ryon/miniconda3/envs/autocti/lib/python3.10/site-packages/PIL/_imaging.cpython-310-darwin.so
  Reason: image not found
