### 2023-01-20 Jammy2211/PyAutoCTI#31 @rhayes777
Sounds like a missing dependency.

Try:
```
conda install libtiff
