### 2023-01-20 Jammy2211/PyAutoCTI#31 @jryon
@rhayes777 Thanks, I tried installing `libtiff`, but then got another error related to `numba` (which I'd installed after `autocti-no-gsl`). I decided to start from scratch with a fresh environment, and ran into a few issues related to python versions. The versions of scipy required for `autocti-no-gsl` appear to be incompatible with python 3.11, so I chose (rather arbitrarily) python 3.9, and the install proceeded smoothly. The `welcome.py` script executed without any trouble. Here are the steps I used:

`conda create -n autocti python=3.9`
`conda activate autocti`
`pip install arcticpy_no_gsl`
`pip install autocti-no-gsl`
`pip install numba`
