### 2023-01-21 Jammy2211/PyAutoCTI#31 @rhayes777
Great so everything is working fine? I don't think we support Python 3.11 @Jammy2211 ? It might be worth us making that clear in the documentation
