### 2023-01-23 Jammy2211/PyAutoCTI#31 @jryon
Yes, I've now run several of the notebooks in the autocti-workspace and everything looks good. Thanks for your help!
