### 2023-01-23 #4904 @jryon
Hi @Jammy2211, Thanks for a great software package! PyAutoCTI clearly represents a substantial amount of work and looks to be a powerful tool for calibrating CTI for new and existing space telescopes. I'd be happy to recommend this submission for publication in JOSS once a few items are addressed.

General:

- The SQLite3 database is highlighted in the paper, but there is little documentation or examples of its usage. Database examples in the `autocti-workspace` would be very helpful, for instance. The readthedocs page on databases is helpful, but the code blocks need import statements (`import autofit as af` and `from autofit import Aggregator`?). Incorporating database usage into the other examples in the documentation (within the Overview section, perhaps) would be useful as well.
- Link to HowToCTI notebook tutorials in the repo readme is broken. I saw references to HowToCTI lectures (in `autocti-workspace` notebooks) as well, but I couldn't find anything with the name HowToCTI - could these materials have been renamed?

Paper:

- In the Statement of Need, please briefly state the intended audience of this package, e.g., users of future space telescopes and their data, support staff for those telescopes, laboratory scientists, or others?
- A short section describing how this software compares to other available packages would be useful. Related software packages are listed, but I think brief discussion is needed. For example, which telescopes are they intended for? Do they produce CTI models, simulate CTI effects, correct for CTI, or etc? Are they open source?
- I'd recommend including additional references for the ACS/WFC and WFC3/UVIS pixel-based CTI correction algorithms, since both have been updated since the Anderson & Bedin 2010 paper:
	- https://ui.adsabs.harvard.edu/abs/2018acs..rept....4A/abstract
	- https://ui.adsabs.harvard.edu/abs/2021wfc..rept....9A/abstract

Documentation:

- I recommend a careful review of the readthedocs pages for typos, page organization, and broken figures:
	- CCD and CTI should always be capitalized
	- Check usage of "electrons" and "pixels" - see https://pyautocti.readthedocs.io/en/latest/overview/overview_2_parallel_and_serial.html#two-dimensional-ccd
	- Content on the [Home page](https://pyautocti.readthedocs.io/en/latest/index.html#) is repeated in top half of the ["What is CTI?" page](https://pyautocti.readthedocs.io/en/latest/overview/overview_1_what_is_cti.html). Was this intended? I found it a bit confusing when navigating the pages
	- Last two figures (gifs of CCD clocking) on the Home page are in the wrong order
	- Several figures are broken on pages in the Overview section
- In the Installation section, please include the "no gsl" package options, and explicitly say Python 3.11 is not supported.
- In the API, the base classes for some of the objects appear to be from PyAutoFit. It may be useful to users to make that more explicit, or add links to the base classes where they appear.
- Check Binder links in the repo readme and readthedocs pages. One looks to be pointing to a nonexistent notebook, and another took me to the PyAutoFit workspace.

autocti_workspace:

- The workspace appears to have been reorganized since the READMEs were written. For example, there are no `results` directories within `notebooks` or `scripts`, and no `database` directory. I think reviewing the READMEs would be helpful.
