### 2023-01-26 90aa836
added methods for 1D plots of regions for Dataset1D
