### 2023-01-26 de0c44a
added data_with_nosie_map visuals to ImagingcI
