+++ autocti/charge_injection/plot/imaging_ci_plotters.py
@@ -101,21 +101,21 @@ class ImagingCIPlotter(Plotter):
         Parameters
         ----------
         image
-            Whether or not to make a 2D plot (via `imshow`) of the image data.
+            Whether to make a 2D plot (via `imshow`) of the image data.
         noise_map
-            Whether or not to make a 2D plot (via `imshow`) of the noise map.
+            Whether to make a 2D plot (via `imshow`) of the noise map.
         inverse_noise_map
-            Whether or not to make a 2D plot (via `imshow`) of the inverse noise map.
+            Whether to make a 2D plot (via `imshow`) of the inverse noise map.
         signal_to_noise_map
-            Whether or not to make a 2D plot (via `imshow`) of the signal-to-noise map.
+            Whether to make a 2D plot (via `imshow`) of the signal-to-noise map.
         absolute_signal_to_noise_map
-            Whether or not to make a 2D plot (via `imshow`) of the absolute signal to noise map.
+            Whether to make a 2D plot (via `imshow`) of the absolute signal to noise map.
         potential_chi_squared_map
-            Whether or not to make a 2D plot (via `imshow`) of the potential chi squared map.
+            Whether to make a 2D plot (via `imshow`) of the potential chi squared map.
         pre_cti_data
-            Whether or not to make a 2D plot (via `imshow`) of the pre-cti data.
+            Whether to make a 2D plot (via `imshow`) of the pre-cti data.
         cosmic_ray_map
-            Whether or not to make a 2D plot (via `imshow`) of the cosmic ray map.
+            Whether to make a 2D plot (via `imshow`) of the cosmic ray map.
         """
 
         self._imaging_meta_plotter.figures_2d(
@@ -154,6 +154,8 @@ class ImagingCIPlotter(Plotter):
         noise_map: bool = False,
         pre_cti_data: bool = False,
         signal_to_noise_map: bool = False,
+        data_with_noise_map: bool = False,
+        data_with_noise_map_logy: bool = False,
     ):
         """
         Plots the individual attributes of the plotter's `ImagingCI` object in 1D.
@@ -172,14 +174,20 @@ class ImagingCIPlotter(Plotter):
             The region on the charge injection image where data is extracted and binned over the parallel or serial
             direction {"parallel_fpr", "parallel_eper", "serial_fpr", "serial_eper"}
         image
-            Whether or not to make a 1D plot (via `plot`) of the image data extracted and binned over the region.
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region.
         noise_map
-            Whether or not to make a 1D plot (via `plot`) of the noise-map extracted and binned over the region.
+            Whether to make a 1D plot (via `plot`) of the noise-map extracted and binned over the region.
         pre_cti_data
-            Whether or not to make a 1D plot (via `plot`) of the pre-cti data extracted and binned over the region.
+            Whether to make a 1D plot (via `plot`) of the pre-cti data extracted and binned over the region.
         signal_to_noise_map
-            Whether or not to make a 1D plot (via `plot`) of the signal-to-noise map data extracted and binned over
+            Whether to make a 1D plot (via `plot`) of the signal-to-noise map data extracted and binned over
             the region.
+        data_with_noise_map
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars.
+        data_with_noise_map_logy
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars and the y-axis on a log10 scale.
         """
 
         if image:
@@ -191,8 +199,8 @@ class ImagingCIPlotter(Plotter):
[…]
