### 2023-01-26 fd48c20
noise map 1D plot bins correctly
