+++ autocti/charge_injection/model/visualizer.py
@@ -123,6 +123,8 @@ class VisualizerImagingCI(Visualizer):
             residual_map=should_plot("residual_map"),
             normalized_residual_map=should_plot("normalized_residual_map"),
             chi_squared_map=should_plot("chi_squared_map"),
+            data_with_noise_map_model=should_plot("data_with_noise_map_model"),
+            data_with_noise_map_model_logy=should_plot("data_with_noise_map_model_logy"),
         )
 
         if not during_analysis:
