+++ autocti/charge_injection/plot/fit_ci_plotters.py
@@ -77,6 +77,10 @@ class FitImagingCIPlotter(Plotter):
     def extract_region_from(self) -> Callable:
         return self.fit.imaging_ci.layout.extract_region_from
 
+    @property
+    def extract_region_noise_map_from(self) -> Callable:
+        return self.fit.imaging_ci.layout.extract_region_noise_map_from
+
     def figures_2d(
         self,
         image: bool = False,
@@ -153,6 +157,8 @@ class FitImagingCIPlotter(Plotter):
         residual_map: bool = False,
         normalized_residual_map: bool = False,
         chi_squared_map: bool = False,
+        data_with_noise_map_model: bool = False,
+        data_with_noise_map_model_logy: bool = False,
     ):
         """
         Plots the individual attributes of the plotter's `FitImagingCI` object in 1D.
@@ -190,6 +196,12 @@ class FitImagingCIPlotter(Plotter):
         chi_squared_map
             Whether to make a 1D plot (via `plot`) of the chi-squared map extracted and binned over the
             region.
+        data_with_noise_map_model
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars and the model-fit overlaid.
+        data_with_noise_map_model_logy
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars and the y-axis on a log10 scale and the model-fit overlaid.
         """
         if image:
 
@@ -320,6 +332,48 @@ class FitImagingCIPlotter(Plotter):
                     ylabel="Image",
                     xlabel="Pixel No.",
                     filename=f"chi_squared_map_{region}",
+                ),
+            )
+
+        if data_with_noise_map_model:
+
+            y = self.extract_region_from(array=self.fit.data, region=region)
+            y_errors = self.extract_region_noise_map_from(array=self.fit.noise_map, region=region)
+            y_extra = self.extract_region_from(array=self.fit.model_data, region=region)
+
+            self.mat_plot_1d.plot_yx(
+                y=y,
+                x=range(len(y)),
+                plot_axis_type_override="errorbar",
+                y_errors=y_errors,
+                y_extra=y_extra,
+                visuals_1d=self.visuals_1d,
+                auto_labels=AutoLabels(
+                    title=f"Data 1D With Noise & Model-Fit [{region}]",
+                    ylabel="Data (e-)",
+                    xlabel="Pixel No.",
+                    filename=f"data_with_noise_map_{region}",
+                ),
+            )
+
+        if data_with_noise_map_model_logy:
+
+            y = self.extract_region_from(array=self.fit.data, region=region)
+            y_errors = self.extract_region_noise_map_from(array=self.fit.noise_map, region=region)
+            y_extra = self.extract_region_from(array=self.fit.model_data, region=region)
+
+            self.mat_plot_1d.plot_yx(
+                y=y,
+                x=range(len(y)),
+                plot_axis_type_override="errorbar_logy",
+                y_errors=y_errors,
+                y_extra=y_extra,
+                visuals_1d=self.visuals_1d,
+                auto_labels=AutoLabels(
+                    title=f"Data 1D With Noise & Model-Fit (log10 y axis) [{region}]",
+                    ylabel="Data (e-)",
+                    xlabel="Pixel No.",
+                    filename=f"data_with_noise_map_logy_{region}",
                 ),
             )
 
