### 2023-01-27 22ab785
unit test on case with no CTI, but CRs added
