+++ test_autocti/charge_injection/imaging/test_simulator_ci.py
@@ -266,6 +266,14 @@ def test__include_cosmics__is_added_to_image_and_trailed(
     ).all()
     assert imaging.layout == layout
 
+    imaging = simulator.via_layout_from(
+        layout=layout,
+        clocker=None,
+        cti=None,
+        cosmic_ray_map=cosmic_ray_map,
+    )
+
+    assert imaging.image[2, 2] > 94.0
 
 def test__from_pre_cti_data(parallel_clocker_2d, traps_x2, ccd):
 
