+++ autocti/extract/two_d/parallel/abstract.py
@@ -17,6 +17,26 @@ class Extract2DParallel(Extract2D):
         """
         return 1
 
+    def _value_list_from(self, array: aa.Array2D, pixels: Tuple[int, int], value:str):
+
+        value_list = []
+
+        arr_list = [
+            np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
+            for region in self.region_list_from(pixels=pixels)
+        ]
+
+        arr_stack = np.ma.stack(arr_list)
+
+        for column_index in range(arr_list[0].shape[1]):
+
+            if value == "median":
+                value_list.append(float(np.ma.median(arr_stack[:, :, column_index])))
+            elif value == "sigma":
+                value_list.append(float(np.ma.std(arr_stack[:, :, column_index])))
+
+        return value_list
+
     def median_list_from(
         self, array: aa.Array2D, pixels: Tuple[int, int]
     ) -> List[float]:
@@ -55,20 +75,47 @@ class Extract2DParallel(Extract2D):
             FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
             as `pixels=(10, 30)` would be used.
         """
-        median_list = []
+        return self._value_list_from(array=array, pixels=pixels, value="median")
 
-        arr_list = [
-            np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
-            for region in self.region_list_from(pixels=pixels)
-        ]
+    def sigma_list_from(
+        self, array: aa.Array2D, pixels: Tuple[int, int]
+    ) -> List[float]:
+        """
+        Returns the standard deviation sigma values of the `Extract2D` object's corresponding region, where the
+        standard deviation is taken over all columns of the region(s).
 
-        arr_stack = np.ma.stack(arr_list)
+        To describe this function we will use the example of estimating the noise of the charge lines in
+        charge injection data, by taking the standard deviation over every individual column of charge
+        injection (e.g. aligned with the FPR).
 
-        for column_index in range(arr_list[0].shape[1]):
+        The inner regions of the FPR of each charge injection line informs us of the noise in the injected level of
+        charge for that injection.
 
-            median_list.append(float(np.ma.median(arr_stack[:, :, column_index])))
+        By taking the standard deviation of values of the charge injection regions (after accounting for those which
+        have had electrons captured and relocated due to CTI), we can therefore estimate the input charge injection
+        normalization.
 
-        return median_list
+        This function does this for every column of every charge injection. If multiple charge injections are performed
+        per column, the standard deviation of all charge regions is taken.
+
+        For example, if there are 3 charge injection regions, this function returns a list where each value
+        estimates the charge injection noise of a given charge injection region. The size of this list
+        is therefore the number of charge injection columns.
+
+        The function `sigma_list_of_lists_from` performs the standard deviation over each individual charge injection
+        region in each column. It therefore estimates multiple injection noise per column. for each individual
+        charge region. Which function one uses depends on the properties of the charge injection on the instrumentation.
+
+        Parameters
+        ----------
+        array
+            The charge injection image from which the charge injection standard deviations / noise are estimated.
+        pixels
+            The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
+            FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
+            as `pixels=(10, 30)` would be used.
+        """
+        return self._value_list_from(array=array, pixels=pixels, value="sigma")
 
     def median_list_of_lists_from(
         self, array: aa.Array2D, pixels: Tuple[int, int]
[…]
