### 2023-01-31 dbd0e1c
implemented Extract2DParallel.sigma_list_of_lists_from
