### 2023-01-31 8aae6e6
not implementing serial - noise is in parallel direction
