### 2023-01-31 bea1e57
implemented noise_map_non_uniform_from
