### 2023-02-08 7815845
added method pixels_within_region
