+++ autocti/charge_injection/plot/imaging_ci_plotters.py
@@ -1,3 +1,4 @@
+import numpy as np
 from typing import Callable
 
 import autoarray.plot as aplt
@@ -262,6 +263,11 @@ class ImagingCIPlotter(Plotter):
 
         if data_with_noise_map:
 
+            # injection_norm_list = self.dataset.layout.extract.parallel_fpr.median_list_from(
+            #     array=data, pixels=(injection_on - 20, injection_on))
+            #
+            # fpr_value = np.mean
+            # )
             y = self.extract_region_from(array=self.dataset.data, region=region)
             y_errors = self.extract_region_noise_map_from(
                 array=self.dataset.noise_map, region=region
