### 2023-02-08 67ee4e1
rename pixels_within methods
