+++ autocti/layout/one_d.py
@@ -43,7 +43,7 @@ class Layout1D(aa.Layout1D):
         super().__init__(shape_1d=shape_1d, prescan=prescan, overscan=overscan)
 
     @property
-    def pixels_between_regions(self):
+    def parallel_rows_between_regions(self):
         return [
             self.region_list[i + 1].x0 - self.region_list[i].x1
             for i in range(len(self.region_list) - 1)
@@ -56,9 +56,9 @@ class Layout1D(aa.Layout1D):
     @property
     def smallest_eper_pixels_to_array_edge(self):
 
-        pixels_between_regions = self.pixels_between_regions
-        pixels_between_regions.append(self.trail_size_to_array_edge)
-        return np.min(pixels_between_regions)
+        parallel_rows_between_regions = self.parallel_rows_between_regions
+        parallel_rows_between_regions.append(self.trail_size_to_array_edge)
+        return np.min(parallel_rows_between_regions)
 
     def extract_region_from(self, array: aa.Array1D, region: str):
 
