### 2023-02-08 66d2e62
smallest_parallel_rows_within_ci_regions
