### 2023-02-08 ada1579
added pixels_from_end to median_list_from for parallel FPR
