+++ autocti/extract/two_d/parallel/fpr.py
@@ -1,4 +1,4 @@
-from typing import List, Tuple
+from typing import List, Optional, Tuple
 
 import autoarray as aa
 
@@ -8,7 +8,9 @@ from autocti.extract.two_d import extract_2d_util
 
 
 class Extract2DParallelFPR(Extract2DParallel):
-    def region_list_from(self, pixels: Tuple[int, int]) -> List[aa.Region2D]:
+    def region_list_from(
+        self, pixels: Optional[Tuple[int, int]] = None, pixels_from_end: Optional[int] = None
+    ) -> List[aa.Region2D]:
         """
          Returns a list of the 2D parallel FPR regions from the `region_list` containing signal  (e.g. the charge
          injection regions of charge injection data), extracted between two input `pixels` indexes.
@@ -67,7 +69,9 @@ class Extract2DParallelFPR(Extract2DParallel):
              corresponding to the 1st, 2nd and 3rd FPR rows).
         """
         return [
-            region.parallel_front_region_from(pixels=pixels)
+            region.parallel_front_region_from(
+                pixels=pixels, pixels_from_end=pixels_from_end
+            )
             for region in self.region_list
         ]
 
