+++ autocti/extract/two_d/parallel/eper.py
@@ -10,50 +10,50 @@ from autocti.extract.two_d import extract_2d_util
 class Extract2DParallelEPER(Extract2DParallel):
     def region_list_from(self, pixels: Tuple[int, int]) -> List[aa.Region2D]:
         """
-          Returns a list of the 2D parallel EPER regions from the `region_list` containing signal  (e.g. the charge
-          injection regions of charge injection data), extracted between two input `pixels` indexes.
+        Returns a list of the 2D parallel EPER regions from the `region_list` containing signal  (e.g. the charge
+        injection regions of charge injection data), extracted between two input `pixels` indexes.
 
-          Negative pixel values are supported to the `pixels` tuple, whereby columns in front of the parallel EPERs
-          (e.g. the FPRs) are also extracted.
+        Negative pixel values are supported to the `pixels` tuple, whereby columns in front of the parallel EPERs
+        (e.g. the FPRs) are also extracted.
 
-          The diagram below illustrates the extraction for `pixels=(0, 1)`:
+        The diagram below illustrates the extraction for `pixels=(0, 1)`:
 
-          [] = read-out electronics
-          [==========] = read-out register
-          [..........] = serial prescan
-          [pppppppppp] = parallel overscan
-          [ssssssssss] = serial overscan
-          [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
-          [tttttttttt] = parallel / serial charge injection region trail
+        [] = read-out electronics
+        [==========] = read-out register
+        [..........] = serial prescan
+        [pppppppppp] = parallel overscan
+        [ssssssssss] = serial overscan
+        [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
+        [tttttttttt] = parallel / serial charge injection region trail
 
-                 [ppppppppppppppppppppp]
-                 [ppppppppppppppppppppp]
-            [...][t1t1t1t1t1t1t1t1t1t1t][sss]
-            [...][c1c1cc1c1cc1cc1ccc1cc][sss]
-          | [...][1c1c1cc1c1cc1ccc1cc1c][sss]     |
-          | [...][t0t0t0t0t0t0t0t0t0t0t][sss]    | Direction
+               [ppppppppppppppppppppp]
+               [ppppppppppppppppppppp]
+          [...][t1t1t1t1t1t1t1t1t1t1t][sss]
+          [...][c1c1cc1c1cc1cc1ccc1cc][sss]
+        | [...][1c1c1cc1c1cc1ccc1cc1c][sss]     |
+        | [...][t0t0t0t0t0t0t0t0t0t0t][sss]    | Direction
         Par [...][0t0t0t0t0t0t0t0t0t0t0][sss]    | of
-          | [...][0ccc0cccc0cccc0cccc0c][sss]    | clocking
+        | [...][0ccc0cccc0cccc0cccc0c][sss]    | clocking
          |/ [...][cc0ccc0cccc0cccc0cccc][sss]   \/
 
-          []     [=====================]
-                 <---------Ser----------
+        []     [=====================]
+               <---------Ser----------
 
-          The extracted regions correspond to the first parallel EPER all charge injection regions:
+        The extracted regions correspond to the first parallel EPER all charge injection regions:
 
-          region_list[0] = [2, 4, 3, 21] (serial prescan is 3 pixels)
-          region_list[1] = [6, 7, 3, 21] (serial prescan is 3 pixels)
+        region_list[0] = [2, 4, 3, 21] (serial prescan is 3 pixels)
+        region_list[1] = [6, 7, 3, 21] (serial prescan is 3 pixels)
 
[…]

[…]
