### 2023-02-08 7b8aa4d
pixels_from_end added to std_list_from in parallel
