+++ autocti/extract/two_d/parallel/abstract.py
@@ -49,8 +49,6 @@ class Extract2DParallel(Extract2D):
                 )
             ]
 
-            print(arr_list)
-
         arr_stack = np.ma.stack(arr_list)
 
         for column_index in range(arr_list[0].shape[1]):
@@ -99,9 +97,13 @@ class Extract2DParallel(Extract2D):
         array
             The charge injection image from which the charge injection normalizations are estimated.
         pixels
-            The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
-            FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
-            as `pixels=(10, 30)` would be used.
+            The row pixel index to extract the region (e.g. FPR / EPER) between. For example, `pixels=(0, 3)`
+            extracts the 1st, 2nd and 3rd FPR / EPER rows). To remove the 10 leading pixels which have lost electrons
+            due to CTI, `pixels=(10, 30)` would be used.
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100) are extracted.
         """
         return self._value_list_from(
             array=array,
@@ -110,7 +112,10 @@ class Extract2DParallel(Extract2D):
             value_str="median",
         )
 
-    def std_list_from(self, array: aa.Array2D, pixels: Tuple[int, int]) -> List[float]:
+    def std_list_from(self, array: aa.Array2D,
+                      pixels: Optional[Tuple[int, int]] = None,
+                      pixels_from_end: Optional[int] = None,
+                      ) -> List[float]:
         """
         Returns the standard deviation (sigma) values of the `Extract2D` object's corresponding region, where the
         standard deviation is taken over all columns of the region(s).
@@ -146,7 +151,7 @@ class Extract2DParallel(Extract2D):
             FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
             as `pixels=(10, 30)` would be used.
         """
-        return self._value_list_from(array=array, pixels=pixels, value_str="std")
+        return self._value_list_from(array=array, pixels=pixels, pixels_from_end=pixels_from_end, value_str="std")
 
     def _value_list_of_lists_from(
         self, array: aa.Array2D, pixels: Tuple[int, int], value_str: str
