+++ test_autocti/extract/two_d/parallel/test_abstract.py
@@ -88,6 +88,42 @@ def test__median_list_of_lists_from(parallel_array, parallel_masked_array):
     assert median_list_of_lists[1] == [5.5, 6.0, 6.0]
 
 
+def test__median_list_of_lists_from__pixels_from_end(
+    parallel_array, parallel_masked_array
+):
+
+    extract = ac.Extract2DParallelFPR(region_list=[(1, 4, 0, 3)])
+
+    # Extracts [1.0, 2.0, 3.0] of every injection in `parallel_array`
+
+    median_list_of_lists = extract.median_list_of_lists_from(
+        array=parallel_array, pixels_from_end=3
+    )
+    assert median_list_of_lists[0] == [2.0, 2.0, 2.0]
+
+    # Reduce pixels to only extract 1.0.
+
+    median_list_of_lists = extract.median_list_of_lists_from(
+        array=parallel_array, pixels_from_end=1
+    )
+    assert median_list_of_lists[0] == [3.0, 3.0, 3.0]
+
+    extract = ac.Extract2DParallelFPR(region_list=[(1, 4, 0, 3), (5, 8, 0, 3)])
+
+    median_list_of_lists = extract.median_list_of_lists_from(
+        array=parallel_array, pixels_from_end=2
+    )
+
+    assert median_list_of_lists[0] == [2.5, 2.5, 2.5]
+    assert median_list_of_lists[1] == [6.5, 6.5, 6.5]
+
+    median_list_of_lists = extract.median_list_of_lists_from(
+        array=parallel_masked_array, pixels_from_end=2
+    )
+    assert median_list_of_lists[0] == [2.5, 3.0, 2.0]
+    assert median_list_of_lists[1] == [6.0, 6.5, 6.5]
+
+
 def test__std_list_from(parallel_array, parallel_masked_array):
 
     extract = ac.Extract2DParallelFPR(region_list=[(1, 4, 0, 3)])
