### 2023-02-08 0e7637d
test on pixels_from_end in std_list_of_lists
