### 2023-02-08 6b1ca4f
fix tests due to pixels_from_end changes
