+++ test_autocti/extract/two_d/parallel/test_parallel_fpr.py
@@ -53,7 +53,8 @@ def test__region_list_from__via_array_2d_list_from(
 def test__region_list_from__via_array_2d_list_from__pixels_from_end(
     parallel_array, parallel_masked_array
 ):
-    extract = ac.Extract2DParallelFPR(region_list=[(1, 4, 0, 3)])
+
+    extract = ac.Extract2DParallelFPR(region_list=[(1, 4, 0, 3)], shape_2d=(5, 5))
 
     fpr_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=1)
     assert (fpr_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
