+++ autocti/extract/two_d/parallel/calibration.py
@@ -277,6 +277,7 @@ class Extract2DParallelCalibration:
             ),
             cosmic_ray_map=cosmic_ray_map,
             noise_scaling_map_dict=noise_scaling_map_dict,
+            fpr_value=imaging_ci.fpr_value
         )
 
         return imaging_ci.apply_mask(mask=mask)
