### 2023-02-09 4259be6
pixels_from_end in Extract2DParallelOverscan
