+++ test_autocti/extract/two_d/parallel/test_parallel_overscan.py
@@ -34,6 +34,31 @@ def test_region_list_from__via_array_2d_list_from(
     ).all()
 
 
+def test_region_list_from__via_array_2d_list_from__pixels_from_end(
+    parallel_array, parallel_masked_array
+):
+
+    extract = ac.Extract2DParallelOverscan(parallel_overscan=(1, 4, 0, 3))
+
+    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=1)
+    assert (overscan_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
+
+    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=3)
+    assert (
+        overscan_list[0]
+        == np.array([[1.0, 1.0, 1.0], [2.0, 2.0, 2.0], [3.0, 3.0, 3.0]])
+    ).all()
+
+    overscan_list = extract.array_2d_list_from(
+        array=parallel_masked_array, pixels_from_end=3
+    )
+
+    assert (
+        overscan_list[0].mask
+        == np.array([[False, False, False], [False, True, False], [False, False, True]])
+    ).all()
+
+
 def test__binned_region_1d_from():
 
     extract = ac.Extract2DParallelOverscan(region_list=[(1, 3, 0, 3)])
