### 2023-02-09 a69024a
added pixels_from_end to Extract2DSerialEPER
