+++ autocti/extract/two_d/serial/eper.py
@@ -57,8 +57,19 @@ class Extract2DSerialEPER(Extract2DSerial):
          Parameters
          ----------
          pixels
-             The column indexes to extract the trails between (e.g. columns(0, 3) extracts the 1st, 2nd and 3rd columns)
+             The column indexes to extract the trails between (e.g. columns(0, 3) extracts the 1st, 2nd and 3rd
+             columns).
+        pixels_from_end
+            Alternative row pixel index specification, which extracts this number of pixels from the end of
+            the EPER. For example, if each EPER is 100 pixels and `pixels_from_end=10`, the last 10 pixels of each
+            EPER (pixels (90, 100)) are extracted.
         """
+
+        if pixels_from_end is not None:
+            pixels = (
+                self.shape_2d[1] - self.region_list[0].x1 - pixels_from_end,
+                self.shape_2d[1] - self.region_list[0].x1
+            )
         return [
             region.serial_trailing_region_from(pixels=pixels)
             for region in self.region_list
