### 2023-02-09 15bbfc9
added pixels_from_end to Extract2DSerialFPR
