+++ test_autocti/extract/two_d/serial/test_serial_eper.py
@@ -70,11 +70,15 @@ def test__region_list_from__array_2d_list_from(serial_array, serial_masked_array
     ).all()
 
 
-def test__region_list_from__array_2d_list_from__pixels_from_end(serial_array, serial_masked_array):
+def test__region_list_from__array_2d_list_from__pixels_from_end(
+    serial_array, serial_masked_array
+):
 
     print(serial_array.shape_native)
 
-    extract = ac.Extract2DSerialEPER(shape_2d=serial_array.shape_native, region_list=[(0, 3, 1, 4)])
+    extract = ac.Extract2DSerialEPER(
+        shape_2d=serial_array.shape_native, region_list=[(0, 3, 1, 4)]
+    )
 
     eper_list = extract.array_2d_list_from(array=serial_array, pixels_from_end=1)
 
