### 2023-02-09 c1bfbfe
added pixels_from_end to Extract2DSerialOverscan
