### 2023-02-09 92c0466
added pixels_from_end to Extract2DSerialPrescan
