### 2023-02-09 858accd
remove if statement in paralle abstract
