### 2023-02-09 aacf6ff
pixels from end added to Extract2DSerial
