+++ autocti/extract/two_d/serial/abstract.py
@@ -1,5 +1,5 @@
 import numpy as np
-from typing import List, Tuple
+from typing import List, Optional, Tuple
 
 import autoarray as aa
 
@@ -17,13 +17,19 @@ class Extract2DSerial(Extract2D):
         return 0
 
     def _value_list_from(
-        self, array: aa.Array2D, pixels: Tuple[int, int], value_str: str
+        self,
+        array: aa.Array2D,
+        value_str: str,
+        pixels: Optional[Tuple[int, int]] = None,
+        pixels_from_end: Optional[int] = None,
     ):
         median_list = []
 
         arr_list = [
             np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
-            for region in self.region_list_from(pixels=pixels)
+            for region in self.region_list_from(
+                pixels=pixels, pixels_from_end=pixels_from_end
+            )
         ]
 
         arr_stack = np.ma.stack(arr_list)
@@ -36,7 +42,10 @@ class Extract2DSerial(Extract2D):
         return median_list
 
     def median_list_from(
-        self, array: aa.Array2D, pixels: Tuple[int, int]
+        self,
+        array: aa.Array2D,
+        pixels: Optional[Tuple[int, int]] = None,
+        pixels_from_end: Optional[int] = None,
     ) -> List[float]:
         """
         Returns the median values of the `Extract2D` object's corresponding region, where the median is taken over
@@ -69,18 +78,33 @@ class Extract2DSerial(Extract2D):
             The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
             FPR rows). To remove the 10 leading pixels which have lost electrons due to CTI, an input such
             as `pixels=(10, 30)` would be used.
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
         """
-        return self._value_list_from(array=array, pixels=pixels, value_str="median")
+        return self._value_list_from(
+            array=array,
+            value_str="median",
+            pixels=pixels,
+            pixels_from_end=pixels_from_end,
+        )
 
     def _value_list_of_lists_from(
-        self, array: aa.Array2D, pixels: Tuple[int, int], value_str: str
+        self,
+        array: aa.Array2D,
+        value_str: str,
+        pixels: Optional[Tuple[int, int]] = None,
+        pixels_from_end: Optional[int] = None,
     ):
 
         median_lists = []
 
         arr_list = [
             np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
-            for region in self.region_list_from(pixels=pixels)
+            for region in self.region_list_from(
+                pixels=pixels, pixels_from_end=pixels_from_end
+            )
         ]
 
         for array_2d in arr_list:
@@ -99,7 +123,10 @@ class Extract2DSerial(Extract2D):
         return median_lists
 
     def median_list_of_lists_from(
-        self, array: aa.Array2D, pixels: Tuple[int, int]
+        self,
+        array: aa.Array2D,
+        pixels: Optional[Tuple[int, int]] = None,
+        pixels_from_end: Optional[int] = None,
     ) -> List[List]:
         """
         Returns the median values of the `Extract2D` object's corresponding region, where the median is taken over
@@ -133,7 +160,14 @@ class Extract2DSerial(Extract2D):
         pixels
             The row pixel index to extract the region between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
             rows of the region).
+        pixels_from_end
[…]
