### 2023-02-09 7ec4ea9
added pixels_from_end to Extract1DFPR
