+++ autocti/layout/one_d.py
@@ -37,7 +37,10 @@ class Layout1D(aa.Layout1D):
         from autocti.extract.one_d.master import Extract1DMaster
 
         self.extract = Extract1DMaster(
-            region_list=region_list, prescan=prescan, overscan=overscan
+            shape_1d=shape_1d,
+            region_list=region_list,
+            prescan=prescan,
+            overscan=overscan,
         )
 
         super().__init__(shape_1d=shape_1d, prescan=prescan, overscan=overscan)
