### 2023-02-11 278a51a
Added method to Extract2D which adds noise to its regions
