### 2023-02-11 178c028
Added method to Extract1D which adds noise to its regions
