+++ autocti/extract/two_d/abstract.py
@@ -123,6 +123,10 @@ class Extract2D:
             The array from which the regions are extracted and put into the returned list of arrays.
         pixels
             The integer range of pixels between which the extraction is performed.
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
         """
         arr_list = [
             array.native[region.slice]
@@ -166,7 +170,11 @@ class Extract2D:
             are extracted and stacked.
         pixels
             The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
-            FPR rows)
+            FPR rows).
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
         """
 
         arr_list = [
@@ -210,7 +218,11 @@ class Extract2D:
             are extracted and stacked.
         pixels
             The row pixel index to extract the FPR between (e.g. `pixels=(0, 3)` extracts the 1st, 2nd and 3rd
-            FPR rows)
+            FPR rows).
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
         """
 
         mask_2d_list = [
@@ -234,9 +246,10 @@ class Extract2D:
         pixels_from_end: Optional[int] = None,
     ) -> aa.Array1D:
         """
-        Extract a region (e.g. the parallel FPR) of every signal region (e.g. the charge injection region of charge injection data) on the CTI calibration data, stack
-        them by taking their mean and then bin them up to a 1D region (e.g. the 1D parallel FPR) by taking the mean
-        across the direction opposite to clocking (e.g. bin over the serial direction for a parallel FPR).
+        Extract a region (e.g. the parallel FPR) of every signal region (e.g. the charge injection region of
+        charge injection data) on the CTI calibration data, stack them by taking their mean and then bin them up to
+        a 1D region (e.g. the 1D parallel FPR) by taking the mean across the direction opposite to
+        clocking (e.g. bin over the serial direction for a parallel FPR).
 
         This returns the 1D average region (e.g. of the parallel FPR) of all of the charge injection regions. When
         binning a uniform charge injection this binning process removes noise to clearly reveal the FPR or EPER.
@@ -252,7 +265,11 @@ class Extract2D:
             stacked.
         pixels
             The column / row pixel index to extract the region (e.g. FPR, EPER) between (e.g. `pixels=(0, 3)` extracts
-            the 1st, 2nd and 3rd columns / rows)
+            the 1st, 2nd and 3rd columns / rows).
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
         """
 
         arr_list = [
@@ -291,7 +308,11 @@ class Extract2D:
             stacked.
         pixels
             The column / row pixel index to extract the region (e.g. FPR, EPER) between (e.g. `pixels=(0, 3)` extracts
-            the 1st, 2nd and 3rd columns / rows)
[…]
