+++ autocti/extract/one_d/abstract.py
@@ -1,3 +1,4 @@
+import copy
 import numpy as np
 from typing import List, Optional, Tuple
 
@@ -183,3 +184,50 @@ class Extract1D:
             new_array[region.x0 : region.x1] += arr
 
         return new_array
+
+    def add_gaussian_noise_to(
+        self,
+        array: aa.Array1D,
+        noise_sigma: float,
+        noise_seed: int = -1,
+        pixels: Optional[Tuple[int, int]] = None,
+        pixels_from_end: Optional[int] = None,
+    ) -> aa.Array2D:
+        """
+        Adds Gaussian noise of an input sigma value to the regions of the `Extract` object and returns the overall
+        input array with this noise added.
+
+        Parameters
+        ----------
+        array
+            The 1D array which contains the charge injection where regions of Gaussian noise is added.
+        noise_sigma
+            The sigma value (standard deviation) of the Gaussian from which noise values are drann.
+        noise_seed
+            The seed of the random number generator, used for the random noises maps.
+        pixels
+            The row pixel index which determines the region of the FPR (e.g. `pixels=(0, 3)` will compute the region
+            corresponding to the 1st, 2nd and 3rd FPR rows).
+        pixels_from_end
+            Alternative row pixex index specification, which extracts this number of pixels from the end of
+            each region (e.g. FPR / EPER). For example, if each FPR is 100 pixels and `pixels_from_end=10`, the
+            last 10 pixels of each FPR (pixels (90, 100)) are extracted.
+        """
+
+        region_list = self.region_list_from(
+            pixels=pixels, pixels_from_end=pixels_from_end
+        )
+
+        array_1d_list = self.array_1d_list_from(
+            array=array, pixels=pixels, pixels_from_end=pixels_from_end
+        )
+
+        array = copy.copy(array.native)
+
+        for arr, region in zip(array_1d_list, region_list):
+
+            array[region.x0 : region.x1] = aa.preprocess.data_with_gaussian_noise_added(
+                data=arr, sigma=noise_sigma, seed=noise_seed
+            )
+
+        return array
