+++ autocti/charge_injection/imaging/simulator.py
@@ -209,7 +209,14 @@ class SimulatorImagingCI(SimulatorImaging):
         if cosmic_ray_map is not None:
             pre_cti_data += cosmic_ray_map.native
 
-        #    if self.charge_noise is not None:
+        if self.charge_noise is not None:
+
+            pre_cti_data = layout.extract.parallel_fpr.add_gaussian_noise_to(
+                array=pre_cti_data,
+                noise_sigma=self.charge_noise,
+                noise_seed=self.noise_seed,
+                pixels_from_end=layout.extract.parallel_fpr.total_rows_min,
+            )
 
         if cti is not None:
             post_cti_data = clocker.add_cti(data=pre_cti_data, cti=cti)
