+++ test_autocti/charge_injection/imaging/test_simulator_ci.py
@@ -213,13 +213,11 @@ def test__include_charge_noise__is_added_before_cti(parallel_clocker_2d, traps_x
 
     image_no_noise = simulator.pre_cti_data_uniform_from(layout=layout)
 
-    # assert imaging.image - image_no_noise.native == pytest.approx(
-    #     np.array(
-    #         [[1.055, -1.180, -1.097], [-1.073, 0.865, -2.301], [1.744, -0.761, 0.319]]
-    #     ),
-    #     1e-1,
-    # )
-    # assert imaging.layout == layout
+    assert imaging.image - image_no_noise.native == pytest.approx(
+        np.array([[1.01064, -1.1632, -1.08214], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]]),
+        1e-1,
+    )
+    assert imaging.layout == layout
 
 
 def test__include_read_noise__is_added_after_cti(parallel_clocker_2d, traps_x2, ccd):
