### 2023-02-27 008ce96
add allow negative pixels to wrapper
