### 2023-03-14 a7132df
renaming test variables for homogeneity
