+++ test_autocti/extract/two_d/parallel/test_parallel_overscan.py
@@ -9,27 +9,27 @@ def test_region_list_from__via_array_2d_list_from(
 
     extract = ac.Extract2DParallelOverscan(parallel_overscan=(1, 4, 0, 3))
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels=(0, 1))
-    assert (overscan_list[0] == np.array([[1.0, 1.0, 1.0]])).all()
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels=(0, 1))
+    assert (array_2d_list[0] == np.array([[1.0, 1.0, 1.0]])).all()
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels=(2, 3))
-    assert (overscan_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels=(2, 3))
+    assert (array_2d_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels=(0, 3))
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels=(0, 3))
     assert (
-        overscan_list[0]
+        array_2d_list[0]
         == np.array([[1.0, 1.0, 1.0], [2.0, 2.0, 2.0], [3.0, 3.0, 3.0]])
     ).all()
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels=(-1, 1))
-    assert (overscan_list[0] == np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]])).all()
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels=(-1, 1))
+    assert (array_2d_list[0] == np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]])).all()
 
-    overscan_list = extract.array_2d_list_from(
+    array_2d_list = extract.array_2d_list_from(
         array=parallel_masked_array, pixels=(0, 3)
     )
 
     assert (
-        overscan_list[0].mask
+        array_2d_list[0].mask
         == np.array([[False, False, False], [False, True, False], [False, False, True]])
     ).all()
 
@@ -40,21 +40,21 @@ def test_region_list_from__via_array_2d_list_from__pixels_from_end(
 
     extract = ac.Extract2DParallelOverscan(parallel_overscan=(1, 4, 0, 3))
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=1)
-    assert (overscan_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=1)
+    assert (array_2d_list[0] == np.array([[3.0, 3.0, 3.0]])).all()
 
-    overscan_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=3)
+    array_2d_list = extract.array_2d_list_from(array=parallel_array, pixels_from_end=3)
     assert (
-        overscan_list[0]
+        array_2d_list[0]
         == np.array([[1.0, 1.0, 1.0], [2.0, 2.0, 2.0], [3.0, 3.0, 3.0]])
     ).all()
 
-    overscan_list = extract.array_2d_list_from(
+    array_2d_list = extract.array_2d_list_from(
         array=parallel_masked_array, pixels_from_end=3
     )
 
     assert (
-        overscan_list[0].mask
+        array_2d_list[0].mask
         == np.array([[False, False, False], [False, True, False], [False, False, True]])
     ).all()
 
[…]
