### 2023-03-14 2b3150f
more refactoring of variable names
