+++ autocti/extract/two_d/parallel/eper.py
@@ -30,18 +30,18 @@ class Extract2DParallelEPER(Extract2DParallel):
         [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
         [tttttttttt] = parallel / serial charge injection region trail
 
-               [ppppppppppppppppppppp]
-               [ppppppppppppppppppppp]
-          [...][t1t1t1t1t1t1t1t1t1t1t][sss]
-          [...][c1c1cc1c1cc1cc1ccc1cc][sss]
+              [ppppppppppppppppppppp]
+              [ppppppppppppppppppppp]
+         [...][t1t1t1t1t1t1t1t1t1t1t][sss]
+         [...][c1c1cc1c1cc1cc1ccc1cc][sss]
         | [...][1c1c1cc1c1cc1ccc1cc1c][sss]     |
         | [...][t0t0t0t0t0t0t0t0t0t0t][sss]    | Direction
         Par [...][0t0t0t0t0t0t0t0t0t0t0][sss]    | of
         | [...][0ccc0cccc0cccc0cccc0c][sss]    | clocking
-         |/ [...][cc0ccc0cccc0cccc0cccc][sss]   \/
+        |/ [...][cc0ccc0cccc0cccc0cccc][sss]   \/
 
         []     [=====================]
-               <---------Ser----------
+              <---------Ser----------
 
         The extracted regions correspond to the first parallel EPER all charge injection regions:
 
@@ -57,11 +57,11 @@ class Extract2DParallelEPER(Extract2DParallel):
         Parameters
         ----------
         pixels
-            The row indexes to extract the trails between (e.g. rows(0, 3) extracts the 1st, 2nd and 3rd rows).
+           The row indexes to extract the trails between (e.g. rows(0, 3) extracts the 1st, 2nd and 3rd rows).
         pixels_from_end
-            Alternative row pixel index specification, which extracts this number of pixels from the end of
-            the EPER. For example, if each EPER is 100 pixels and `pixels_from_end=10`, the last 10 pixels of each
-            EPER (pixels (90, 100)) are extracted.
+           Alternative row pixel index specification, which extracts this number of pixels from the end of
+           the EPER. For example, if each EPER is 100 pixels and `pixels_from_end=10`, the last 10 pixels of each
+           EPER (pixels (90, 100)) are extracted.
         """
 
         region_list = []
@@ -83,6 +83,65 @@ class Extract2DParallelEPER(Extract2DParallel):
 
         return region_list
 
+    def array_2d_from(self, array: aa.Array2D) -> aa.Array2D:
+        """
+        Extract from an `Array2D` the parallel EPERs and return them as a new `Array2D` where everything else (e.g.
+        the charge injection regions, serial EPERs) are zeros.
+
+        The dimensions of the input array therefore do not change (unlike other `Layout2DCI` methods).
+
+        Negative pixel values can be input into the `pixels` tuple, whereby rows in front of the parallel EPERs (e.g.
+        the FPR) are also extracted.
+
+        The diagram below illustrates the extraction:
+
+        [] = read-out electronics
+        [==========] = read-out register
+        [..........] = serial prescan
+        [pppppppppp] = parallel overscan
+        [ssssssssss] = serial overscan
+        [f#ff#f#f#f] = signal region (FPR) (0 / 1 indicate the region index)
+        [tttttttttt] = parallel / serial charge injection region trail
+
+             [tptpptptptpptpptpptpt]
+             [tptptptpptpttptptptpt]
+        [...][ttttttttttttttttttttt][sss]
+        [...][ccccccccccccccccccccc][sss]
+        | [...][ccccccccccccccccccccc][sss]    |
+        | [...][ttttttttttttttttttttt][sss]    | Direction
+        Par[...][ttttttttttttttttttttt][sss]    | of
[…]

[…]
