+++ autocti/extract/one_d/eper.py
@@ -18,7 +18,7 @@ class Extract1DEPER(Extract1D):
             These are used for extracting the EPER regions of 1D data.
 
             Negative pixel values can be input into the `pixels` tuple, whereby pixels in front of the EPERs (e.g.
-            the FPR) are also extracted.
+            the FPR) are extracted.
 
             Parameters
         ----------
