### 2023-03-15 d4e6315
added pixels_from_end=-1 ovveride to Extract2DParallelEPER which extracts full EPER trails between every charge injection region
