+++ autocti/extract/two_d/serial/overscan_no_eper.py
@@ -13,6 +13,7 @@ class Extract2DSerialOverscanNoEPER(Extract2DSerial):
         self,
         pixels: Optional[Tuple[int, int]] = None,
         pixels_from_end: Optional[int] = None,
+        force_same_row_size : bool = False,
     ) -> List[aa.Region2D]:
         """
         Returns a list of the 2D serial overscan regions without EPER trails, between two input `pixels` indexes.
@@ -62,12 +63,12 @@ class Extract2DSerialOverscanNoEPER(Extract2DSerial):
             The column pixel index which determines the region of the overscan without serial EPERs
             (e.g. `pixels=(0, 3)` will compute the region corresponding to the 1st, 2nd and 3rd overscan columns).
         """
-        #
-        # if pixels_from_end is not None:
-        #    pixels = (
-        #        self.serial_overscan.total_columns - pixels_from_end,
-        #        self.serial_overscan.total_columns,
-        #    )
+
+        if pixels_from_end is not None:
+            pixels = (
+                self.serial_overscan.total_columns - pixels_from_end,
+                self.serial_overscan.total_columns,
+            )
 
         extract_parallel_eper = Extract2DParallelEPER(
             shape_2d=self.shape_2d,
@@ -77,23 +78,23 @@ class Extract2DSerialOverscanNoEPER(Extract2DSerial):
             serial_prescan=self.serial_prescan,
         )
 
-        #        parallel_eper_region_list = extract_parallel_eper.region_list_from(pixels_from_end=
+        parallel_eper_region_list = extract_parallel_eper.region_list_from(
+            pixels_from_end=-1
+        )
 
         region_list = []
 
         for i, region in enumerate(self.region_list):
 
-            parallel_trailing_region = region.parallel_trailing_region_from(
-                pixels=pixels
-            )
-
-            serial_overscan_no_eper_extract = aa.Region2D(
-                (
-                    parallel_trailing_region.y0,
-                    parallel_trailing_region.y1,
-                    self.serial_overscan.x0 + pixels[0],
-                    self.serial_overscan.x0 + pixels[1],
+            region_list.append(
+                aa.Region2D(
+                    (
+                        parallel_eper_region_list[i].y0,
+                        parallel_eper_region_list[i].y1,
+                        self.serial_overscan.x0 + pixels[0],
+                        self.serial_overscan.x0 + pixels[1],
+                    )
                 )
             )
 
-        return [serial_overscan_extract]
+        return region_list
