### 2023-03-15 1896391
added forfce_rows_same_size input to array_2d_list_from in abstract Extract
