### 2023-03-15 4a5213b
Extraction inputs refactored into Settings object
