+++ test_autocti/extract/two_d/serial/test_serial_eper.py
@@ -7,27 +7,39 @@ def test__region_list_from__array_2d_list_from(serial_array, serial_masked_array
 
     extract = ac.Extract2DSerialEPER(region_list=[(0, 3, 1, 4)])
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[4.0], [4.0], [4.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(1, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(1, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[5.0], [5.0], [5.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(2, 3))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(2, 3))
+    )
 
     assert (array_2d_list[0] == np.array([[6.0], [6.0], [6.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[4.0, 5.0], [4.0, 5.0], [4.0, 5.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(-1, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(-1, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[3.0, 4.0], [3.0, 4.0], [3.0, 4.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(1, 4))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(1, 4))
+    )
 
     assert (
         array_2d_list[0]
@@ -36,22 +48,30 @@ def test__region_list_from__array_2d_list_from(serial_array, serial_masked_array
 
     extract = ac.Extract2DSerialEPER(region_list=[(0, 3, 1, 4), (0, 3, 5, 8)])
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[4.0], [4.0], [4.0]])).all()
     assert (array_2d_list[1] == np.array([[8.0], [8.0], [8.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(1, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(1, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[5.0], [5.0], [5.0]])).all()
     assert (array_2d_list[1] == np.array([[9.0], [9.0], [9.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(2, 3))
+    array_2d_list = extract.array_2d_list_from(
[…]
