+++ test_autocti/extract/two_d/serial/test_serial_prescan.py
@@ -6,36 +6,50 @@ import autocti as ac
 def test__region_list_from(serial_array, serial_masked_array):
     extract = ac.Extract2DSerialPrescan(serial_prescan=(0, 3, 1, 4))
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[1.0], [1.0], [1.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(1, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(1, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[2.0], [2.0], [2.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(2, 3))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(2, 3))
+    )
 
     assert (array_2d_list[0] == np.array([[3.0], [3.0], [3.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(-1, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(-1, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]])).all()
 
     extract = ac.Extract2DSerialPrescan(serial_prescan=(0, 3, 1, 5))
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[1.0, 2.0], [1.0, 2.0], [1.0, 2.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(1, 4))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(1, 4))
+    )
 
     assert (
         array_2d_list[0]
         == np.array([[2.0, 3.0, 4.0], [2.0, 3.0, 4.0], [2.0, 3.0, 4.0]])
     ).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_masked_array, pixels=(0, 3))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_masked_array, settings=ac.SettingsExtract(pixels=(0, 3))
+    )
 
     assert (
         (array_2d_list[0].mask)
@@ -48,20 +62,26 @@ def test__region_list_from(serial_array, serial_masked_array):
 def test__region_list_from__pixels_from_end(serial_array, serial_masked_array):
     extract = ac.Extract2DSerialPrescan(serial_prescan=(0, 3, 1, 4))
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels_from_end=1)
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels_from_end=1)
+    )
 
     assert (array_2d_list[0] == np.array([[3.0], [3.0], [3.0]])).all()
 
[…]
