+++ test_autocti/extract/two_d/serial/test_serial_overscan_no_eper.py
@@ -11,11 +11,15 @@ def test__region_list_from(serial_array, serial_masked_array):
         serial_overscan=(0, 3, 8, 10),
     )
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 1))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 1))
+    )
 
     assert (array_2d_list[0] == np.array([[8.0], [8.0]])).all()
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[8.0, 9.0], [8.0, 9.0]])).all()
 
@@ -37,7 +41,9 @@ def test__region_list_from(serial_array, serial_masked_array):
         serial_overscan=(0, 5, 7, 10),
     )
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels=(0, 2))
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels=(0, 2))
+    )
 
     assert (array_2d_list[0] == np.array([[2.0, 12.0]])).all()
     assert (array_2d_list[1] == np.array([[5.0, 15.0], [6.0, 16.0]])).all()
@@ -51,7 +57,9 @@ def test__region_list_from__pixels_from_end(serial_array, serial_masked_array):
         serial_overscan=(0, 3, 8, 10),
     )
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels_from_end=2)
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels_from_end=2)
+    )
 
     assert (array_2d_list[0] == np.array([[8.0, 9.0], [8.0, 9.0]])).all()
 
@@ -73,7 +81,9 @@ def test__region_list_from__pixels_from_end(serial_array, serial_masked_array):
         serial_overscan=(0, 5, 7, 10),
     )
 
-    array_2d_list = extract.array_2d_list_from(array=serial_array, pixels_from_end=2)
+    array_2d_list = extract.array_2d_list_from(
+        array=serial_array, settings=ac.SettingsExtract(pixels_from_end=2)
+    )
 
     assert (array_2d_list[0] == np.array([[12.0, 22.0]])).all()
     assert (array_2d_list[1] == np.array([[15.0, 25.0], [16.0, 26.0]])).all()

[…]
