### 2023-04-04 a0df982
full dataset visualization now rlies on deepcop + analysis input
