+++ autocti/charge_injection/model/result.py
@@ -7,7 +7,7 @@ class ResultImagingCI(ResultDataset):
     def max_log_likelihood_full_fit(self) -> FitImagingCI:
         return self.analysis.fit_via_instance_and_dataset_from(
             instance=self.instance,
-            imaging_ci=self.analysis.dataset.imaging_full,
+            imaging_ci=self.analysis.dataset_full,
             hyper_noise_scale=True,
         )
 
@@ -15,7 +15,7 @@ class ResultImagingCI(ResultDataset):
     def max_log_likelihood_full_fit_no_hyper_scaling(self):
         return self.analysis.fit_via_instance_and_dataset_from(
             instance=self.instance,
-            imaging_ci=self.analysis.dataset.imaging_full,
+            imaging_ci=self.analysis.dataset_full,
             hyper_noise_scale=False,
         )
 
