+++ autocti/__init__.py
@@ -54,7 +54,9 @@ from .dataset_1d.dataset_1d.dataset_1d import Dataset1D
 from .dataset_1d.dataset_1d.simulator import SimulatorDataset1D
 from .dataset_1d.fit import FitDataset1D
 from .dataset_1d.model.analysis import AnalysisDataset1D
+from .dataset_1d.model.result import ResultDataset1D
 from .charge_injection.model.analysis import AnalysisImagingCI
+from .charge_injection.model.result import ResultImagingCI
 from .model.model_util import CTI1D
 from .model.model_util import CTI2D
 from .model.settings import SettingsCTI1D
