### 2023-04-04 7a5059e
refactor Dataset1d to support full dataset
