+++ autocti/dataset_1d/model/analysis.py
@@ -1,11 +1,7 @@
 import os
-from typing import List
+from typing import List, Optional
 
-from autofit.non_linear.samples import SamplesPDF
-from autofit.mapper.prior_model.collection import Collection
-from autofit.mapper.model import ModelInstance
-from autofit.non_linear.abstract_search import Analysis
-from autofit.non_linear.paths.directory import DirectoryPaths
+import autofit as af
 
 from autocti.dataset_1d.dataset_1d.dataset_1d import Dataset1D
 from autocti.dataset_1d.fit import FitDataset1D
@@ -16,13 +12,13 @@ from autocti.model.settings import SettingsCTI1D
 from autocti.clocker.one_d import Clocker1D
 
 
-class AnalysisDataset1D(Analysis):
+class AnalysisDataset1D(af.Analysis):
     def __init__(
         self,
         dataset: Dataset1D,
         clocker: Clocker1D,
         settings_cti: SettingsCTI1D = SettingsCTI1D(),
-        results: List[ResultDataset] = None,
+        dataset_full: Optional[Dataset1D] = None,
     ):
 
         super().__init__()
@@ -30,9 +26,42 @@ class AnalysisDataset1D(Analysis):
         self.dataset = dataset
         self.clocker = clocker
         self.settings_cti = settings_cti
-        self.results = results
+        self.dataset_full = dataset_full
 
-    def log_likelihood_function(self, instance: ModelInstance) -> float:
+    def modify_before_fit(self, paths: af.DirectoryPaths, model: af.Collection):
+        """
+        PyAutoFit calls this function immediately before the non-linear search begins, therefore it can be used to
+        perform tasks using the final model parameterization.
+
+        This function:
+
+         1) Visualizes the 1D dataset, which does not change during the analysis and thus can be done once.
+
+        Parameters
+        ----------
+        paths
+            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
+            visualization and the pickled objects used by the aggregator output by this function.
+        model
+            The PyAutoFit model object, which includes model components representing the galaxies that are fitted to
+            the imaging data.
+        """
+        if not paths.is_complete:
+
+            if not os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
+
+                visualizer = VisualizerDataset1D(visualize_path=paths.image_path)
+                visualizer.visualize_dataset_1d(dataset_1d=self.dataset)
+
+                if self.dataset_full is not None:
+
+                    visualizer.visualize_dataset_1d(
+                        dataset_1d=self.dataset_full, folder_suffix="_full"
+                    )
+
+        return self
+
+    def log_likelihood_function(self, instance: af.ModelInstance) -> float:
         """
         Determine the fitness of a particular model
 
@@ -53,7 +82,7 @@ class AnalysisDataset1D(Analysis):
         return fit.log_likelihood
 
     def fit_via_instance_and_dataset_from(
-        self, instance: ModelInstance, dataset: Dataset1D
+        self, instance: af.ModelInstance, dataset: Dataset1D
     ) -> FitDataset1D:
 
         post_cti_data = self.clocker.add_cti(
@@ -62,31 +91,37 @@ class AnalysisDataset1D(Analysis):
 
         return FitDataset1D(dataset=dataset, post_cti_data=post_cti_data)
 
-    def fit_via_instance_from(self, instance: ModelInstance) -> FitDataset1D:
+    def fit_via_instance_from(self, instance: af.ModelInstance) -> FitDataset1D:
 
         return self.fit_via_instance_and_dataset_from(
             instance=instance, dataset=self.dataset
         )
 
     def visualize(
[…]
