### 2023-04-06 df60419
visualizer now plots multiple analysis results on subplot, with unit test
