### 2023-04-06 0e70a13
test__visualize_fit_ci_regions_combined
