### 2023-04-06 be2e8bc
remoe over testing in ImagingCIVisualizer
