+++ test_autocti/charge_injection/model/test_visualizer_ci.py
@@ -30,9 +30,6 @@ def test__visualizes_imaging_ci_using_configs(imaging_ci_7x7, plot_path, plot_pa
     assert path.join(plot_path, "subplot_imaging_ci.png") in plot_patch.paths
     assert path.join(plot_path, "image_2d.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
-    assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
-    assert path.join(plot_path, "pre_cti_data.png") in plot_patch.paths
-    assert path.join(plot_path, "cosmic_ray_map.png") in plot_patch.paths
 
 
 def test__visualizes_imaging_ci_regions_using_configs(
@@ -53,11 +50,6 @@ def test__visualizes_imaging_ci_regions_using_configs(
     assert path.join(plot_path, "subplot_1d_ci_parallel_fpr.png") in plot_patch.paths
     assert path.join(plot_path, "image_parallel_fpr.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map_parallel_fpr.png") not in plot_patch.paths
-    assert (
-        path.join(plot_path, "signal_to_noise_map_parallel_fpr.png")
-        not in plot_patch.paths
-    )
-    assert path.join(plot_path, "pre_cti_data_parallel_fpr.png") in plot_patch.paths
 
 
 def test___visualizes_fit_ci_using_configs(fit_ci_7x7, plot_path, plot_patch):
@@ -74,12 +66,6 @@ def test___visualizes_fit_ci_using_configs(fit_ci_7x7, plot_path, plot_patch):
     assert path.join(plot_path, "subplot_fit_ci.png") in plot_patch.paths
     assert path.join(plot_path, "image_2d.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
-    assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
-    assert path.join(plot_path, "pre_cti_data.png") in plot_patch.paths
-    assert path.join(plot_path, "post_cti_data.png") in plot_patch.paths
-    assert path.join(plot_path, "residual_map.png") not in plot_patch.paths
-    assert path.join(plot_path, "normalized_residual_map.png") in plot_patch.paths
-    assert path.join(plot_path, "chi_squared_map.png") in plot_patch.paths
 
     plot_patch.paths = []
 
@@ -88,12 +74,6 @@ def test___visualizes_fit_ci_using_configs(fit_ci_7x7, plot_path, plot_patch):
     assert path.join(plot_path, "subplot_fit_ci.png") in plot_patch.paths
     assert path.join(plot_path, "image_2d.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map.png") in plot_patch.paths
-    assert path.join(plot_path, "signal_to_noise_map.png") in plot_patch.paths
-    assert path.join(plot_path, "pre_cti_data.png") in plot_patch.paths
-    assert path.join(plot_path, "post_cti_data.png") in plot_patch.paths
-    assert path.join(plot_path, "residual_map.png") in plot_patch.paths
-    assert path.join(plot_path, "normalized_residual_map.png") in plot_patch.paths
-    assert path.join(plot_path, "chi_squared_map.png") in plot_patch.paths
 
 
 def test___visualizes_fit_ci_regions_using_configs(fit_ci_7x7, plot_path, plot_patch):
@@ -114,18 +94,6 @@ def test___visualizes_fit_ci_regions_using_configs(fit_ci_7x7, plot_path, plot_p
     )
[…]
