### 2023-04-06 bfb93ce
clean up Dataset1DVisualizer
