+++ test_autocti/dataset_1d/model/test_visualizer_1d.py
@@ -14,59 +14,41 @@ def make_visualizer_plotter_setup():
     return path.join("{}".format(directory), "files")
 
 
-class TestVisualizerDataset1D:
-    def test__visualizes_dataset_1d_using_configs(
-        self, dataset_1d_7, plot_path, plot_patch
-    ):
+def test__visualizes_dataset_1d_using_configs(dataset_1d_7, plot_path, plot_patch):
 
-        if path.exists(plot_path):
-            shutil.rmtree(plot_path)
+    if path.exists(plot_path):
+        shutil.rmtree(plot_path)
 
-        visualizer = VisualizerDataset1D(visualize_path=plot_path)
+    visualizer = VisualizerDataset1D(visualize_path=plot_path)
 
-        visualizer.visualize_dataset_1d(dataset_1d=dataset_1d_7)
+    visualizer.visualize_dataset_1d(dataset_1d=dataset_1d_7)
 
-        plot_path = path.join(plot_path, "dataset_1d")
+    plot_path = path.join(plot_path, "dataset_1d")
 
-        assert path.join(plot_path, "subplot_dataset_1d.png") in plot_patch.paths
-        assert path.join(plot_path, "data.png") in plot_patch.paths
-        assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
-        assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
-        assert path.join(plot_path, "pre_cti_data.png") in plot_patch.paths
+    assert path.join(plot_path, "subplot_dataset_1d.png") in plot_patch.paths
+    assert path.join(plot_path, "data.png") in plot_patch.paths
+    assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
 
-    def test___visualizes_fit_line_using_configs(
-        self, fit_line_7, plot_path, plot_patch
-    ):
 
-        if os.path.exists(plot_path):
-            shutil.rmtree(plot_path)
+def test___visualizes_fit_line_using_configs(fit_line_7, plot_path, plot_patch):
 
-        visualizer = VisualizerDataset1D(visualize_path=plot_path)
+    if os.path.exists(plot_path):
+        shutil.rmtree(plot_path)
 
-        visualizer.visualize_fit_line(fit=fit_line_7, during_analysis=True)
+    visualizer = VisualizerDataset1D(visualize_path=plot_path)
 
-        plot_path = path.join(plot_path, "fit_dataset_1d")
+    visualizer.visualize_fit_line(fit=fit_line_7, during_analysis=True)
 
-        assert path.join(plot_path, "subplot_fit_dataset_1d.png") in plot_patch.paths
-        assert path.join(plot_path, "data.png") in plot_patch.paths
-        assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
-        assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
-        assert path.join(plot_path, "pre_cti_data.png") in plot_patch.paths
-        assert path.join(plot_path, "post_cti_data.png") in plot_patch.paths
-        assert path.join(plot_path, "residual_map.png") not in plot_patch.paths
-        assert path.join(plot_path, "normalized_residual_map.png") in plot_patch.paths
-        assert path.join(plot_path, "chi_squared_map.png") in plot_patch.paths
+    plot_path = path.join(plot_path, "fit_dataset_1d")
 
-        plot_patch.paths = []
+    assert path.join(plot_path, "subplot_fit_dataset_1d.png") in plot_patch.paths
+    assert path.join(plot_path, "data.png") in plot_patch.paths
+    assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
 
[…]
