### 2023-04-06 a62a955
started writing, but commented out, visualize_fit_1d_combined
