### 2023-04-06 63eb5ee
condense region plotting for Dataset1DPlotter
