+++ autocti/dataset_1d/plot/dataset_1d_plotters.py
@@ -1,5 +1,4 @@
-import numpy as np
-from typing import Callable
+from typing import Callable, Optional
 
 import autoarray.plot as aplt
 
@@ -59,6 +58,9 @@ class Dataset1DPlotter(Plotter):
 
     def figures_1d(
         self,
+        region: Optional[str] = None,
+        data_with_noise_map: bool = False,
+        data_with_noise_map_logy: bool = False,
         data: bool = False,
         noise_map: bool = False,
         signal_to_noise_map: bool = False,
@@ -70,76 +72,8 @@ class Dataset1DPlotter(Plotter):
         The API is such that every plottable attribute of the `Imaging` object is an input parameter of type bool of
         the function, which if switched to `True` means that it is plotted.
 
-        Parameters
-        ----------
-        data
-            Whether to make a 1D plot (via `plot`) of the data.
-        noise_map
-            Whether to make a 1D plot (via `plot`) of the noise map.
-        signal_to_noise_map
-            Whether to make a 1D plot (via `plot`) of the signal-to-noise map.
-        pre_cti_data
-            Whether to make a 1D plot (via `plot`) of the pre-cti data.
-        """
-
-        if data:
-
-            self.mat_plot_1d.plot_yx(
-                y=self.dataset_1d.data,
-                x=self.dataset_1d.data.grid_radial,
-                visuals_1d=self.get_visuals_1d(),
-                auto_labels=AutoLabels(title="Dataset 1D Data", filename="data"),
-            )
-
-        if noise_map:
-
-            self.mat_plot_1d.plot_yx(
-                y=self.dataset_1d.noise_map,
-                x=self.dataset_1d.noise_map.grid_radial,
-                visuals_1d=self.get_visuals_1d(),
-                auto_labels=AutoLabels(
-                    title="Dataset 1D Noise Map", filename="noise_map"
-                ),
-            )
-
-        if signal_to_noise_map:
-
-            self.mat_plot_1d.plot_yx(
-                y=self.dataset_1d.signal_to_noise_map,
-                x=self.dataset_1d.signal_to_noise_map.grid_radial,
-                visuals_1d=self.get_visuals_1d(),
-                auto_labels=AutoLabels(
-                    title="Dataset 1D Signal-To-Noise Map",
-                    filename="signal_to_noise_map",
-                ),
-            )
-
-        if pre_cti_data:
-
-            self.mat_plot_1d.plot_yx(
-                y=self.dataset_1d.pre_cti_data,
-                x=self.dataset_1d.pre_cti_data.grid_radial,
-                visuals_1d=self.get_visuals_1d(),
-                auto_labels=AutoLabels(
-                    title="Dataset 1D Pre CTI Data", filename="pre_cti_data"
-                ),
-            )
-
-    def figures_1d_of_region(
-        self,
-        region: str,
-        data: bool = False,
-        noise_map: bool = False,
-        pre_cti_data: bool = False,
-        signal_to_noise_map: bool = False,
-        data_with_noise_map: bool = False,
-        data_with_noise_map_logy: bool = False,
-    ):
-        """
-        Plots the individual attributes of the plotter's `Dataset1D` object in 1D.
-
-        These 1D plots correspond to regions in 1D on the charge injection image, which are binned up to produce a
-        1D plot.
+        If a `region` string is input, the 1D plots correspond to regions in 1D on the 1D dataset, which are binned up
+        to produce a1D plot.
 
         For example, for the input `region=fpr`, this function extracts the FPR over each charge region and bins them
         such that the 1D plot shows the average FPR.
@@ -151,6 +85,12 @@ class Dataset1DPlotter(Plotter):
         ----------
         region
             The region on the 1D dataset where data is extracted and binned {fpr", "eper"}
+        data_with_noise_map
[…]
