+++ autocti/dataset_1d/plot/fit_plotters.py
@@ -1,3 +1,5 @@
+from typing import Callable, Optional
+
 import autoarray.plot as aplt
 
 from autoarray.plot.auto_labels import AutoLabels
@@ -46,16 +48,23 @@ class FitDataset1DPlotter(Plotter):
     def get_visuals_1d(self) -> aplt.Visuals1D:
         return self.visuals_1d
 
+    @property
+    def extract_region_from(self) -> Callable:
+        return self.fit.dataset.layout.extract_region_from
+
     def figures_1d(
         self,
-        data=False,
-        noise_map=False,
-        signal_to_noise_map=False,
-        pre_cti_data=False,
-        post_cti_data=False,
-        residual_map=False,
-        normalized_residual_map=False,
-        chi_squared_map=False,
+        region: Optional[str] = None,
+        data_with_noise_map: bool = False,
+        data_with_noise_map_logy: bool = False,
+        data: bool = False,
+        noise_map: bool = False,
+        signal_to_noise_map: bool = False,
+        pre_cti_data: bool = False,
+        post_cti_data: bool = False,
+        residual_map: bool = False,
+        normalized_residual_map: bool = False,
+        chi_squared_map: bool = False,
     ):
         """
         Plots the individual attributes of the plotter's `FitDataset1D` object in 1D.
@@ -65,7 +74,15 @@ class FitDataset1DPlotter(Plotter):
 
         Parameters
         ----------
-        image
+        region
+            The region on the 1D dataset where data is extracted and binned {fpr", "eper"}
+        data_with_noise_map
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars.
+        data_with_noise_map_logy
+            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
+            noise-map values included as error bars and the y-axis on a log10 scale.
+        data
             Whether to make a 1D plot (via `plot`) of the image data.
         noise_map
             Whether to make a 1D plot (via `plot`) of the noise map.
@@ -82,99 +99,191 @@ class FitDataset1DPlotter(Plotter):
         chi_squared_map
             Whether to make a 1D plot (via `plot`) of the chi-squared map.
         """
-        if data:
+
+        suffix = f"_{region}" if region is not None else ""
+
+        if data_with_noise_map:
+
+            y = self.extract_region_from(array=self.fit.data, region=region)
+            y_errors = self.extract_region_from(
+                array=self.fit.noise_map, region=region
+            )
 
             self.mat_plot_1d.plot_yx(
-                y=self.fit.data,
-                x=self.fit.data.grid_radial,
+                y=y,
+                x=range(len(y)),
+                plot_axis_type_override="errorbar",
+                y_errors=y_errors,
                 visuals_1d=self.get_visuals_1d(),
-                auto_labels=AutoLabels(title="Image", filename="data"),
+                auto_labels=AutoLabels(
+                    title=f"Data 1D With Noise {region}",
+                    ylabel="Data (e-)",
+                    xlabel="Pixel No.",
+                    filename=f"data_with_noise_map{suffix}",
+                ),
+            )
+
+        if data_with_noise_map_logy:
+
+            y = self.extract_region_from(array=self.fit.data, region=region)
+            y_errors = self.extract_region_from(
+                array=self.fit.noise_map, region=region
+            )
+
+            self.mat_plot_1d.plot_yx(
+                y=y,
+                x=range(len(y)),
+                plot_axis_type_override="errorbar_logy",
+                y_errors=y_errors,
+                visuals_1d=self.get_visuals_1d(),
+                auto_labels=AutoLabels(
[…]
