+++ test_autocti/charge_injection/plot/test_imaging_ci_plotters.py
@@ -17,7 +17,9 @@ def make_imaging_ci_plotter_setup():
     )
 
 
-def test__individual_attributes_are_output(imaging_ci_7x7, plot_path, plot_patch):
+def test__figures_2d__individual_attributes_are_output(
+    imaging_ci_7x7, plot_path, plot_patch
+):
 
     imaging_ci_plotter = aplt.ImagingCIPlotter(
         dataset=imaging_ci_7x7,
@@ -51,14 +53,16 @@ def test__individual_attributes_are_output(imaging_ci_7x7, plot_path, plot_patch
     assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
 
 
-def test__individual_1d_of_region_are_output(imaging_ci_7x7, plot_path, plot_patch):
+def test__figures_1d__individual_1d_of_region_are_output(
+    imaging_ci_7x7, plot_path, plot_patch
+):
 
     imaging_ci_plotter = aplt.ImagingCIPlotter(
         dataset=imaging_ci_7x7,
         mat_plot_1d=aplt.MatPlot1D(output=aplt.Output(plot_path, format="png")),
     )
 
-    imaging_ci_plotter.figures_1d_of_region(
+    imaging_ci_plotter.figures_1d(
         region="parallel_fpr",
         image=True,
         noise_map=True,
@@ -75,9 +79,7 @@ def test__individual_1d_of_region_are_output(imaging_ci_7x7, plot_path, plot_pat
 
     plot_patch.paths = []
 
-    imaging_ci_plotter.figures_1d_of_region(
-        region="parallel_fpr", image=True, pre_cti_data=True
-    )
+    imaging_ci_plotter.figures_1d(region="parallel_fpr", image=True, pre_cti_data=True)
 
     assert path.join(plot_path, "image_parallel_fpr.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map_parallel_fpr.png") not in plot_patch.paths
@@ -99,5 +101,5 @@ def test__subplots__output(imaging_ci_7x7, plot_path, plot_patch):
     imaging_ci_plotter.subplot_imaging_ci()
     assert path.join(plot_path, "subplot_imaging_ci.png") in plot_patch.paths
 
-    imaging_ci_plotter.subplot_1d_of_region(region="parallel_fpr")
+    imaging_ci_plotter.subplot_1d(region="parallel_fpr")
     assert path.join(plot_path, "subplot_1d_ci_parallel_fpr.png") in plot_patch.paths
