### 2023-04-07 31a20af
added test__visualize_fit_ci_combined
