+++ test_autocti/dataset_1d/model/test_visualizer_1d.py
@@ -14,7 +14,7 @@ def make_visualizer_plotter_setup():
     return path.join("{}".format(directory), "files")
 
 
-def test__visualizes_dataset_1d_using_configs(dataset_1d_7, plot_path, plot_patch):
+def test__visualize_dataset_1d__uses_configs(dataset_1d_7, plot_path, plot_patch):
 
     if path.exists(plot_path):
         shutil.rmtree(plot_path)
@@ -30,7 +30,7 @@ def test__visualizes_dataset_1d_using_configs(dataset_1d_7, plot_path, plot_patc
     assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
 
 
-def test___visualizes_fit_1d_using_configs(fit_1d_7, plot_path, plot_patch):
+def test__visualize_fit_1d__uses_configs(fit_1d_7, plot_path, plot_patch):
 
     if os.path.exists(plot_path):
         shutil.rmtree(plot_path)
@@ -52,3 +52,22 @@ def test___visualizes_fit_1d_using_configs(fit_1d_7, plot_path, plot_patch):
     assert path.join(plot_path, "subplot_fit_dataset_1d.png") in plot_patch.paths
     assert path.join(plot_path, "data.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map.png") in plot_patch.paths
+
+
+def test__visualize_fit_ci_combined(fit_1d_7, plot_path, plot_patch):
+
+    if os.path.exists(plot_path):
+        shutil.rmtree(plot_path)
+
+    visualizer = VisualizerDataset1D(visualize_path=plot_path)
+
+    visualizer.visualize_fit_1d_combined(
+        fit_list=[fit_1d_7, fit_1d_7], during_analysis=True
+    )
+
+    plot_path = path.join(plot_path, "fit_dataset_1d_combined")
+
+    assert path.join(plot_path, "subplot_residual_map_list.png") in plot_patch.paths
+    assert (
+        path.join(plot_path, "subplot_chi_squared_map_list.png") not in plot_patch.paths
+    )
