+++ test_autocti/charge_injection/model/test_visualizer_ci.py
@@ -14,7 +14,7 @@ def make_visualizer_plotter_setup():
     return path.join("{}".format(directory), "files")
 
 
-def test__visualizes_imaging_ci_using_configs(imaging_ci_7x7, plot_path, plot_patch):
+def test__visualizes_imaging_ci__uses_configs(imaging_ci_7x7, plot_path, plot_patch):
 
     if path.exists(plot_path):
         shutil.rmtree(plot_path)
@@ -32,7 +32,7 @@ def test__visualizes_imaging_ci_using_configs(imaging_ci_7x7, plot_path, plot_pa
     assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
 
 
-def test__visualizes_imaging_ci_regions_using_configs(
+def test__visualizes_imaging_ci_regions__uses_configs(
     imaging_ci_7x7, plot_path, plot_patch
 ):
 
@@ -52,7 +52,7 @@ def test__visualizes_imaging_ci_regions_using_configs(
     assert path.join(plot_path, "noise_map_parallel_fpr.png") not in plot_patch.paths
 
 
-def test___visualizes_fit_ci_using_configs(fit_ci_7x7, plot_path, plot_patch):
+def test___visualizes_fit_ci__uses_configs(fit_ci_7x7, plot_path, plot_patch):
 
     if os.path.exists(plot_path):
         shutil.rmtree(plot_path)
@@ -76,7 +76,7 @@ def test___visualizes_fit_ci_using_configs(fit_ci_7x7, plot_path, plot_patch):
     assert path.join(plot_path, "noise_map.png") in plot_patch.paths
 
 
-def test___visualizes_fit_ci_regions_using_configs(fit_ci_7x7, plot_path, plot_patch):
+def test___visualizes_fit_ci_regions__uses_configs(fit_ci_7x7, plot_path, plot_patch):
 
     if os.path.exists(plot_path):
         shutil.rmtree(plot_path)
