### 2023-04-07 b7706c2
visualize_fit_1d_region_combined
