### 2023-04-17 07e3613
add method region_list_from which inspects the model and returns the regions of the data that should be visualized
