### 2023-04-17 73501f3
remove with_noise_map tag from visualization
