+++ autocti/charge_injection/plot/imaging_ci_plotters.py
@@ -154,9 +154,8 @@ class ImagingCIPlotter(Plotter):
     def figures_1d(
         self,
         region: str,
-        data_with_noise_map: bool = False,
-        data_with_noise_map_logy: bool = False,
         data: bool = False,
+        data_logy: bool = False,
         noise_map: bool = False,
         pre_cti_data: bool = False,
         signal_to_noise_map: bool = False,
@@ -180,14 +179,12 @@ class ImagingCIPlotter(Plotter):
         region
             The region on the charge injection image where data is extracted and binned over the parallel or serial
             direction {"parallel_fpr", "parallel_eper", "serial_fpr", "serial_eper"}
-        data_with_noise_map
+        data
             Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
             noise-map values included as error bars.
-        data_with_noise_map_logy
+        data_logy
             Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region, with the
             noise-map values included as error bars and the y-axis on a log10 scale.
-        data
-            Whether to make a 1D plot (via `plot`) of the image data extracted and binned over the region.
         noise_map
             Whether to make a 1D plot (via `plot`) of the noise-map extracted and binned over the region.
         pre_cti_data
@@ -199,17 +196,43 @@ class ImagingCIPlotter(Plotter):
 
         if data:
 
-            y = self.extract_region_from(array=self.imaging.image, region=region)
+            y = self.extract_region_from(array=self.dataset.data, region=region)
+            y_errors = self.extract_region_noise_map_from(
+                array=self.dataset.noise_map, region=region
+            )
 
             self.mat_plot_1d.plot_yx(
                 y=y,
                 x=range(len(y)),
+                plot_axis_type_override="errorbar",
+                y_errors=y_errors,
                 visuals_1d=self.get_visuals_1d(),
                 auto_labels=AutoLabels(
-                    title=f"Data {region}",
+                    title=f"Data w/ Noise {region} (FPR = {self.dataset.fpr_value} e-)",
                     ylabel="Data (e-)",
                     xlabel="Pixel No.",
-                    filename=f"image_{region}",
+                    filename=f"data_{region}",
+                ),
+            )
+
+        if data_logy:
+
+            y = self.extract_region_from(array=self.dataset.data, region=region)
+            y_errors = self.extract_region_noise_map_from(
+                array=self.dataset.noise_map, region=region
+            )
+
+            self.mat_plot_1d.plot_yx(
+                y=y,
+                x=range(len(y)),
+                plot_axis_type_override="errorbar_logy",
+                y_errors=y_errors,
+                visuals_1d=self.get_visuals_1d(),
+                auto_labels=AutoLabels(
+                    title=f"Data w/ Noise {region} [log10 y] (FPR = {self.dataset.fpr_value} e-)",
+                    ylabel="Data (e-)",
+                    xlabel="Pixel No.",
+                    filename=f"data_logy_{region}",
                 ),
             )
 
@@ -260,48 +283,6 @@ class ImagingCIPlotter(Plotter):
                     ylabel="Signal To Noise (e-)",
                     xlabel="Pixel No.",
                     filename=f"signal_to_noise_map_{region}",
-                ),
-            )
-
-        if data_with_noise_map:
-
-            y = self.extract_region_from(array=self.dataset.data, region=region)
-            y_errors = self.extract_region_noise_map_from(
-                array=self.dataset.noise_map, region=region
-            )
-
-            self.mat_plot_1d.plot_yx(
-                y=y,
-                x=range(len(y)),
[…]
