+++ test_autocti/config/visualize.yaml
@@ -1,7 +1,7 @@
 plots:
   dataset:
-    data_with_noise_map: true
-    data_with_noise_map_logy: false
+    data: true
+    data_logy: false
     absolute_signal_to_noise_map: false
     cosmic_ray_map: true
     data: true
@@ -16,8 +16,8 @@ plots:
     all_at_end_fits: false
     all_at_end_png: true
     chi_squared_map: false
-    data_with_noise_map: true
-    data_with_noise_map_logy: false
+    data: true
+    data_logy: false
     data: true
     noise_map: false
     noise_scaling_map_dict: false

[…]
