### 2023-04-17 a71026e
added errorbars to residual map plot
