### 2023-04-20 a405689
support new visualizer methods
