+++ autocti/charge_injection/model/visualizer.py
@@ -15,21 +15,21 @@ logger.setLevel(level="INFO")
 
 
 class VisualizerImagingCI(Visualizer):
-    def visualize_imaging_ci(self, imaging_ci, folder_suffix: str = ""):
+    def visualize_dataset(self, dataset, folder_suffix: str = ""):
         def should_plot(name):
             return plot_setting(section="dataset", name=name)
 
-        mat_plot_2d = self.mat_plot_2d_from(subfolders=f"imaging_ci{folder_suffix}")
+        mat_plot_2d = self.mat_plot_2d_from(subfolders=f"dataset{folder_suffix}")
 
-        imaging_ci_plotter = aplt.ImagingCIPlotter(
-            dataset=imaging_ci, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
+        dataset_plotter = aplt.ImagingCIPlotter(
+            dataset=dataset, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
         )
 
         if should_plot("subplot_dataset"):
 
-            imaging_ci_plotter.subplot_imaging_ci()
+            dataset_plotter.subplot_imaging_ci()
 
-        imaging_ci_plotter.figures_2d(
+        dataset_plotter.figures_2d(
             data=should_plot("data"),
             noise_map=should_plot("noise_map"),
             inverse_noise_map=should_plot("inverse_noise_map"),
@@ -40,16 +40,14 @@ class VisualizerImagingCI(Visualizer):
             cosmic_ray_map=should_plot("cosmic_ray_map"),
         )
 
-    def visualize_imaging_ci_regions(
-        self, imaging_ci, region_list, folder_suffix: str = ""
-    ):
+    def visualize_dataset_regions(self, dataset, region_list, folder_suffix: str = ""):
         def should_plot(name):
             return plot_setting(section="dataset", name=name)
 
-        mat_plot_1d = self.mat_plot_1d_from(subfolders=f"imaging_ci{folder_suffix}")
+        mat_plot_1d = self.mat_plot_1d_from(subfolders=f"dataset{folder_suffix}")
 
-        imaging_ci_plotter = aplt.ImagingCIPlotter(
-            dataset=imaging_ci, mat_plot_1d=mat_plot_1d, include_2d=self.include_2d
+        dataset_plotter = aplt.ImagingCIPlotter(
+            dataset=dataset, mat_plot_1d=mat_plot_1d, include_2d=self.include_2d
         )
 
         for region in region_list:
@@ -58,9 +56,9 @@ class VisualizerImagingCI(Visualizer):
 
                 if should_plot("subplot_dataset"):
 
-                    imaging_ci_plotter.subplot_1d(region=region)
+                    dataset_plotter.subplot_1d(region=region)
 
-                imaging_ci_plotter.figures_1d(
+                dataset_plotter.figures_1d(
                     region=region,
                     data=should_plot("data"),
                     data_logy=should_plot("data_logy"),
@@ -75,11 +73,75 @@ class VisualizerImagingCI(Visualizer):
                     f"VISUALIZATION - Could not visualize the ImagingCI 1D {region}"
                 )
 
-    def visualize_fit_ci(self, fit, during_analysis, folder_suffix: str = ""):
+    def visualize_dataset_combined(self, dataset_list, folder_suffix: str = ""):
+        def should_plot(name):
+            return plot_setting(section="dataset", name=name)
+
+        mat_plot_2d = self.mat_plot_2d_from(
+            subfolders=f"dataset_combined{folder_suffix}"
+        )
+
+        dataset_plotter_list = [
+            aplt.ImagingCIPlotter(
[…]
