+++ autocti/extract/two_d/parallel/calibration.py
@@ -278,6 +278,7 @@ class Extract2DParallelCalibration:
             cosmic_ray_map=cosmic_ray_map,
             noise_scaling_map_dict=noise_scaling_map_dict,
             fpr_value=imaging_ci.fpr_value,
+            settings_dict=imaging_ci.settings_dict,
         )
 
         return imaging_ci.apply_mask(mask=mask)

[…]
