### 2023-04-20 3b92b5b
improve figures of 1D fits
