### 2023-04-21 830f5d4
remove absolute sn map / potential chi squared / inversion noise map
