+++ test_autocti/charge_injection/plot/test_fit_ci_plotters.py
@@ -16,7 +16,6 @@ def make_fit_ci_plotter_setup():
 def test__individual_attribute_plots__all_plot_correctly(
     fit_ci_7x7, plot_path, plot_patch
 ):
-
     fit_ci_plotter = aplt.FitImagingCIPlotter(
         fit=fit_ci_7x7,
         mat_plot_2d=aplt.MatPlot2D(output=aplt.Output(plot_path, format="png")),
@@ -66,7 +65,6 @@ def test__individual_attribute_plots__all_plot_correctly(
 def test__individual_line_attriutes_plot__all_plot_correctly_output(
     fit_ci_7x7, plot_path, plot_patch
 ):
-
     fit_ci_plotter = aplt.FitImagingCIPlotter(
         fit=fit_ci_7x7,
         mat_plot_2d=aplt.MatPlot2D(output=aplt.Output(plot_path, format="png")),
@@ -124,15 +122,14 @@ def test__individual_line_attriutes_plot__all_plot_correctly_output(
 
 
 def test__fit_ci_subplots_are_output(fit_ci_7x7, plot_path, plot_patch):
-
     fit_ci_plotter = aplt.FitImagingCIPlotter(
         fit=fit_ci_7x7,
         mat_plot_2d=aplt.MatPlot2D(output=aplt.Output(plot_path, format="png")),
         mat_plot_1d=aplt.MatPlot1D(output=aplt.Output(plot_path, format="png")),
     )
 
-    fit_ci_plotter.subplot_fit_ci()
-    assert path.join(plot_path, "subplot_fit_ci.png") in plot_patch.paths
+    fit_ci_plotter.subplot_fit()
+    assert path.join(plot_path, "subplot_fit.png") in plot_patch.paths
 
     fit_ci_plotter.subplot_1d(region="parallel_fpr")
     assert (
