+++ autocti/dataset_1d/model/result.py
@@ -2,5 +2,4 @@ from autocti.model.result import ResultDataset
 
 
 class ResultDataset1D(ResultDataset):
-
     pass
