### 2023-05-10 1f79569
Dataset1D aggregator works
