+++ autocti/aggregator/__init__.py
@@ -1,0 +1,1 @@
+from autocti.aggregator.dataset_1d import Dataset1DAgg
