+++ autocti/aggregator/dataset_1d.py
@@ -2,9 +2,7 @@ from functools import partial
 from typing import Optional
 
 import autofit as af
-import autoarray as aa
 
-from autocti.dataset_1d.dataset_1d.dataset_1d import Dataset1D
 from autocti.dataset_1d.dataset_1d.settings import SettingsDataset1D
 
 
@@ -23,36 +21,25 @@ def _dataset_1d_from(fit: af.Fit, settings_dataset: Optional[SettingsDataset1D] 
         A PyAutoFit aggregator's SearchOutput object containing the generators of the results of model-fits.
     """
 
-    data = fit.value(name="data")
-    noise_map = fit.value(name="noise_map")
-    pre_cti_data = fit.value(name="pre_cti_data")
-    layout = fit.value(name="layout")
+    dataset = fit.value(name="dataset")
 
     settings_dataset = settings_dataset or fit.value(name="settings_dataset")
 
-    dataset_1d = Dataset1D(
-        data=data,
-        noise_map=noise_map,
-        pre_cti_data=pre_cti_data,
-        settings=settings_dataset,
-        pad_for_convolver=True,
-    )
+    dataset.apply_settings(settings=settings_dataset)
 
-    dataset_1d.apply_settings(settings=settings_dataset)
-
-    return dataset_1d
+    return dataset
 
 
 class Dataset1DAgg:
     def __init__(self, aggregator: af.Aggregator):
         self.aggregator = aggregator
 
-    def dataset_1d_gen_from(self, settings_dataset: Optional[aa.SettingsDataset1D] = None):
+    def dataset_gen_from(self, settings_dataset: Optional[SettingsDataset1D] = None):
         """
         Returns a generator of `Dataset1D` objects from an input aggregator, which generates a list of the
         `Dataset1D` objects for every set of results loaded in the aggregator.
 
-        This is performed by mapping the `dataset_1d_from_agg_obj` with the aggregator, which sets up each
+        This is performed by mapping the `dataset_from_agg_obj` with the aggregator, which sets up each
         dataset_1d using only generators ensuring that manipulating the dataset_1d of large sets of results is done in a
         memory efficient way.
 
