### 2023-05-10 c77517c
FitDataset1DAgg works
