+++ autocti/dataset_1d/plot/dataset_1d_plotters.py
@@ -173,7 +173,6 @@ class Dataset1DPlotter(Plotter):
             )
 
         if signal_to_noise_map:
-
             y = self.extract_region_from(
                 array=self.dataset.signal_to_noise_map, region=region
             )
