+++ autocti/aggregator/__init__.py
@@ -1,1 +1,2 @@
 from autocti.aggregator.dataset_1d import Dataset1DAgg
+from autocti.aggregator.fit_dataset_1d import FitDataset1DAgg
