### 2023-05-10 bf7094b
fiox aggregator tests
