+++ autocti/aggregator/__init__.py
@@ -1,2 +1,3 @@
+from autocti.aggregator.cti import CTIAgg
 from autocti.aggregator.dataset_1d import Dataset1DAgg
 from autocti.aggregator.fit_dataset_1d import FitDataset1DAgg
