+++ test_autocti/aggregator/test_cti.py
@@ -8,12 +8,12 @@ import autocti as ac
 from test_autogalaxy.aggregator.conftest import clean
 
 
-def test__fit_dataset_1d_randomly_drawn_via_pdf_gen_from(
+def test__cti_randomly_drawn_via_pdf_gen_from(
     dataset_1d_7, clocker_1d, samples_1d, model_1d
 ):
-    path_prefix = "aggregator_fit_dataset_1d_gen"
+    path_prefix = "aggregator_cti_gen"
 
-    database_file = path.join(conf.instance.output_path, "fit_dataset_1d.sqlite")
+    database_file = path.join(conf.instance.output_path, "cti.sqlite")
     result_path = path.join(conf.instance.output_path, path_prefix)
 
     clean(database_file=database_file, result_path=result_path)
@@ -28,30 +28,27 @@ def test__fit_dataset_1d_randomly_drawn_via_pdf_gen_from(
     agg = af.Aggregator.from_database(filename=database_file)
     agg.add_directory(directory=result_path)
 
-    fit_dataset_1d_agg = ac.agg.FitDataset1DAgg(aggregator=agg)
-    fit_dataset_1d_pdf_gen = fit_dataset_1d_agg.randomly_drawn_via_pdf_gen_from(
-        total_samples=2
-    )
+    cti_agg = ac.agg.CTIAgg(aggregator=agg)
+    cti_pdf_gen = cti_agg.randomly_drawn_via_pdf_gen_from(total_samples=2)
 
     i = 0
 
-    for fit_dataset_1d_gen in fit_dataset_1d_pdf_gen:
-        for fit_dataset_1d in fit_dataset_1d_gen:
+    for cti_gen in cti_pdf_gen:
+        for cti in cti_gen:
             i += 1
 
-            assert fit_dataset_1d.post_cti_data[0] == pytest.approx(1.0, 1.0e-4)
+            assert cti.trap_list[0].density == pytest.approx(0.1, 1.0e-4)
 
     assert i == 2
 
     clean(database_file=database_file, result_path=result_path)
 
 
-def test__fit_dataset_1d_all_above_weight_gen(
-    dataset_1d_7, clocker_1d, samples_1d, model_1d
-):
-    path_prefix = "aggregator_fit_dataset_1d_gen"
+#
+def test__cti_all_above_weight_gen(dataset_1d_7, clocker_1d, samples_1d, model_1d):
+    path_prefix = "aggregator_cti_gen"
 
-    database_file = path.join(conf.instance.output_path, "fit_dataset_1d.sqlite")
+    database_file = path.join(conf.instance.output_path, "cti.sqlite")
     result_path = path.join(conf.instance.output_path, path_prefix)
 
     clean(database_file=database_file, result_path=result_path)
@@ -66,16 +63,20 @@ def test__fit_dataset_1d_all_above_weight_gen(
     agg = af.Aggregator.from_database(filename=database_file)
     agg.add_directory(directory=result_path)
 
-    fit_dataset_1d_agg = ac.agg.FitDataset1DAgg(aggregator=agg)
-    fit_dataset_1d_pdf_gen = fit_dataset_1d_agg.all_above_weight_gen_from(
-        minimum_weight=-1.0
-    )
+    cti_agg = ac.agg.CTIAgg(aggregator=agg)
+    cti_pdf_gen = cti_agg.all_above_weight_gen_from(minimum_weight=-1.0)
 
     i = 0
 
-    for fit_dataset_1d_gen in fit_dataset_1d_pdf_gen:
-        for fit_dataset_1d in fit_dataset_1d_gen:
+    for cti_gen in cti_pdf_gen:
+        for cti in cti_gen:
             i += 1
+
[…]
