+++ test_autocti/aggregator/test_aggregator_dataset_1d.py
@@ -26,27 +26,25 @@ def test__dataset_gen_from(dataset_1d_7, clocker_1d, samples_1d, model_1d):
     agg.add_directory(directory=result_path)
 
     dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
-    dataset_gen = dataset_agg.dataset_gen_from()
+    dataset_gen = dataset_agg.dataset_list_gen_from()
 
-    for dataset in dataset_gen:
-        assert (dataset.data == dataset_1d_7.data).all()
+    for dataset_list in dataset_gen:
+        assert (dataset_list[0].data == dataset_1d_7.data).all()
 
     clean(database_file=database_file, result_path=result_path)
 
-    # analysis_list = [analysis, analysis]
-    #
-    # analysis = sum(analysis_list)
-    #
-    # search.fit(model=model_1d, analysis=analysis)
-    #
-    # agg = af.Aggregator.from_database(filename=database_file)
-    # agg.add_directory(directory=result_path)
-    #
-    # dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
-    # dataset_gen = dataset_agg.dataset_gen_from()
-    #
-    # print(list(dataset_gen))
-    # print(list(dataset_gen))
-    # print(list(dataset_gen))
-    # print(list(dataset_gen))
-    # aaaa
+    analysis_list = [analysis, analysis]
+
+    analysis = sum(analysis_list)
+
+    search.fit(model=model_1d, analysis=analysis)
+
+    agg = af.Aggregator.from_database(filename=database_file)
+    agg.add_directory(directory=result_path)
+
+    dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
+    dataset_gen = dataset_agg.dataset_list_gen_from()
+
+    for dataset_list in dataset_gen:
+        assert (dataset_list[0].data == dataset_1d_7.data).all()
+        assert (dataset_list[1].data == dataset_1d_7.data).all()
