+++ autocti/aggregator/fit_dataset_1d.py
@@ -11,14 +11,12 @@ import autofit as af
 
 from autocti.aggregator.abstract import AbstractAgg
 from autocti.aggregator.dataset_1d import _dataset_1d_list_from
-from autocti.dataset_1d.dataset_1d.settings import SettingsDataset1D
 
 
 def _fit_dataset_1d_list_from(
     fit: af.Fit,
     cti: Union[CTI1D, CTI2D],
-    clocker: Optional[AbstractClocker] = None,
-    settings_dataset: Optional[SettingsDataset1D] = None,
+    clocker_list: Optional[AbstractClocker] = None,
 ) -> List[FitDataset1D]:
     """
     Returns a `FitDataset1D` object from a PyAutoFit database `Fit` object and an instance of galaxies from a non-linear
@@ -39,14 +37,16 @@ def _fit_dataset_1d_list_from(
 
     from autocti.dataset_1d.fit import FitDataset1D
 
-    dataset_1d_list = _dataset_1d_list_from(fit=fit, settings_dataset=settings_dataset)
+    dataset_1d_list = _dataset_1d_list_from(fit=fit)
 
-    clocker_list = [fit.value(name="clocker")]
-    if clocker_list[0] is None:
-        clocker_list = fit.child_values(name="clocker")
+    if clocker_list is None:
+        clocker_list = [fit.value(name="clocker")]
+        if clocker_list[0] is None:
+            clocker_list = fit.child_values(name="clocker")
 
     post_cti_data_list = [
-        clocker.add_cti(data=dataset_1d.data, cti=cti) for dataset_1d, clocker in zip(dataset_1d_list, clocker_list)
+        clocker.add_cti(data=dataset_1d.data, cti=cti)
+        for dataset_1d, clocker in zip(dataset_1d_list, clocker_list)
     ]
 
     return [
@@ -62,8 +62,7 @@ class FitDataset1DAgg(AbstractAgg):
     def __init__(
         self,
         aggregator: af.Aggregator,
-        clocker: Optional[AbstractClocker] = None,
-        settings_dataset: Optional[SettingsDataset1D] = None,
+        clocker_list: Optional[List[AbstractClocker]] = None,
     ):
         """
         Wraps a PyAutoFit aggregator in order to create generators of fits to dataset_1d data, corresponding to the
@@ -77,9 +76,7 @@ class FitDataset1DAgg(AbstractAgg):
         settings_dataset
             The settings of the `Dataset1D` object fitted by the non-linear search.
         """
-        super().__init__(aggregator=aggregator, clocker=clocker)
-
-        self.settings_dataset = settings_dataset
+        super().__init__(aggregator=aggregator, clocker_list=clocker_list)
 
     def object_via_gen_from(self, fit, cti: Union[CTI1D, CTI2D]) -> FitDataset1D:
         """
@@ -99,6 +96,5 @@ class FitDataset1DAgg(AbstractAgg):
         return _fit_dataset_1d_list_from(
             fit=fit,
             cti=cti,
-            clocker=self.clocker,
-            settings_dataset=self.settings_dataset,
+            clocker_list=self.clocker_list,
         )
