+++ autocti/dataset_1d/model/analysis.py
@@ -91,8 +91,8 @@ class AnalysisDataset1D(af.Analysis):
     def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
         """
         Before the model-fit via the non-linear search begins, this routine saves attributes of the `Analysis` object
-        to the `pickles` folder such that they can be loaded after the analysis using PyAutoFit's database and aggregator
-        tools.
+        to the `pickles` folder such that they can be loaded after the analysis using PyAutoFit's database and
+        aggregator tools.
 
         For this analysis the following are output:
 

[…]
