+++ autocti/aggregator/__init__.py
@@ -2,3 +2,4 @@ from autocti.aggregator.cti import CTIAgg
 from autocti.aggregator.dataset_1d import Dataset1DAgg
 from autocti.aggregator.imaging_ci import ImagingCIAgg
 from autocti.aggregator.fit_dataset_1d import FitDataset1DAgg
+from autocti.aggregator.fit_imaging_ci import FitImagingCIAgg
