### 2023-06-03 4a8bd0d
fix test autogalaxy import
