+++ optional_requirements.txt
@@ -1,5 +1,4 @@
-numba
-lacosmic
+numba<=0.56.4
 pylops>=1.10.0,<=1.18.3
 pynufft
 ultranest==3.2.0
