+++ .github/workflows/main.yml
@@ -68,6 +68,7 @@ jobs:
         pip3 install -r PyAutoArray/requirements.txt
         pip3 install -r PyAutoArray/optional_requirements.txt
         pip3 install -r PyAutoCTI/requirements.txt
+        pip3 install -r PyAutoCTI/optional_requirements.txt
     - name: Run tests
       run: |
         export ROOT_DIR=`pwd`
