+++ optional_requirements.txt
@@ -1,7 +1,1 @@
 numba<=0.56.4
-pylops>=1.10.0,<=1.18.3
-pynufft
-ultranest==3.2.0
-zeus-mcmc==2.4.1
-jax==0.3.1
-jaxlib==0.3.0
