### 2023-06-07 396cc18
test commenting out allow_negative_pixels
