+++ autocti/clocker/two_d.py
@@ -206,35 +206,64 @@ class Clocker2D(AbstractClocker):
             parallel_window_offset = self.parallel_window_offset
             serial_window_offset = self.serial_window_offset
 
+        try:
+            image_post_cti = add_cti(
+                image=data,
+                parallel_ccd=parallel_ccd,
+                parallel_roe=self.parallel_roe,
+                parallel_traps=parallel_trap_list,
+                parallel_express=self.parallel_express,
+                parallel_window_offset=parallel_window_offset,
+                parallel_window_start=self.parallel_window_start,
+                parallel_window_stop=self.parallel_window_stop,
+                parallel_time_start=self.parallel_time_start,
+                parallel_time_stop=self.parallel_time_stop,
+                parallel_prune_n_electrons=self.parallel_prune_n_electrons,
+                parallel_prune_frequency=self.parallel_prune_frequency,
+                serial_ccd=serial_ccd,
+                serial_roe=self.serial_roe,
+                serial_traps=serial_trap_list,
+                serial_express=self.serial_express,
+                serial_window_offset=serial_window_offset,
+                serial_window_start=self.serial_window_start,
+                serial_window_stop=self.serial_window_stop,
+                serial_time_start=self.serial_time_start,
+                serial_time_stop=self.serial_time_stop,
+                serial_prune_n_electrons=self.serial_prune_n_electrons,
+                serial_prune_frequency=self.serial_prune_frequency,
+                allow_negative_pixels=self.allow_negative_pixels,
+                pixel_bounce=self.pixel_bounce,
+                verbosity=self.verbosity,
+            )
+        except TypeError:
+            image_post_cti = add_cti(
+                image=data,
+                parallel_ccd=parallel_ccd,
+                parallel_roe=self.parallel_roe,
+                parallel_traps=parallel_trap_list,
+                parallel_express=self.parallel_express,
+                parallel_window_offset=parallel_window_offset,
+                parallel_window_start=self.parallel_window_start,
+                parallel_window_stop=self.parallel_window_stop,
+                parallel_time_start=self.parallel_time_start,
+                parallel_time_stop=self.parallel_time_stop,
+                parallel_prune_n_electrons=self.parallel_prune_n_electrons,
+                parallel_prune_frequency=self.parallel_prune_frequency,
+                serial_ccd=serial_ccd,
+                serial_roe=self.serial_roe,
+                serial_traps=serial_trap_list,
+                serial_express=self.serial_express,
+                serial_window_offset=serial_window_offset,
+                serial_window_start=self.serial_window_start,
+                serial_window_stop=self.serial_window_stop,
+                serial_time_start=self.serial_time_start,
+                serial_time_stop=self.serial_time_stop,
+                serial_prune_n_electrons=self.serial_prune_n_electrons,
+                serial_prune_frequency=self.serial_prune_frequency,
+                pixel_bounce=self.pixel_bounce,
+                verbosity=self.verbosity,
+            )
 
-        image_post_cti = add_cti(
-            image=data,
-            parallel_ccd=parallel_ccd,
-            parallel_roe=self.parallel_roe,
-            parallel_traps=parallel_trap_list,
-            parallel_express=self.parallel_express,
-            parallel_window_offset=parallel_window_offset,
-            parallel_window_start=self.parallel_window_start,
-            parallel_window_stop=self.parallel_window_stop,
-            parallel_time_start=self.parallel_time_start,
-            parallel_time_stop=self.parallel_time_stop,
-            parallel_prune_n_electrons=self.parallel_prune_n_electrons,
[…]
