+++ paper/paper.bib
@@ -734,3 +734,24 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2010PASP..122.1035A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+@MISC{Anderson2018,
+       author = {{Anderson}, Jay and {Ryon}, Jenna E.},
+        title = "{Improving the Pixel-Based CTE-correction Model for ACS/WFC}",
+ howpublished = {Instrument Science Report ACS 2018-04, 37 pages},
+         year = 2018,
+        month = aug,
+        pages = {4},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018acs..rept....4A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@MISC{Anderson2021,
+       author = {{Anderson}, Jay and {Baggett}, Sylvia and {Kuhn}, Ben},
+        title = "{Updating the WFC3/UVIS CTE model and Mitigation Strategies}",
+     keywords = {Hubble Space Telescope, HST, STScI, Space Telescope Science Institute, WFC3, Wide Field Camera 3, UVIS, Charge-transfer efficiency, Pixel-based correction, Calibration, Photometry, Pipeline},
+ howpublished = {Instrument Science Report 2021-9, 44 pages},
+         year = 2021,
+        month = jun,
+        pages = {9},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021wfc..rept....9A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
