### 2023-06-07 baf170b
added pre_cti_data_residual_map method to ImagingCI
