### 2023-06-12 9273e9f
fix unit tests due to .png output rename
