+++ test_autocti/dataset_1d/model/test_visualizer_1d.py
@@ -98,9 +98,9 @@ def test__visualize_fit_combined(fit_1d_7, plot_path, plot_patch):
 
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
-    assert path.join(plot_path, "subplot_residual_map_list.png") in plot_patch.paths
+    assert path.join(plot_path, "subplot_residual_map.png") in plot_patch.paths
     assert (
-        path.join(plot_path, "subplot_chi_squared_map_list.png") not in plot_patch.paths
+        path.join(plot_path, "subplot_chi_squared_map.png") not in plot_patch.paths
     )
 
 
@@ -116,8 +116,8 @@ def test__visualize_fit_region_combined(fit_1d_7, plot_path, plot_patch):
 
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
-    assert path.join(plot_path, "subplot_residual_map_fpr_list.png") in plot_patch.paths
+    assert path.join(plot_path, "subplot_residual_map_fpr.png") in plot_patch.paths
     assert (
-        path.join(plot_path, "subplot_chi_squared_map_fpr_list.png")
+        path.join(plot_path, "subplot_chi_squared_map_fpr.png")
         not in plot_patch.paths
     )
