+++ test_autocti/charge_injection/model/test_visualizer_ci.py
@@ -115,9 +115,9 @@ def test__visualize_fit_combined(fit_ci_7x7, plot_path, plot_patch):
 
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
-    assert path.join(plot_path, "subplot_residual_map_list.png") in plot_patch.paths
+    assert path.join(plot_path, "subplot_residual_map.png") in plot_patch.paths
     assert (
-        path.join(plot_path, "subplot_chi_squared_map_list.png") not in plot_patch.paths
+        path.join(plot_path, "subplot_chi_squared_map.png") not in plot_patch.paths
     )
 
 
@@ -136,9 +136,9 @@ def test__visualize_fit_regions_combined(fit_ci_7x7, plot_path, plot_patch):
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
     assert (
-        path.join(plot_path, "subplot_data_parallel_fpr_list.png") in plot_patch.paths
+        path.join(plot_path, "subplot_data_parallel_fpr.png") in plot_patch.paths
     )
     assert (
-        path.join(plot_path, "subplot_data_logy_parallel_fpr_list.png")
+        path.join(plot_path, "subplot_data_logy_parallel_fpr.png")
         not in plot_patch.paths
     )
