+++ autocti/__init__.py
@@ -72,4 +72,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2023.7.7.3"
+__version__ = "2023.6.12.3"
