+++ test_autocti/charge_injection/model/test_visualizer_ci.py
@@ -116,9 +116,7 @@ def test__visualize_fit_combined(fit_ci_7x7, plot_path, plot_patch):
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
     assert path.join(plot_path, "subplot_residual_map.png") in plot_patch.paths
-    assert (
-        path.join(plot_path, "subplot_chi_squared_map.png") not in plot_patch.paths
-    )
+    assert path.join(plot_path, "subplot_chi_squared_map.png") not in plot_patch.paths
 
 
 def test__visualize_fit_regions_combined(fit_ci_7x7, plot_path, plot_patch):
@@ -135,9 +133,7 @@ def test__visualize_fit_regions_combined(fit_ci_7x7, plot_path, plot_patch):
 
     plot_path = path.join(plot_path, "fit_dataset_combined")
 
-    assert (
-        path.join(plot_path, "subplot_data_parallel_fpr.png") in plot_patch.paths
-    )
+    assert path.join(plot_path, "subplot_data_parallel_fpr.png") in plot_patch.paths
     assert (
         path.join(plot_path, "subplot_data_logy_parallel_fpr.png")
         not in plot_patch.paths
