+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 arcticpy==2.01.0
-autoconf==2023.7.7.2
+autoconf==2023.6.12.5
 numpydoc>=1.0.0
 pyprojroot==0.2.0
 sphinx==5.2.3
