+++ files/citations.md
@@ -10,7 +10,7 @@ This work uses the following software packages:
 
 - `Astropy` https://github.com/astropy/astropy [@astropy1] [@astropy2]
 - `corner.py` https://github.com/dfm/corner.py [@corner]
-- `dynesty` https://github.com/joshspeagle/dynesty [@dynesty]
+- `dynesty` https://github.com/joshspeagle/dynesty [@dynesty] [@dynesty1] [@dynesty2] [@dynesty3] [@dynesty4]
 - `emcee` https://github.com/dfm/emcee [@emcee]
 - `matplotlib` https://github.com/matplotlib/matplotlib [@matplotlib]
 - `numba` https://github.com/numba/numba [@numba]
