+++ files/citations.bib
@@ -229,3 +229,67 @@ archivePrefix = {arXiv},
     archivePrefix={arXiv},
     primaryClass={stat.ML}
 }
+
+
+@INPROCEEDINGS{dynesty1,
+       author = {{Skilling}, John},
+        title = "{Nested Sampling}",
+     keywords = {02.50.Tt, Inference methods},
+    booktitle = {Bayesian Inference and Maximum Entropy Methods in Science and Engineering: 24th International Workshop on Bayesian Inference and Maximum Entropy Methods in Science and Engineering},
+         year = 2004,
+       editor = {{Fischer}, Rainer and {Preuss}, Roland and {Toussaint}, Udo Von},
+       series = {American Institute of Physics Conference Series},
+       volume = {735},
+        month = nov,
+        pages = {395-405},
+          doi = {10.1063/1.1835238},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2004AIPC..735..395S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{dynesty2,
+abstract = {Nested sampling estimates directly how the likelihood function relates to prior mass. The evidence (alternatively the marginal likelihood, marginal density of the data, or the prior predictive) is immediately obtained by summation. It is the prime result of the computation, and is accompanied by an estimate of numerical uncertainty. Samples from the posterior distribution are an optional by-product, obtainable for any temperature. The method relies on sampling within a hard constraint on likelihood value, as opposed to the softened likelihood of annealing methods. Progress depends only on the shape of the "nested"contours of likelihood, and not on the likelihood values. This invariance (over monotonic relabelling) allows the method to deal with a class of phasechange problems which effectively defeat thermal annealing. {\textcopyright} 2006 International Society for Bayesian Analysis.},
+archivePrefix = {arXiv},
+arxivId = {arXiv:1011.1669v3},
+author = {Skilling, John},
+doi = {10.1214/06-BA127},
+eprint = {arXiv:1011.1669v3},
+isbn = {9788578110796},
+issn = {19360975},
+journal = {Bayesian Analysis},
+keywords = {Algorithm,Annealing,Bayesian computation,Evidence,Marginal likelihood,Model selection,Nest,Phase change},
+number = {4},
+pages = {833--860},
+pmid = {25246403},
+title = {{Nested sampling for general Bayesian computation}},
+url = {http://projecteuclid.org/euclid.ba/1340370944},
+volume = {1},
+year = {2006}
+}
+
+@article{dynesty3,
[…]
