+++ files/citations.tex
@@ -20,7 +20,7 @@ This work uses the following software packages:
 
 \item
 \href{https://github.com/joshspeagle/dynesty}{\textt{dynesty}}
-\citep{dynesty}
+\citep{dynesty, dynesty1, dynesty2, dynesty3, dynesty4}
 
 \item
 \href{https://github.com/dfm/emcee}{\textt{emcee}}
