### 2023-06-29 #4904 @jryon
Hi @Jammy2211, Thanks for addressing my comments!

While looking into your responses, I attempted to install PyAutoCTI from scratch using the instructions on the readthedocs. Because I cannot install GSL on my work machine, I followed the instructions on the Troubleshooting page. `arcticpy-no-gsl` installed without problems, however, `pip install autocti --ignore-installed arcticpy` still attempted to install `arcticpy`. It looks like the `--ignore-installed flag` does not acts as expected. `pip install autocti-no-gsl` was successful, however, the version installed was a release from January, not the most recent version. It seems there are possibly two issues here - installation instructions need to be updated and/or the autocti-no-gsl package needs a new release to PyPI.

I will try to install autocti on my personal machine to test the installation instructions with GSL in the next couple of days.
