+++ autocti/dataset_1d/model/analysis.py
@@ -119,7 +119,6 @@ class AnalysisDataset1D(af.Analysis):
             paths.save_object("dataset_full", self.dataset_full)
 
     def visualize_before_fit(self, paths: af.DirectoryPaths, model: af.Collection):
-
         region_list = self.region_list_from()
 
         visualizer = VisualizerDataset1D(visualize_path=paths.image_path)
@@ -202,7 +201,6 @@ class AnalysisDataset1D(af.Analysis):
         instance: af.ModelInstance,
         during_analysis: bool,
     ):
-
         if analyses is None:
             return
 
