+++ test_autocti/extract/two_d/serial/test_serial_calibration.py
@@ -254,7 +254,7 @@ def test__imaging_ci_from(imaging_ci_7x7):
 
     dataset = extract.imaging_ci_from(dataset=imaging_ci_7x7, rows=(0, 2))
 
-    assert (dataset.data.native == imaging_ci_7x7.image.native[0:2, :]).all()
+    assert (dataset.data.native == imaging_ci_7x7.data.native[0:2, :]).all()
     assert (dataset.noise_map.native == imaging_ci_7x7.noise_map.native[0:2, :]).all()
     assert (
         dataset.pre_cti_data.native == imaging_ci_7x7.pre_cti_data.native[0:2, :]
