### 2023-07-05 10f72b8
move arcticpy to optional requiremenets
