+++ optional_requirements.txt
@@ -1,1 +1,2 @@
 numba<=0.56.4
+arcticpy==2.2
