+++ requirements.txt
@@ -1,1 +1,0 @@
-arcticpy==2.2
