### 2023-07-05 #4904 @Jammy2211
I have done a release `2023.7.5.2` which makes it so you can install `arcticpy-no-gsl` and `autocti` separately.

It should work with a `pip install autocti==2023.7.5.2`

https://pyautocti.readthedocs.io/en/latest/installation/troubleshooting.html
