### 2023-07-05 #4904 @jryon
Great, I was able to install autocti and arcticpy after a little extra effort. I was getting the following scipy version conflict error when following this set of steps:

`conda create -n autocti astropy numpy scikit-image scikit-learn scipy`
`conda activate autocti`
`pip install autocti==2023.7.5.2`

Resulting in:
"ERROR: Cannot install autocti because these package versions have conflicting dependencies.

The conflict is caused by:
    autoarray 2023.7.5.2 depends on scipy<=1.8.1 and >=1.5.4
    autofit 2023.7.5.2 depends on scipy==1.8.1

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts"

Instead, I pinned scipy to the overlapping version from the error, and everything installed successfully:

`conda create -n autocti astropy numpy scikit-image scikit-learn scipy==1.8.1`
`conda activate autocti`
`pip install autocti==2023.7.5.2`
`pip install arcticpy_no_gsl`

Not clear to me if this scipy version issue is just my problem or if this would be something to note on the readthedocs pages.
