### 2023-07-07 #4904 @jryon
@Jammy2211 After taking a second look at everything, I have one final request, having to do with this item:

> > A short section describing how this software compares to other available packages would be useful. Related software packages are listed, but I think brief discussion is needed. For example, which telescopes are they intended for? Do they produce CTI models, simulate CTI effects, correct for CTI, or etc? Are they open source?
>
> To my knowledge, no other package exists for CTI calibration (e.g. arctic, C3TM, CDM03) are all packages for clocking and modeling CTI, but not calibrating it).
>
> I opted not to discuss other packages as I didnt want to say anything that might not be fully truly (e.g. maybe CDM03 has a calibration mode I am not aware of?). I can add a sentence if you require but I would rather leave it up to to the reader to checkout the package themselves.

The JOSS review checklist does require discussion of the "state of the field" in the paper, so please do add a sentence. Something to the effect of "While other packages to clock and model CTI exist (e.g., etc etc), to our knowledge, this is the first package for CTI calibration" would be sufficient.

@adonath Once that is done, I'd be happy to recommend this work for publication in JOSS.
