+++ test_autocti/aggregator/conftest.py
@@ -45,7 +45,8 @@ def make_samples_1d(model_1d):
     )
 
     return ac.m.MockSamples(
-        model=model_1d, sample_list=sample_list, max_log_likelihood_instance=cti
+        model=model_1d, sample_list=sample_list, max_log_likelihood_instance=cti,
+        gaussian_tuples=[(1.0, 2.0)] * model_1d.prior_count
     )
 
 
@@ -80,5 +81,6 @@ def make_samples_2d(model_2d):
     )
 
     return ac.m.MockSamples(
-        model=model_2d, sample_list=sample_list, max_log_likelihood_instance=cti
+        model=model_2d, sample_list=sample_list, max_log_likelihood_instance=cti,
+        gaussian_tuples=[(1.0, 2.0)] * model_2d.prior_count
     )
