+++ autocti/dataset_1d/model/analysis.py
@@ -88,7 +88,7 @@ class AnalysisDataset1D(af.Analysis):
             instance=instance, dataset=self.dataset
         )
 
-    def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
+    def save_attributes(self, paths: af.DirectoryPaths):
         """
         Before the model-fit via the non-linear search begins, this routine saves attributes of the `Analysis` object
         to the `pickles` folder such that they can be loaded after the analysis using PyAutoFit's database and
@@ -240,8 +240,5 @@ class AnalysisDataset1D(af.Analysis):
     def make_result(
         self,
         samples: af.SamplesPDF,
-        sigma=1.0,
-        use_errors=True,
-        use_widths=False,
     ) -> ResultDataset1D:
         return ResultDataset1D(samples=samples, analysis=self)
