+++ autocti/charge_injection/imaging/imaging.py
@@ -1,5 +1,6 @@
 import numpy as np
-from typing import Optional, List, Dict
+from pathlib import Path
+from typing import Optional, List, Dict, Union
 
 import autoarray as aa
 
@@ -171,38 +172,82 @@ class ImagingCI(aa.Imaging):
     @classmethod
     def from_fits(
         cls,
-        layout,
-        pixel_scales,
-        data_path=None,
-        image=None,
-        data_hdu=0,
-        noise_map_path=None,
-        noise_map_hdu=0,
-        noise_map_from_single_value=None,
-        pre_cti_data_path=None,
-        pre_cti_data_hdu=0,
-        pre_cti_data=None,
-        cosmic_ray_map_path=None,
-        cosmic_ray_map_hdu=0,
+        pixel_scales: aa.type.PixelScales,
+        layout : Layout2DCI,
+        data_path : Optional[Union[Path, str]] = None,
+        data_hdu : int = 0,
+        data: aa.Array2D = None,
+        noise_map_path  : Optional[Union[Path, str]] = None,
+        noise_map_hdu : int = 0,
+        noise_map_from_single_value : float = None,
+        pre_cti_data_path  : Optional[Union[Path, str]] = None,
+        pre_cti_data_hdu : int = 0,
+        pre_cti_data : aa.Array2D = None,
+        cosmic_ray_map_path  : Optional[Union[Path, str]] = None,
+        cosmic_ray_map_hdu : int = 0,
         settings_dict: Optional[Dict] = None,
     ) -> "ImagingCI":
-        if data_path is not None and image is None:
-            ci_image = aa.Array2D.from_fits(
+        """
+        Load charge injection imaging from multiple .fits file. 
+        
+        For each attribute of the charge injection data (e.g. `data`, `noise_map`, `pre_cti_data`) the path to 
+        the .fits and the `hdu` containing the data can be specified.
+
+        The `noise_map` assumes the noise value in each `data` value are independent, where these values are the
+        the RMS standard deviation error in each pixel.
+        
+        If the dataset has a mask associated with it (e.g. in a `mask.fits` file) the file must be loaded separately
+        via the `Mask2D` object and applied to the imaging after loading via fits using the `from_fits` method.
+ 
+        Parameters
+        ----------
+        layout
+            The layout of the charge injection, containing information like where the parallel and serial FPR and 
+            EPER are located.
+        pixel_scales
+            The (y,x) arcsecond-to-pixel units conversion factor of every pixel. If this is input as a `float`,
+            it is converted to a (float, float).
+        data_path
+            The path to the data .fits file containing the image data (e.g. '/path/to/data.fits').
+        data_hdu
+            The hdu the image data is contained in the .fits file specified by `data_path`.
+        data
+            Manually input the data as an `Array2D` instead of loading it via a .fits file.
+        noise_map_path
+            The path to the noise_map .fits file containing the noise_map (e.g. '/path/to/noise_map.fits').
+        noise_map_hdu
+            The hdu the noise map is contained in the .fits file specified by `noise_map_path`.
+        noise_map_from_single_value
+            Creates a `noise_map` of constant values if this is input instead of loading via .fits.
+        pre_cti_data_path
+            The path to the pre CTI data .fits file containing the image data (e.g. '/path/to/pre_cti_data.fits').
+        pre_cti_data_hdu
+            The hdu the pre cti data is contained in the .fits file specified by `pre_cti_data_path`.
+        pre_cti_data
+            Manually input the pre CTI data as an `Array2D` instead of loading it via a .fits file.
+        cosmic_ray_map_path
[…]
