+++ autocti/charge_injection/model/analysis.py
@@ -213,7 +213,15 @@ class AnalysisImagingCI(af.Analysis):
             The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
             visualization,and the pickled objects used by the aggregator output by this function.
         """
-        paths.save_object("dataset", self.dataset)
+
+        self.dataset.output_to_fits(
+            data_path=paths._files_path / "data.fits",
+            noise_map_path=paths._files_path / "noise_map.fits",
+            pre_cti_data_path=paths._files_path / "pre_cti_data.fits",
+            cosmic_ray_map_path=paths._files_path / "cosmic_ray_map.fits",
+            overwrite=True
+        )
+
         paths.save_object("clocker", self.clocker)
         paths.save_object("settings_cti", self.settings_cti)
         if self.dataset_full is not None:
