### 2023-07-20 5605e9c
imaging ci output mask to fits now works
