### 2023-07-20 8e823c2
Dataset 1D from_fits docstring
