### 2023-07-20 4ec768b
remove output to fits mask metho
