### 2023-07-20 124883a
save_attributes outputs data to .fits files
