### 2023-07-20 47da7ee
settings updated with docstrings, made dictable
