+++ test_autocti/model/test_settings_dict.py
@@ -8,9 +8,9 @@ import autocti as ac
 @pytest.fixture(name="settings_dict")
 def make_settings_dict():
     return {
-        'type': 'autocti.model.settings.SettingsCTI2D',
-        'parallel_total_density_range': None,
-        'serial_total_density_range': None
+        "type": "autocti.model.settings.SettingsCTI2D",
+        "parallel_total_density_range": None,
+        "serial_total_density_range": None,
     }
 
 
@@ -19,7 +19,6 @@ def test_settings_from_dict(settings_dict):
 
 
 def test_file():
-
     filename = "/tmp/temp.json"
     ac.SettingsCTI2D().output_to_json(filename)
 
