+++ autocti/dataset_1d/model/analysis.py
@@ -1,4 +1,3 @@
-import os
 from typing import List, Optional
 
 import autofit as af
@@ -6,7 +5,6 @@ import autofit as af
 from autocti.dataset_1d.dataset_1d.dataset_1d import Dataset1D
 from autocti.dataset_1d.fit import FitDataset1D
 from autocti.dataset_1d.model.visualizer import VisualizerDataset1D
-from autocti.model.result import ResultDataset
 from autocti.dataset_1d.model.result import ResultDataset1D
 from autocti.model.settings import SettingsCTI1D
 from autocti.clocker.one_d import Clocker1D
@@ -20,6 +18,30 @@ class AnalysisDataset1D(af.Analysis):
         settings_cti: SettingsCTI1D = SettingsCTI1D(),
         dataset_full: Optional[Dataset1D] = None,
     ):
+        """
+        Fits a CTI model to a 1D CTI dataset via a non-linear search.
+
+        The `Analysis` class defines the `log_likelihood_function` which fits the model to the dataset and returns the
+        log likelihood value defining how well the model fitted the data.
+
+        It handles many other tasks, such as visualization, outputting results to hard-disk and storing results in
+        a format that can be loaded after the model-fit is complete.
+
+        This class is used for model-fits which fit a CTI model via a `CTI1D` object to a charge injection
+        imaging dataset.
+
+        Parameters
+        ----------
+        dataset
+            The 1D CTI dataset that the model is fitted to.
+        clocker
+            The CTI arctic clocker used by the non-linear search and model-fit.
+        settings_cti
+            The settings controlling aspects of the CTI model in this model-fit.
+        dataset_full
+            The full dataset, which is visualized separate from the `dataset` that is fitted, which for example may
+            not have the FPR masked and thus enable visualization of the FPR.
+        """
         super().__init__()
 
         self.dataset = dataset
@@ -32,8 +54,8 @@ class AnalysisDataset1D(af.Analysis):
 
     def modify_before_fit(self, paths: af.DirectoryPaths, model: af.Collection):
         """
-        PyAutoFit calls this function immediately before the non-linear search begins, therefore it can be used to
-        perform tasks using the final model parameterization.
+        This function is called immediately before the non-linear search begins and performs final tasks and checks
+        before it begins.
 
         This function:
 
@@ -120,13 +142,27 @@ class AnalysisDataset1D(af.Analysis):
             pre_cti_data_path=dataset_path / "pre_cti_data.fits",
             overwrite=True,
         )
-        self.dataset.mask.output_to_fits(file_path=dataset_path / "mask.fits")
+        self.dataset.mask.output_to_fits(
+            file_path=dataset_path / "mask.fits", overwrite=True
+        )
+
+        if self.dataset_full is not None:
+            dataset_path = paths._files_path / "dataset_full"
+
+            self.dataset_full.output_to_fits(
+                data_path=dataset_path / "data.fits",
+                noise_map_path=dataset_path / "noise_map.fits",
+                pre_cti_data_path=dataset_path / "pre_cti_data.fits",
+                overwrite=True,
+            )
+            self.dataset_full.mask.output_to_fits(
+                file_path=dataset_path / "mask.fits", overwrite=True
+            )
+
         self.clocker.output_to_json(file_path=paths._files_path / "clocker.json")
         self.settings_cti.output_to_json(
             file_path=paths._files_path / "settings_cti.json"
         )
-        if self.dataset_full is not None:
-            paths.save_object("dataset_full", self.dataset_full)
 
     def visualize_before_fit(self, paths: af.DirectoryPaths, model: af.Collection):
         region_list = self.region_list_from()
