+++ optional_requirements.txt
@@ -1,2 +1,2 @@
 numba<=0.56.4
-arcticpy==2.2
+arcticpy==2.3
