+++ optional_requirements.txt
@@ -1,2 +1,3 @@
 numba<=0.56.4
+pyyaml==5.3.1
 arcticpy==2.4 
