### 2023-07-22 #4904 @Jammy2211
I have loosened the requirement on `scipy` in `autofit`: https://github.com/rhayes777/PyAutoFit/blob/main/requirements.txt

I have added the following paragraph under statement of need:

```
The majority of software packages for CTI, such as `arCTIc`, `C3TM` and `CDM03`, do not perform CTI calibration. They
instead model the CCD clocking procedure, including a CTI model describing the traps on a CCD, enabling the addition
and correction of CTI to imaging data. `PyAutoCTI` wraps one of these algorithm (specifically `arCTIc`) in order to
measure the CTI model from calibration data. The open-source software `Pyxel` has a "calibration mode" which also wraps
these algorithms to perform CTI calibration. However, `Pyxel` is more general purpose and has tools for the entire
CCD clocking procedure (e.g. bias, dark frames, flat fields). `PyAutoCTI` is therefore more specialized and
has a more extensive suite of tools for CTI calibration.
```
