+++ autocti/aggregator/dataset_1d.py
@@ -1,39 +1,146 @@
 from functools import partial
+
 import autofit as af
+import autoarray as aa
+
+from autocti.dataset_1d.dataset_1d.dataset_1d import Dataset1D
 
 
 def _dataset_1d_list_from(fit: af.Fit, use_dataset_full: bool = False):
     """
-     Returns a `Dataset1D` object from an aggregator's `SearchOutput` class, which we call an 'agg_obj' to describe
-     that it acts as the aggregator object for one result in the `Aggregator`. This uses the aggregator's generator
-     outputs such that the function can use the `Aggregator`'s map function to create a `Dataset1D` generator.
+    Returns a `Dataset1D` object from a `PyAutoFit` sqlite database `Fit` object.
 
-     The `Dataset1D` is created following the same method as the PyAutoCTI `Search` classes, including using the
-    `SettingsDataset1D` instance output by the Search to load inputs of the `Dataset1D`.
+    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
 
-     Parameters
-     ----------
-     fit
-         A PyAutoFit aggregator's SearchOutput object containing the generators of the results of model-fits.
-     use_dataset_full
-         If True, the full dataset is used to create the dataset list, else the masked dataset is used. This is used
-         when trying to plot regions of the dataset that are masked out (e.g. the FPR).
+    - The 1D dataset data as a .fits file (`dataset/data.fits`).
+    - The noise-map as a .fits file (`dataset/noise_map.fits`).
+    - The pre CTI data as a .fits file (`dataset/pre_cti_data.fits`).
+    - The layout of the `Dataset1D` data structure used in the fit (`dataset/layout.json`).
+    - The mask used to mask the `Dataset1D` data structure in the fit (`dataset/mask.fits`).
+
+    Each individual attribute can be loaded from the database via the `fit.value()` method.
+
+    This method combines all of these attributes and returns a `Dataset1D` object, has the mask applied to the
+    `Dataset1D` data structure and its settings updated to the values used by the model-fit.
+
+    If multiple `Dataset1D` objects were fitted simultaneously via analysis summing, the `fit.child_values()` method
+    is instead used to load lists of the data, noise-map, pre CTI data, layout and mask and combine them into a list of
+    `Dataset1D` objects.
+
+    If a `dataset_full` is input into the `Analysis` class when a model-fit is performed and therefore accessible
+    to the database, the input `use_dataset_full` can be switched in to load instead the full `Dataset1D` objects.
+
+    Parameters
+    ----------
+    fit
+        A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry in a sqlite database.
+    use_dataset_full
+        If a `dataset_full` is input into the `Analysis` class when a model-fit is performed and therefore accessible
+        to the database, the input `use_dataset_full` can be switched in to load instead the full `Dataset1D` objects.
     """
 
-    name = "dataset"
+    # TODO : Rich make a really nice API for making this concise...
 
     if use_dataset_full:
-        name = "dataset_full"
+        folder = "dataset_full"
+    else:
+        folder = "dataset"
 
-    dataset_list = [fit.value(name=name)]
-    if dataset_list[0] is None:
-        dataset_list = fit.child_values(name=name)
+    if fit.value(name=f"{folder}.data") is not None:
+        layout = fit.value(name=f"{folder}.layout")
+
+        data = aa.Array1D.from_primary_hdu(primary_hdu=fit.value(name=f"{folder}.data"))
+        noise_map = aa.Array1D.from_primary_hdu(
+            primary_hdu=fit.value(name=f"{folder}.noise_map")
+        )
+        pre_cti_data = aa.Array1D.from_primary_hdu(
[…]
