+++ test_autocti/aggregator/test_aggregator_dataset_1d.py
@@ -1,4 +1,6 @@
+import copy
 from os import path
+import pytest
 
 from autoconf import conf
 import autofit as af
@@ -7,44 +9,112 @@ import autocti as ac
 from test_autocti.aggregator.conftest import clean
 
 
-# def test__dataset_gen_from(dataset_1d_7, clocker_1d, samples_1d, model_1d):
-#     path_prefix = "aggregator_dataset_gen"
-#
-#     database_file = path.join(conf.instance.output_path, "dataset.sqlite")
-#     result_path = path.join(conf.instance.output_path, path_prefix)
-#
-#     clean(database_file=database_file, result_path=result_path)
-#
-#     search = ac.m.MockSearch(samples=samples_1d)
-#     search.paths = af.DirectoryPaths(path_prefix=path_prefix)
-#
-#     analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
-#
-#     search.fit(model=model_1d, analysis=analysis)
-#
-#     agg = af.Aggregator.from_database(filename=database_file)
-#     agg.add_directory(directory=result_path)
-#
-#     dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
-#     dataset_gen = dataset_agg.dataset_list_gen_from()
-#
-#     for dataset_list in dataset_gen:
-#         assert (dataset_list[0].data == dataset_1d_7.data).all()
-#
-#     clean(database_file=database_file, result_path=result_path)
-#
-#     analysis_list = [analysis, analysis]
-#
-#     analysis = sum(analysis_list)
-#
-#     search.fit(model=model_1d, analysis=analysis)
-#
-#     agg = af.Aggregator.from_database(filename=database_file)
-#     agg.add_directory(directory=result_path)
-#
-#     dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
-#     dataset_gen = dataset_agg.dataset_list_gen_from()
-#
-#     for dataset_list in dataset_gen:
-#         assert (dataset_list[0].data == dataset_1d_7.data).all()
-#         assert (dataset_list[1].data == dataset_1d_7.data).all()
+def test__dataset_gen_from__analysis_has_single_dataset(
+    dataset_1d_7, clocker_1d, samples_1d, model_1d
+):
+    path_prefix = "aggregator_dataset_1d_gen"
+
+    database_file = path.join(conf.instance.output_path, "dataset.sqlite")
+    result_path = path.join(conf.instance.output_path, path_prefix)
+
+    clean(database_file=database_file, result_path=result_path)
+
+    search = ac.m.MockSearch(samples=samples_1d)
+    search.paths = af.DirectoryPaths(path_prefix=path_prefix)
+
+    analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
+
+    search.fit(model=model_1d, analysis=analysis)
+
+    agg = af.Aggregator.from_database(filename=database_file)
+    agg.add_directory(directory=result_path)
+
+    dataset_agg = ac.agg.Dataset1DAgg(aggregator=agg)
+    dataset_gen = dataset_agg.dataset_list_gen_from()
+
+    for dataset in dataset_gen:
+        assert (dataset.data == dataset_1d_7.data).all()
+        assert dataset.layout.prescan[1] == pytest.approx(
+            dataset_1d_7.layout.prescan[1], 1.0e-4
+        )
+
+
+def test__dataset_gen_from__analysis_has_multi_dataset(
+    dataset_1d_7, clocker_1d, samples_1d, model_1d
+):
+    path_prefix = "aggregator_dataset_1d_gen"
+
[…]
