+++ test_autocti/aggregator/test_aggregator_dataset_1d.py
@@ -118,4 +118,3 @@ def test__dataset_gen_from__analysis_use_dataset_full(
         assert dataset_list[1].layout.prescan[1] == pytest.approx(
             dataset_1d_7.layout.prescan[1], 1.0e-4
         )
-
