### 2023-07-26 1fce628
uncommencted agg CTI tests
