### 2023-07-26 41da4d9
aggregator fit 1D dataset
