### 2023-08-09 4ca50e0
CTI aggregator now supports lists
