+++ autocti/aggregator/cti.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+import logging
 from typing import TYPE_CHECKING, List, Union
 
 if TYPE_CHECKING:
@@ -9,52 +10,98 @@ import autofit as af
 
 from autocti.aggregator.abstract import AbstractAgg
 
+logger = logging.getLogger(__name__)
 
-def _cti_from(fit: af.Fit) -> Union[CTI1D, CTI2D]:
+def _cti_from(fit: af.Fit) -> List[Union[CTI1D, CTI2D]]:
     """
-    Returns a `CTI` object from a PyAutoFit database `Fit` object and an instance of a CTI object from a non-linear
-    search model-fit.
+    Returns a list of `CTI` objects from a `PyAutoFit` sqlite database `Fit` object.
 
-    This function adds the `adapt_model_image` and `adapt_cti_image_path_dict` to the clocker before constructing
-    the `CTI`, if they were used.
+    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+
+    - The model and its best fit parameters (e.g. `model.json`).
+
+    Each individual attribute can be loaded from the database via the `fit.value()` method.
+
+    This method combines all of these attributes and returns a `CTI` object for a given non-linear search sample
+    (e.g. the maximum likelihood model). This includes associating adapt images with their respective galaxies.
+
+    If multiple `CTI` objects were fitted simultaneously via analysis summing, the `fit.child_values()` method
+    is instead used to load lists of Tracers. This is necessary if each Tracer has different galaxies (e.g. certain
+    parameters vary across each dataset and `Analysis` object).
 
     Parameters
     ----------
     fit
         A PyAutoFit database Fit object containing the generators of the results of model-fits.
-    clocker
-        The CTI arctic clocker used by the non-linear search and model-fit.
 
     Returns
     -------
     CTI
-        The cti computed via an instance of clocker.
+        The list of CTI models computed via an instance of clocker.
     """
 
-    return fit.instance
+    if len(fit.children) > 0:
+        logger.info(
+            """
+            Using database for a fit with multiple summed Analysis objects.
+
+            CTI objects do not fully support this yet (e.g. adapt images may not be set up correctly)
+            so proceed with caution!
+            """
+        )
+
+        return [fit.instance] * len(fit.children)
+
+    return [fit.instance]
 
 
 class CTIAgg(AbstractAgg):
     """
-    Wraps a PyAutoFit aggregator in order to create generators of ctis corresponding to the results of a non-linear
-    search model-fit.
+    Interfaces with an `PyAutoFit` aggregator object to create instances of `CTI` objects from the results
+    of a model-fit.
+
+    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+
+    - The model and its best fit parameters (e.g. `model.json`).
+
+    The `aggregator` contains the path to each of these files, and they can be loaded individually. This class
+    can load them all at once and create an `CTI` object via the `_cti_from` method.
+
+    This class's methods returns generators which create the instances of the `CTI` objects. This ensures
+    that large sets of results can be efficiently loaded from the hard-disk and do not require storing all
+    `CTI` instances in the memory at once.
+
+    For example, if the `aggregator` contains 3 model-fits, this class can be used to create a generator which
+    creates instances of the corresponding 3 `CTI` objects.
+
+    If multiple `CTI` objects were fitted simultaneously via analysis summing, the `fit.child_values()` method
+    is instead used to load lists of CTIs. This is necessary if each CTI has different galaxies (e.g. certain
+    parameters vary across each dataset and `Analysis` object).
+
+    This can be done manually, but this object provides a more concise API.
+
[…]
