### 2023-08-09 b436d1f
FitImagingCI agg works
