+++ autocti/aggregator/fit_imaging_ci.py
@@ -20,34 +20,50 @@ def _fit_imaging_ci_list_from(
     clocker_list: Optional[AbstractClocker] = None,
 ) -> List[FitImagingCI]:
     """
-    Returns a `FitImagingCI` object from a PyAutoFit database `Fit` object and an instance of galaxies from a non-linear
-    search model-fit.
+    Returns a list of `FitImagingCI` object from a `PyAutoFit` sqlite database `Fit` object.
+
+    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+
+    - The masked dataset (e.g. data / noise map / pre cti data) as .fits files (contained in `dataset` folder).
+    - The clocker used to add  CTI in the fit (`dataset/clocker.json`).
+    - The settings used for clocking CIT (contained in `dataset/settings_cti.json`).
+
+    Each individual attribute can be loaded from the database via the `fit.value()` method.
+
+    This method combines all of these attributes and returns a  list of `FitImagingCI` objects, by loading the masked
+    dataset adding CTI to its pre-cti data via the cti model and clocking and fitting the model image to the dataset.
+
+    If multiple `ImagingCI` objects were fitted simultaneously via analysis summing, the `fit.child_values()` method
+    is instead used to load lists of the datasets, perform the fit and return a list of `FitImagingCI` objects.
+
+    If a `dataset_full` is input into the `Analysis` class when a model-fit is performed and therefore accessible
+    to the database, the input `use_dataset_full` can be switched in to fit the full dataset instead.
 
     Parameters
     ----------
     fit
-        A PyAutoFit database Fit object containing the generators of the results of model-fits.
-    settings_dataset
-        The settings of the `ImagingCI` object fitted by the non-linear search.
-
-    Returns
-    -------
-    FitImagingCI
-        The fit to the imaging_ci dataset computed via an instance of galaxies.
+        A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry in a sqlite database.
+    use_dataset_full
+        If a `dataset_full` is input into the `Analysis` class when a model-fit is performed and therefore accessible
+        to the database, the input `use_dataset_full` can be switched in to load instead the full `ImagingCI` objects.
+    clocker_list
+        If input, overwrites the clocker used in the fit with a new clocker which is used to perform the fit.
     """
 
     from autocti.charge_injection.fit import FitImagingCI
 
-    imaging_ci_list = _imaging_ci_list_from(fit=fit, use_dataset_full=use_dataset_full)
+    dataset_list = _imaging_ci_list_from(fit=fit, use_dataset_full=use_dataset_full)
 
     if clocker_list is None:
-        clocker_list = [fit.value(name="clocker")]
-        if clocker_list[0] is None:
+
+        if not fit.children:
+            clocker_list = [fit.value(name="clocker")]
+        else:
             clocker_list = fit.child_values(name="clocker")
 
     post_cti_data_list = [
-        clocker.add_cti(data=dataset.data, cti=cti)
-        for dataset, clocker in zip(imaging_ci_list, clocker_list)
+        clocker.add_cti(data=dataset.pre_cti_data, cti=cti)
+        for dataset, clocker in zip(dataset_list, clocker_list)
     ]
 
     return [
@@ -55,7 +71,7 @@ def _fit_imaging_ci_list_from(
             dataset=dataset,
             post_cti_data=post_cti_data,
         )
-        for dataset, post_cti_data in zip(imaging_ci_list, post_cti_data_list)
+        for dataset, post_cti_data in zip(dataset_list, post_cti_data_list)
     ]
 
 
@@ -67,16 +83,43 @@ class FitImagingCIAgg(AbstractAgg):
[…]
