### 2023-08-09 791f9d1
refactored and clean up aggregator tests
