+++ test_autocti/aggregator/test_aggregator_cti.py
@@ -1,32 +1,23 @@
-from os import path
 import pytest
 
-from autoconf import conf
-import autofit as af
 import autocti as ac
 
-from test_autocti.aggregator.conftest import clean
+from test_autocti.aggregator.conftest import clean, aggregator_from
 
+database_file =  "db_cti_gen"
 
 def test__cti_randomly_drawn_via_pdf_gen_from(
     dataset_1d_7, clocker_1d, samples_1d, model_1d
 ):
-    path_prefix = "aggregator_cti_gen"
 
-    database_file = path.join(conf.instance.output_path, "cti.sqlite")
-    result_path = path.join(conf.instance.output_path, path_prefix)
-
-    clean(database_file=database_file, result_path=result_path)
-
-    search = ac.m.MockSearch(
-        samples=samples_1d, result=ac.m.MockResult(model=model_1d, samples=samples_1d)
-    )
-    search.paths = af.DirectoryPaths(path_prefix=path_prefix)
     analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
-    search.fit(model=model_1d, analysis=analysis)
 
-    agg = af.Aggregator.from_database(filename=database_file)
-    agg.add_directory(directory=result_path)
+    agg = aggregator_from(
+        database_file=database_file,
+        analysis=analysis,
+        model=model_1d,
+        samples=samples_1d,
+    )
 
     cti_agg = ac.agg.CTIAgg(aggregator=agg)
     cti_pdf_gen = cti_agg.randomly_drawn_via_pdf_gen_from(total_samples=2)
@@ -41,26 +32,19 @@ def test__cti_randomly_drawn_via_pdf_gen_from(
 
     assert i == 2
 
-    clean(database_file=database_file, result_path=result_path)
+    clean(database_file=database_file)
 
 
 def test__cti_all_above_weight_gen(dataset_1d_7, clocker_1d, samples_1d, model_1d):
-    path_prefix = "aggregator_cti_gen"
 
-    database_file = path.join(conf.instance.output_path, "cti.sqlite")
-    result_path = path.join(conf.instance.output_path, path_prefix)
-
-    clean(database_file=database_file, result_path=result_path)
-
-    search = ac.m.MockSearch(
-        samples=samples_1d, result=ac.m.MockResult(model=model_1d, samples=samples_1d)
-    )
-    search.paths = af.DirectoryPaths(path_prefix=path_prefix)
     analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
-    search.fit(model=model_1d, analysis=analysis)
 
-    agg = af.Aggregator.from_database(filename=database_file)
-    agg.add_directory(directory=result_path)
+    agg = aggregator_from(
+        database_file=database_file,
+        analysis=analysis,
+        model=model_1d,
+        samples=samples_1d,
+    )
 
     cti_agg = ac.agg.CTIAgg(aggregator=agg)
     cti_pdf_gen = cti_agg.all_above_weight_gen_from(minimum_weight=-1.0)
@@ -79,4 +63,4 @@ def test__cti_all_above_weight_gen(dataset_1d_7, clocker_1d, samples_1d, model_1
 
     assert i == 2
 
-    clean(database_file=database_file, result_path=result_path)
+    clean(database_file=database_file)
