### 2023-08-09 32294a1
aggregator refactor complete
