+++ test_autocti/aggregator/test_aggregator_fit_dataset_1d.py
@@ -6,10 +6,10 @@ from test_autocti.aggregator.conftest import clean, aggregator_from
 
 database_file = "db_fit_dataset_1d"
 
+
 def test__fit_dataset_1d_randomly_drawn_via_pdf_gen_from(
     dataset_1d_7, clocker_1d, samples_1d, model_1d
 ):
-
     analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
 
     agg = aggregator_from(
@@ -28,7 +28,6 @@ def test__fit_dataset_1d_randomly_drawn_via_pdf_gen_from(
 
     for fit_dataset_1d_gen in fit_dataset_1d_pdf_gen:
         for fit_dataset_1d_list in fit_dataset_1d_gen:
-
             i += 1
 
             assert fit_dataset_1d_list[0].post_cti_data[0] == pytest.approx(1.0, 1.0e-4)
@@ -41,7 +40,6 @@ def test__fit_dataset_1d_randomly_drawn_via_pdf_gen_from(
 def test__fit_dataset_1d_all_above_weight_gen(
     dataset_1d_7, clocker_1d, samples_1d, model_1d
 ):
-
     analysis = ac.AnalysisDataset1D(dataset=dataset_1d_7, clocker=clocker_1d)
 
     agg = aggregator_from(
