### 2023-09-14 a497726
Dataset1D uses new autofit outputs
