### 2023-09-14 12f42bc
ImagingCI uses new autofit outputs
