+++ autocti/charge_injection/model/analysis.py
@@ -239,57 +239,60 @@ class AnalysisImagingCI(af.Analysis):
             visualization,and the pickled objects used by the aggregator output by this function.
         """
 
-        self.clocker.output_to_json(file_path=paths._files_path / "clocker.json")
-        self.settings_cti.output_to_json(
-            file_path=paths._files_path / "settings_cti.json"
+        paths.save_json(
+            name="clocker",
+            object_dict=self.clocker,
         )
+
+        paths.save_json(
+            name="clocker",
+            object_dict=self.settings_cti,
+        )
+
 
         if conf.instance["visualize"]["plots"]["combined_only"]:
             return
 
-        dataset_path = paths._files_path / "dataset"
+        def output_dataset(dataset):
 
-        self.dataset.output_to_fits(
-            data_path=dataset_path / "data.fits",
-            noise_map_path=dataset_path / "noise_map.fits",
-            pre_cti_data_path=dataset_path / "pre_cti_data.fits",
-            cosmic_ray_map_path=dataset_path / "cosmic_ray_map.fits",
-            overwrite=True,
-        )
-        self.dataset.layout.output_to_json(
-            file_path=dataset_path / "layout.json",
-        )
-
-        if self.dataset.settings_dict is not None:
-            with open(dataset_path / "settings_dict.json", "w+") as outfile:
-                json.dump(self.dataset.settings_dict, outfile)
-
-        self.dataset.mask.output_to_fits(
-            file_path=dataset_path / "mask.fits", overwrite=True
-        )
-
-        if self.dataset_full is not None:
-            dataset_path = paths._files_path / "dataset_full"
-
-            self.dataset_full.output_to_fits(
-                data_path=dataset_path / "data.fits",
-                noise_map_path=dataset_path / "noise_map.fits",
-                pre_cti_data_path=dataset_path / "pre_cti_data.fits",
-                cosmic_ray_map_path=dataset_path / "cosmic_ray_map.fits",
-                overwrite=True,
+            paths.save_fits(
+                name="data",
+                hdu=dataset.data.hdu_for_output,
+                prefix="dataset",
             )
-
-            self.dataset.layout.output_to_json(
-                file_path=dataset_path / "layout.json",
+            paths.save_fits(
+                name="noise_map",
+                hdu=dataset.noise_map.hdu_for_output,
+                prefix="dataset",
+            )
+            paths.save_fits(
+                name="pre_cti_data",
+                hdu=dataset.pre_cti_data.hdu_for_output,
+                prefix="dataset",
+            )
+            paths.save_fits(
+                name="layout",
+                hdu=dataset.layout.hdu_for_output,
+                prefix="dataset",
+            )
+            paths.save_fits(
+                name="mask",
+                hdu=dataset.mask.hdu_for_output,
+                prefix="dataset",
             )
 
             if self.dataset.settings_dict is not None:
-                with open(dataset_path / "settings_dict.json", "w+") as outfile:
-                    json.dump(self.dataset.settings_dict, outfile)
 
-            self.dataset_full.mask.output_to_fits(
-                file_path=dataset_path / "mask.fits", overwrite=True
-            )
+                paths.save_json(
+                    name="settings_dict",
+                    object_dict=self.dataset.settings_dict,
+                    prefix="dataset",
+                )
+
+        output_dataset(dataset=self.dataset)
+
+        if self.dataset_full is not None:
+            output_dataset(dataset=self.dataset_full)
 
     def in_ascending_fpr_order_from(self, quantity_list, fpr_value_list):
         if not conf.instance["visualize"]["general"]["general"][
