### 2023-09-14 7d329fc
CTI2D now uses from_json
