+++ autocti/model/settings.py
@@ -2,6 +2,8 @@ from typing import List, Optional, Tuple
 
 from arcticpy import TrapInstantCapture
 
+from autoconf.dictable import from_dict, from_json, output_to_json
+
 from autocti import exc
 
 
@@ -18,6 +20,16 @@ class AbstractSettingsCTI:
             ):
                 raise exc.PriorException
 
+    @classmethod
+    def from_json(cls, file_path):
+        return from_json(file_path=file_path)
+
+    @classmethod
+    def from_dict(cls, dictionary):
+        return from_dict(dictionary)
+
+    def output_to_json(self, file_path):
+        output_to_json(obj=self, file_path=file_path)
 
 class SettingsCTI1D(AbstractSettingsCTI):
     def __init__(
