+++ autocti/clocker/abstract.py
@@ -4,6 +4,8 @@ from arcticpy import CCD
 from arcticpy import CCDPhase
 from arcticpy import TrapInstantCapture
 
+from autoconf.dictable import from_dict, from_json, output_to_json
+
 from autocti import exc
 
 
@@ -72,3 +74,14 @@ class AbstractClocker:
         """
         if not any(ccd_list):
             raise exc.ClockerException("No CCD object was passed to the add_cti method")
+
+    @classmethod
+    def from_json(cls, file_path):
+        return from_json(file_path=file_path)
+
+    @classmethod
+    def from_dict(cls, dictionary):
+        return from_dict(dictionary)
+
+    def output_to_json(self, file_path):
+        output_to_json(obj=self, file_path=file_path)
