+++ autocti/aggregator/imaging_ci.py
@@ -54,6 +54,7 @@ def _imaging_ci_list_from(fit: af.Fit, use_dataset_full: bool = False):
     dataset_list = []
 
     for fit in fit_list:
+
         layout = fit.value(name=f"{folder}.layout")
 
         data = aa.Array2D.from_primary_hdu(primary_hdu=fit.value(name=f"{folder}.data"))
@@ -63,9 +64,12 @@ def _imaging_ci_list_from(fit: af.Fit, use_dataset_full: bool = False):
         pre_cti_data = aa.Array2D.from_primary_hdu(
             primary_hdu=fit.value(name=f"{folder}.pre_cti_data")
         )
-        cosmic_ray_map = aa.Array2D.from_primary_hdu(
-            primary_hdu=fit.value(name=f"{folder}.cosmic_ray_map")
-        )
+        try:
+            cosmic_ray_map = aa.Array2D.from_primary_hdu(
+                primary_hdu=fit.value(name=f"{folder}.cosmic_ray_map")
+            )
+        except AttributeError:
+            cosmic_ray_map = None
 
         settings_dict = fit.value(name="dataset.settings_dict")
 
