+++ test_autocti/model/test_settings_dict.py
@@ -2,6 +2,8 @@ import os
 
 import pytest
 
+from autoconf.dictable import from_dict
+
 import autocti as ac
 
 
@@ -10,13 +12,14 @@ def make_settings_dict():
     return {
         "type": "autocti.model.settings.SettingsCTI2D",
         "class_path": "autocti.model.settings.SettingsCTI2D",
+        "arguments": {},
         "parallel_total_density_range": None,
         "serial_total_density_range": None,
     }
 
 
 def test_settings_from_dict(settings_dict):
-    assert isinstance(ac.SettingsCTI2D.from_dict(settings_dict), ac.SettingsCTI2D)
+    assert isinstance(from_dict(settings_dict), ac.SettingsCTI2D)
 
 
 def test_file():
