+++ optional_requirements.txt
@@ -1,4 +1,4 @@
 numba<=0.56.4
 pyyaml==5.3.1
-arcticpy==2.4
+arcticpy==2.5
 ultranest==3.6.2
