### 2023-09-19 bad670e
reverting to pre injeciton API
