+++ autocti/__init__.py
@@ -28,6 +28,7 @@ from .extract.settings import SettingsExtract
 from .extract.two_d.parallel.overscan import Extract2DParallelOverscan
 from .extract.two_d.parallel.fpr import Extract2DParallelFPR
 from .extract.two_d.parallel.eper import Extract2DParallelEPER
+from .extract.two_d.parallel.pre_injection import Extract2DParallelPreInjection
 from .extract.two_d.parallel.pedestal import Extract2DParallelPedestal
 from .extract.two_d.serial.prescan import Extract2DSerialPrescan
 from .extract.two_d.serial.overscan import Extract2DSerialOverscan
