+++ test_autocti/extract/two_d/parallel/test_parallel_pre_injection.py
@@ -37,3 +37,31 @@ def test__region_list_from__via_array_2d_list_from(
         array_2d_list[0].mask
         == np.array([[False, False, False], [False, False, False], [False, True, False]])
     ).all()
+
+
+def test__region_list_from__via_array_2d_list_from__pixels_from_end(
+    parallel_array, parallel_masked_array
+):
+    extract = ac.Extract2DParallelPreInjection(region_list=[(2, 4, 0, 3)], shape_2d=(5, 5))
+
+    array_2d_list = extract.array_2d_list_from(
+        array=parallel_array, settings=ac.SettingsExtract(pixels_from_end=1)
+    )
+
+    assert (array_2d_list[0] == np.array([[1.0, 1.0, 1.0]])).all()
+
+    array_2d_list = extract.array_2d_list_from(
+        array=parallel_array, settings=ac.SettingsExtract(pixels_from_end=2)
+    )
+    assert (array_2d_list[0] == np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]])).all()
+
+    extract = ac.Extract2DParallelPreInjection(region_list=[(1, 4, 0, 3), (5, 8, 0, 3)])
+
+    array_2d_list = extract.array_2d_list_from(
+        array=parallel_masked_array, settings=ac.SettingsExtract(pixels_from_end=1)
+    )
+
+    assert (
+        array_2d_list[0].mask
+        == np.array([[False, False, False]])
+    ).all()
