+++ autocti/extract/two_d/parallel/pre_injection.py
@@ -75,16 +75,17 @@ class Extract2DParallelPreInjection(Extract2DParallel):
         pixels = settings.pixels
         pixels_from_end = settings.pixels_from_end
 
-        # if pixels_from_end is not None:
-        #     pixels = (self.total_rows - pixels_from_end, self.total_rows)
+        if pixels_from_end is not None:
 
-        y0_min = min([region.y0 for region in self.region_list])
+            y0_min = min([region.y0 for region in self.region_list])
+
+            pixels = (y0_min - pixels_from_end, y0_min)
 
         y0 = pixels[1]
 
         print((pixels[0], y0, self.region_list[0].x0, self.region_list[0].x1))
 
-        return [aa.Region2D(region=(0 + pixels[0], y0, self.region_list[0].x0, self.region_list[0].x1))]
+        return [aa.Region2D(region=(pixels[0], y0, self.region_list[0].x0, self.region_list[0].x1))]
 
     def binned_region_1d_from(self, settings: SettingsExtract) -> aa.Region1D:
         """
