### 2023-10-15 #4904 @mwcraig
@Jammy2211 -- apologies for the delay in getting to this review.

I'm hitting some pretty substantial install issues, then an error on importing the package. I'll open an issue in the repo regarding the import issue.

The biggest immediate installation issue is that the instructions need to be clear that the highest scipy version supported is `1.8.1`, which means, for almost all users, that the highest python version is `3.10` because there are not wheels or conda packages for scipy for more recent versions of python.

On the conda side I would fix this by changing your environment creation instructions to something like:

```
mamba create -n pyautoci python=3.10 scipy=1.8
```

That should pin both packages in the new environment.

I'm not sure how to do that in the pip instructions....

Longer run, the version constraints on scipy and numpy in PyAutoArray are likely to start causing problems for increasing numbers of users over the next year or two.

To be explicit, I think it is sufficient for the JOSS paper to just explicitly address the scipy/python version restrictions.
