### 2023-10-15 Jammy2211/PyAutoCTI#53 @mwcraig
Error on import with latest released version
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4904) I tried installing `autocti`.

Though there were no errors during installation, I got an error in `welcome.py` which I can reproduce by simply doing `import autocti as ac`:

```python-traceback
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/mattcraig/mambaforge/envs/pyautoci/lib/python3.10/site-packages/autocti/__init__.py", line 10, in <module>
    from arcticpy import ROE
  File "/Users/mattcraig/mambaforge/envs/pyautoci/lib/python3.10/site-packages/arcticpy/__init__.py", line 1, in <module>
    from arcticpy.cti import add_cti, remove_cti, CTI_model_for_HST_ACS
  File "/Users/mattcraig/mambaforge/envs/pyautoci/lib/python3.10/site-packages/arcticpy/cti.py", line 8, in <module>
    from arcticpy.ccd import CCDPhase, CCD
  File "/Users/mattcraig/mambaforge/envs/pyautoci/lib/python3.10/site-packages/arcticpy/ccd.py", line 3, in <module>
    from autoconf.dictable import Dictable
ImportError: cannot import name 'Dictable' from 'autoconf.dictable' (/Users/mattcraig/mambaforge/envs/pyautoci/lib/python3.10/site-packages/autoconf/dictable.py)
```

## my installation steps

```bash
mamba create -n pyautoci python=3.10 scipy=1.8
conda activate pyautoci
pip install autocti
pip install arcticpy_no_gsl
```
