+++ test_autocti/model/test_settings_dict.py
@@ -3,26 +3,30 @@ import os
 import pytest
 
 import autocti as ac
+from autoconf.dictable import from_dict, output_to_json, from_json
 
 
 @pytest.fixture(name="settings_dict")
 def make_settings_dict():
     return {
-        "type": "autocti.model.settings.SettingsCTI2D",
-        "parallel_total_density_range": None,
-        "serial_total_density_range": None,
+        "class_path": "autocti.model.settings.SettingsCTI2D",
+        "type": "instance",
+        "arguments": {
+            "parallel_total_density_range": None,
+            "serial_total_density_range": None,
+        },
     }
 
 
 def test_settings_from_dict(settings_dict):
-    assert isinstance(ac.SettingsCTI2D.from_dict(settings_dict), ac.SettingsCTI2D)
+    assert isinstance(from_dict(settings_dict), ac.SettingsCTI2D)
 
 
 def test_file():
     filename = "/tmp/temp.json"
-    ac.SettingsCTI2D().output_to_json(filename)
+    output_to_json(ac.SettingsCTI2D(), filename)
 
     try:
-        assert isinstance(ac.SettingsCTI2D().from_json(filename), ac.SettingsCTI2D)
+        assert isinstance(from_json(filename), ac.SettingsCTI2D)
     finally:
         os.remove(filename)
