### 2023-10-16 Jammy2211/PyAutoCTI#53 @rhayes777
Apologies I think this is due to changes to the API in a library. I've pushed a fix although @Jammy2211 will have to advise when that will be released.
