### 2023-10-16 Jammy2211/PyAutoCTI#53 @rhayes777
@Jammy2211

Looks like arcticpy also needs updating:

`ImportError while loading conftest '/home/runner/work/PyAutoCTI/PyAutoCTI/PyAutoCTI/test_autocti/conftest.py'.
test_autocti/conftest.py:6: in <module>
    import autocti as ac
autocti/__init__.py:10: in <module>
    from arcticpy import ROE
/opt/hostedtoolcache/Python/3.9.[18](https://github.com/Jammy2211/PyAutoCTI/actions/runs/6529920316/job/17728316475#step:10:19)/x64/lib/python3.9/site-packages/arcticpy/__init__.py:1: in <module>
    from arcticpy.cti import add_cti, remove_cti, CTI_model_for_HST_ACS
/opt/hostedtoolcache/Python/3.9.18/x64/lib/python3.9/site-packages/arcticpy/cti.py:8: in <module>
    from arcticpy.ccd import CCDPhase, CCD
/opt/hostedtoolcache/Python/3.9.18/x64/lib/python3.9/site-packages/arcticpy/ccd.py:3: in <module>
    from autoconf.dictable import Dictable
E   ImportError: cannot import name 'Dictable' from 'autoconf.dictable' (/home/runner/work/PyAutoCTI/PyAutoCTI/PyAutoConf/autoconf/dictable.py)`
