### 2023-10-16 cc4157d
simplify requirements
