+++ optional_requirements.txt
@@ -1,5 +1,4 @@
-numba<=0.56.4
-pyyaml==5.3.1
+numba
 arcticpy==2.5
 ultranest==3.6.2
 
