### 2023-10-16 Jammy2211/PyAutoCTI#53 @Jammy2211
Just to add, I am implementing Python3.11 support this week :).
