### 2023-10-16 Jammy2211/PyAutoCTI#53 @mwcraig
> Just to add, I am implementing Python3.11 support this week :).

Wonderful! Let me know when it is ready to tes...
