+++ optional_requirements.txt
@@ -1,4 +1,5 @@
-numba
+numba<=0.56.4
+pyyaml
 arcticpy==2.5
 ultranest==3.6.2
 
