### 2023-10-17 dcf1bca
successfully removed np.ma from binned_array_1d_from
