### 2023-10-17 1715d1a
removed from stacked_array_2d_from
