### 2023-10-17 60fd9bf
remove np.ma from binned_array_1d_total_pixels_from
