### 2023-10-17 162043f
removed np.ma from _value_list_from
