### 2023-10-17 84304ad
removed np.ma from _value_list_of_lists_from
