+++ autocti/extract/two_d/parallel/abstract.py
@@ -140,20 +140,31 @@ class Extract2DParallel(Extract2D):
         value_lists = []
 
         arr_list = [
-            np.ma.array(data=array.native[region.slice], mask=array.mask[region.slice])
+            array.native[region.slice]
             for region in self.region_list_from(settings=settings)
         ]
 
-        for array_2d in arr_list:
+        mask_list = [
+            array.mask[region.slice]
+            for region in self.region_list_from(settings=settings)
+        ]
+
+        for array_2d, mask in zip(arr_list, mask_list):
             value_list = []
 
             for column_index in range(array_2d.shape[1]):
+
+                arr_extract = array_2d[:, column_index]
+                mask_extract = mask[:, column_index]
+
+                arr_unmasked = arr_extract[np.invert(mask_extract)]
+
                 if value_str == "median":
-                    value = float(np.ma.median(array_2d[:, column_index]))
+                    value = float(np.median(arr_unmasked))
                 elif value_str == "mean":
-                    value = float(np.ma.mean(array_2d[:, column_index]))
+                    value = float(np.mean(arr_unmasked))
                 elif value_str == "std":
-                    value = float(np.ma.std(array_2d[:, column_index]))
+                    value = float(np.std(arr_unmasked))
 
                 value_list.append(value)
 
