### 2023-10-17 03e6a71
use pytest to create inputs
