+++ test_autocti/extract/test_integration.py
@@ -3,7 +3,9 @@ import pytest
 
 import autocti as ac
 
-def test__stacked_array_2d_from():
+@pytest.fixture(name="parallel_masked_array")
+def make_parallel_masked_array(parallel_array):
+
     parallel_array = ac.Array2D.no_mask(
         values=[
             [0.0, 0.0, 0.0, 4.0, 5.0],
@@ -36,7 +38,36 @@ def test__stacked_array_2d_from():
         pixel_scales=1.0,
     )
 
-    arr = ac.Array2D(values=parallel_array.native, mask=mask)
+    return ac.Array2D(values=parallel_array.native, mask=mask)
+
+
+# @pytest.fixture(name="serial_array")
+# def make_serial_array():
+#     return ac.Array2D.no_mask(
+#         values=[
+#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+#         ],
+#         pixel_scales=1.0,
+#     )
+#
+#
+# @pytest.fixture(name="serial_masked_array")
+# def make_serial_masked_array(serial_array):
+#     mask = ac.Mask2D(
+#         mask=[
+#             [False, False, False, False, False, True, False, False, False, False],
+#             [False, False, True, False, False, False, True, False, False, False],
+#             [False, False, False, False, False, False, False, True, False, False],
+#         ],
+#         pixel_scales=1.0,
+#     )
+#
+#     return ac.Array2D(values=serial_array.native, mask=mask)
+
+
+def test__stacked_array_2d_from(parallel_masked_array):
 
     extract = ac.Extract2DParallelEPER(
         region_list=[
@@ -46,7 +77,7 @@ def test__stacked_array_2d_from():
         ])
 
     stacked_array = extract.stacked_array_2d_from(
-        array=arr,
+        array=parallel_masked_array,
         settings=ac.SettingsExtract(pixels=(0, 2))
     )
 
@@ -54,47 +85,14 @@ def test__stacked_array_2d_from():
     assert stacked_array[1] == pytest.approx([6.5       , 6.66666667, 6.66666667, 0.        , 2.], 1.0e-4)
 
     stacked_array_2d_total_pixels = extract.stacked_array_2d_total_pixels_from(
-        array=arr,
+        array=parallel_masked_array,
         settings=ac.SettingsExtract(pixels=(0, 2))
     )
 
     assert stacked_array_2d_total_pixels[0] == pytest.approx([3, 3, 2, 3, 2], 1.0e-4)
     assert stacked_array_2d_total_pixels[1] == pytest.approx([2, 3, 3, 0, 1], 1.0e-4)
 
-def test__binned_array_1d_from():
-    parallel_array = ac.Array2D.no_mask(
-        values=[
-            [0.0, 0.0, 0.0, 4.0, 5.0],
-            [1.0, 1.0, 1.0, 9.0, 3.0],
-            [2.0, 2.0, 2.0, 8.0, 2.0],
-            [3.0, 3.0, 3.0, 7.0, 1.0],
-            [4.0, 4.0, 4.0, 6.0, 0.0],
-            [5.0, 5.0, 5.0, 5.0, 5.0],
-            [6.0, 6.0, 6.0, 4.0, 4.0],
-            [7.0, 7.0, 7.0, 3.0, 3.0],
-            [8.0, 8.0, 8.0, 9.0, 5.0],
[…]
