### 2023-10-17 749b4d0
setup serial integration test
