+++ test_autocti/extract/test_integration.py
@@ -1,10 +1,10 @@
-
+import numpy as np
 import pytest
 
 import autocti as ac
 
 @pytest.fixture(name="parallel_masked_array")
-def make_parallel_masked_array(parallel_array):
+def make_parallel_masked_array():
 
     parallel_array = ac.Array2D.no_mask(
         values=[
@@ -41,30 +41,32 @@ def make_parallel_masked_array(parallel_array):
     return ac.Array2D(values=parallel_array.native, mask=mask)
 
 
-# @pytest.fixture(name="serial_array")
-# def make_serial_array():
-#     return ac.Array2D.no_mask(
-#         values=[
-#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
-#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
-#             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
-#         ],
-#         pixel_scales=1.0,
-#     )
-#
-#
-# @pytest.fixture(name="serial_masked_array")
-# def make_serial_masked_array(serial_array):
-#     mask = ac.Mask2D(
-#         mask=[
-#             [False, False, False, False, False, True, False, False, False, False],
-#             [False, False, True, False, False, False, True, False, False, False],
-#             [False, False, False, False, False, False, False, True, False, False],
-#         ],
-#         pixel_scales=1.0,
-#     )
-#
-#     return ac.Array2D(values=serial_array.native, mask=mask)
+@pytest.fixture(name="serial_masked_array")
+def make_serial_masked_array():
+    
+    serial_array = ac.Array2D.no_mask(
+        values=[
+            [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+            [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+            [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
+            [5.0, 2.0, 4.0, 6.0, 4.0, 7.0, 6.0, 9.0, 3.0, 5.0],
+            [4.0, 3.0, 2.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0],
+        ],
+        pixel_scales=1.0,
+    )
+    
+    mask = ac.Mask2D(
+        mask=[
+            [False, False, False, False, False, True, False, False, False, False],
+            [False, False, True, False, False, False, True, False, False, False],
+            [False, False, False, False, False, False, False, True, False, False],
+            [True, True, True, False, False, True, False, True, True, True],
+            [True, False, True, True, True, False, True, False, True, True],
+        ],
+        pixel_scales=1.0,
+    )
+
+    return ac.Array2D(values=serial_array.native, mask=mask)
 
 
 def test__stacked_array_2d_from(parallel_masked_array):
@@ -113,7 +115,7 @@ def test__binned_array_1d_from(parallel_masked_array):
 
     assert binned_array_1d_total_pixels == pytest.approx([13, 9], 1.0e-4)
 
-def test__abstract___value_list_from(parallel_masked_array):
+def test__parallel__value_list_from(parallel_masked_array):
 
     extract = ac.Extract2DParallelEPER(
         region_list=[
@@ -139,3 +141,30 @@ def test__abstract___value_list_from(parallel_masked_array):
[…]
