### 2023-10-17 3294774
value_list_from in serial extract removed np.ma
