### 2023-10-17 4cf6cf8
value_list_of_lists_from in serial extract removed np.ma
