+++ test_autocti/extract/test_integration.py
@@ -165,6 +165,6 @@ def test__serial__value_list_from(serial_masked_array):
         settings=ac.SettingsExtract(pixels=(0, 2))
     )
 
-    assert median_list_of_lists[0] == pytest.approx([3.5, 3.5, 3.5, 5.0, np.nan], 1.0e-4)
+    assert np.nan_to_num(median_list_of_lists[0]) == pytest.approx([3.5, 3.5, 3.5, 5.0, 0.0], 1.0e-4)
     assert median_list_of_lists[1] == pytest.approx([6.5, 7.0, 6.0, 6.0, 3.0], 1.0e-4)
-    assert median_list_of_lists[2] == pytest.approx([8.5, 8.5, 8.5, np.nan, np.nan], 1.0e-4)
+    assert np.nan_to_num(median_list_of_lists[2]) == pytest.approx([8.5, 8.5, 8.5, 0.0, 0.0], 1.0e-4)
