### 2023-10-17 085bedb
np.ma removed
