+++ test_autocti/extract/test_integration.py
@@ -3,9 +3,9 @@ import pytest
 
 import autocti as ac
 
+
 @pytest.fixture(name="parallel_masked_array")
 def make_parallel_masked_array():
-
     parallel_array = ac.Array2D.no_mask(
         values=[
             [0.0, 0.0, 0.0, 4.0, 5.0],
@@ -43,7 +43,6 @@ def make_parallel_masked_array():
 
 @pytest.fixture(name="serial_masked_array")
 def make_serial_masked_array():
-    
     serial_array = ac.Array2D.no_mask(
         values=[
             [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0],
@@ -54,7 +53,7 @@ def make_serial_masked_array():
         ],
         pixel_scales=1.0,
     )
-    
+
     mask = ac.Mask2D(
         mask=[
             [False, False, False, False, False, True, False, False, False, False],
@@ -70,38 +69,33 @@ def make_serial_masked_array():
 
 
 def test__stacked_array_2d_from(parallel_masked_array):
-
     extract = ac.Extract2DParallelEPER(
-        region_list=[
-            (1, 3, 0, 5),
-            (4, 6, 0, 5),
-            (7, 8, 0, 5)
-        ])
-
-    stacked_array = extract.stacked_array_2d_from(
-        array=parallel_masked_array,
-        settings=ac.SettingsExtract(pixels=(0, 2))
+        region_list=[(1, 3, 0, 5), (4, 6, 0, 5), (7, 8, 0, 5)]
     )
 
-    assert stacked_array[0] == pytest.approx([5.66666667, 5.66666667, 7.        , 6.66666667, 2.5], 1.0e-4)
-    assert stacked_array[1] == pytest.approx([6.5       , 6.66666667, 6.66666667, 0.        , 2.], 1.0e-4)
+    stacked_array = extract.stacked_array_2d_from(
+        array=parallel_masked_array, settings=ac.SettingsExtract(pixels=(0, 2))
+    )
+
+    assert stacked_array[0] == pytest.approx(
+        [5.66666667, 5.66666667, 7.0, 6.66666667, 2.5], 1.0e-4
+    )
+    assert stacked_array[1] == pytest.approx(
+        [6.5, 6.66666667, 6.66666667, 0.0, 2.0], 1.0e-4
+    )
 
     stacked_array_2d_total_pixels = extract.stacked_array_2d_total_pixels_from(
-        array=parallel_masked_array,
-        settings=ac.SettingsExtract(pixels=(0, 2))
+        array=parallel_masked_array, settings=ac.SettingsExtract(pixels=(0, 2))
     )
 
     assert stacked_array_2d_total_pixels[0] == pytest.approx([3, 3, 2, 3, 2], 1.0e-4)
     assert stacked_array_2d_total_pixels[1] == pytest.approx([2, 3, 3, 0, 1], 1.0e-4)
 
-def test__binned_array_1d_from(parallel_masked_array):
 
+def test__binned_array_1d_from(parallel_masked_array):
     extract = ac.Extract2DParallelEPER(
-        region_list=[
-            (1, 3, 0, 5),
-            (4, 6, 0, 5),
-            (7, 8, 0, 5)
-        ])
+        region_list=[(1, 3, 0, 5), (4, 6, 0, 5), (7, 8, 0, 5)]
+    )
 
     binned_array_1d = extract.binned_array_1d_from(
         array=parallel_masked_array, settings=ac.SettingsExtract(pixels=(0, 2))
@@ -115,19 +109,16 @@ def test__binned_array_1d_from(parallel_masked_array):
 
[…]
