+++ test_autocti/aggregator/test_aggregator_cti.py
@@ -28,7 +28,7 @@ def test__cti_randomly_drawn_via_pdf_gen_from(
         for cti in cti_gen:
             i += 1
 
-            assert cti.cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
+            assert cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
 
     assert i == 2
 
@@ -55,10 +55,10 @@ def test__cti_all_above_weight_gen(dataset_1d_7, clocker_1d, samples_1d, model_1
             i += 1
 
             if i == 1:
-                assert cti.cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
+                assert cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
 
             if i == 2:
-                assert cti.cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
+                assert cti.trap_list[0].density == pytest.approx(10.0, 1.0e-4)
 
     assert i == 2
 
