+++ autocti/__init__.py
@@ -75,4 +75,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2023.9.18.4"
+__version__ = "2023.10.23.2"
