### 2023-10-23 bf6a3db
doc specifies Python3.11 support
