### 2023-10-23 #4904 @Jammy2211
@mwcraig Support for Python3.11 is included in todays release: https://github.com/Jammy2211/PyAutoCTI/releases/tag/2023.10.23.3

Everything installs fine on our build server, so here's hoping is smooth for you... let us know if there are issues.
