### 2023-10-26 #4904 @mwcraig
@Jammy2211 -- installation proceeds smoothly now with python 3.11, but I'm still getting the import error reported here: https://github.com/Jammy2211/PyAutoCTI/issues/53

I should have been more specific there (happy to open a new issue) -- I installed the no-GSL version of arcticpy because I got an error about GSL when installing the regular arcticpy.

The import error being raised is one coming from arcticpy, I think.
