### 2023-10-26 Jammy2211/PyAutoCTI#53 @Jammy2211
`pip install arcticpy_no_gsl==2.5` should fix the issue.
