### 2023-11-07 4efca3d
method to mask rows of read noise persistance added
