### 2023-11-07 0aed45d
docstirng for read noise persistence mask
