+++ autocti/mask/mask_2d.py
@@ -325,7 +325,41 @@ class Mask2D(aa.Mask2D):
         settings: "SettingsMask2D",
         pixel_scales: aa.type.PixelScales,
         invert: bool = False,
-    ):
+    ) -> "Mask2D":
+        """
+        Read noise persistence is a feature of CCDs whereby the signal from a high signal pixel (e.g. cosmic ray) can
+        persist into the signal of subsequent rows of pixels.
+
+        This leads to a 'streak' of signal values in the x direction, which typically need to be masked out.
+
+        This function produces a read noise persistence mask from a list of row values, where the values are the
+        average signal in each row of the image after other features (e.g. the charge injection) have been removed.
+
+        All rows with a signal above an input `read_noise_persistence_threshold` are masked out, where this threshold
+        should be estimated from the data itself or based on the CCD's properties.
+
+        Parameters
+        ----------
+        layout
+            The layout of the CCD (where the parallel overscan begins and ends, where the charge injection
+            regions are, etc.).
+        row_value_list
+            The average signal in each row of the image after other features (e.g. the charge injection) have been
+            removed.
+        read_noise_persistence_threshold
+            The threshold above which a row is masked out, assuming that this threshold means that a signal is
+            so bright that it must be due to read noise persistence.
+        settings
+            The settings of the mask (e.g. the number of pixels to mask out).
+        pixel_scales
+            The pixel scales of the CCD in arc-seconds per pixel, which is passed to the mask.
+        invert
+            If `True`, the mask is inverted such that all pixels that are masked are unmasked and visa versa.
+
+        Returns
+        -------
+        The read noise persistence mask.
+        """
         mask_row = [
             row_value > read_noise_persistence_threshold for row_value in row_value_list
         ]
@@ -336,4 +370,7 @@ class Mask2D(aa.Mask2D):
             if mask_row[y]:
                 mask[y, :] = True
 
-        return mask
+        if invert:
+            mask = np.invert(mask)
+
+        return Mask2D(mask=mask.astype("bool"), pixel_scales=pixel_scales)
