### 2023-11-07 45ff2a3
can mask rows in front and behind each detected read noise persistence row
