### 2023-11-09 fe1e1ba
added function to get mask_from an Extract2D object
