### 2023-11-09 5ebd928
added function figures_1d_data_binned
