### 2023-11-09 1150b01
unit test added, raise exeption for variable FPR
