### 2023-11-09 88aed5e
added plot to visualizer
