+++ autocti/charge_injection/model/visualizer.py
@@ -153,6 +153,14 @@ class VisualizerImagingCI(Visualizer):
                         filename_suffix=f"{filename_suffix}_{region}",
                     )
 
+                multi_plotter.subplot_of_figure(
+                    func_name="figures_1d_data_binned",
+                    figure_name="columns_fpr",
+                    filename_suffix=f"{filename_suffix}",
+                )
+
+                ddd
+
             except (exc.RegionException, TypeError, ValueError):
                 logger.info(
                     f"VISUALIZATION - Could not visualize the ImagingCI 1D {region}"
