### 2023-11-09 bad309b
all plots added to ImagingCIPloter
