### 2023-11-09 5969ba2
fix issues with masking of data in plotter
