+++ autocti/charge_injection/model/visualizer.py
@@ -153,18 +153,24 @@ class VisualizerImagingCI(Visualizer):
                         filename_suffix=f"{filename_suffix}_{region}",
                     )
 
-                for figure_name in ["columns_fpr"]:
-                    if should_plot(figure_name):
-                        multi_plotter.subplot_of_figure(
-                            func_name="figures_1d_data_binned",
-                            figure_name=figure_name,
-                            filename_suffix=f"{filename_suffix}",
-                        )
-
             except (exc.RegionException, TypeError, ValueError):
                 logger.info(
                     f"VISUALIZATION - Could not visualize the ImagingCI 1D {region}"
                 )
+
+        if folder_suffix == "_full":
+            for figure_name in [
+                "rows_fpr",
+                "rows_no_fpr",
+                "columns_fpr",
+                "columns_no_fpr",
+            ]:
+                if should_plot(figure_name):
+                    multi_plotter.subplot_of_figure(
+                        func_name="figures_1d_data_binned",
+                        figure_name=figure_name,
+                        filename_suffix=f"{filename_suffix}",
+                    )
 
     def visualize_fit(self, fit, during_analysis, folder_suffix: str = ""):
         def should_plot(name):
