### 2023-11-09 292e791
config for plots
