### 2023-11-09 2913c3a
FitImagingCIPlotter has binned plots
