+++ autocti/charge_injection/model/visualizer.py
@@ -369,6 +369,19 @@ class VisualizerImagingCI(Visualizer):
                     f"VISUALIZATION - Could not visualize the ImagingCI 1D {region}"
                 )
 
+        if folder_suffix == "_full":
+            for figure_name in [
+                "rows_fpr",
+                "rows_no_fpr",
+                "columns_fpr",
+                "columns_no_fpr",
+            ]:
+                if should_plot(figure_name):
+                    multi_plotter.subplot_of_figure(
+                        func_name="figures_1d_data_binned",
+                        figure_name=figure_name,
+                    )
+
     def visualize_fit_in_fits(self, fit):
         mat_plot_2d = self.mat_plot_2d_from(subfolders="fit_imaging/fit", format="fit")
 
