+++ autocti/charge_injection/plot/fit_ci_plotters.py
@@ -423,10 +423,15 @@ class FitImagingCIPlotter(Plotter):
         )
 
         serial_prescan = self.dataset.layout.extract.serial_prescan.serial_prescan
-        fpr_mask[serial_prescan.y0:serial_prescan.y1, serial_prescan.x0:serial_prescan.x1] = True
+        fpr_mask[
+            serial_prescan.y0 : serial_prescan.y1, serial_prescan.x0 : serial_prescan.x1
+        ] = True
 
         serial_overscan = self.dataset.layout.extract.serial_overscan.serial_overscan
-        fpr_mask[serial_overscan.y0:serial_overscan.y1, serial_overscan.x0:serial_overscan.x1] = True
+        fpr_mask[
+            serial_overscan.y0 : serial_overscan.y1,
+            serial_overscan.x0 : serial_overscan.x1,
+        ] = True
 
         if any(
             [
