### 2023-11-13 ff14f3a
added method to extract pixel indexes of rows and columns
