### 2023-11-13 d9d204c
added method to extract pixel indexes of rows and columsn for SerialExtract objects
