### 2023-11-16 4033943
read_noise_persistence -> readout_persistence
