### 2023-11-21 5744c57
fpr_mask now made in abstract_plotters
