+++ autocti/plot/abstract_plotters.py
@@ -10,7 +10,9 @@ from autoarray.plot.abstract_plotters import AbstractPlotter
 
 from autocti.plot.get_visuals.one_d import GetVisuals1D
 from autocti.plot.get_visuals.two_d import GetVisuals2D
+from autocti.extract.settings import SettingsExtract
 
+from autocti import exc
 
 class Plotter(AbstractPlotter):
     def __init__(
@@ -107,3 +109,36 @@ class Plotter(AbstractPlotter):
             return True
 
         return False
+
+    def fpr_mask_from(self):
+
+        fpr_size = self.dataset.layout.parallel_rows_within_regions[0]
+
+        if any(
+            [
+                fpr_size != fpr_size_of_row
+                for fpr_size_of_row in self.dataset.layout.parallel_rows_within_regions
+            ]
+        ):
+            raise exc.PlottingException(
+                "The FPR in this dataset have a variable number of rows. This means that masknig the FPR in the"
+                "figures_1d_data_binned method is not supported."
+            )
+
+        fpr_mask = self.dataset.layout.extract.parallel_fpr.mask_from(
+            settings=SettingsExtract(pixels=(0, fpr_size)),
+            pixel_scales=self.dataset.pixel_scales,
+        )
+
+        serial_prescan = self.dataset.layout.extract.serial_prescan.serial_prescan
+        fpr_mask[
+            serial_prescan.y0 : serial_prescan.y1, serial_prescan.x0 : serial_prescan.x1
+        ] = True
+
+        serial_overscan = self.dataset.layout.extract.serial_overscan.serial_overscan
+        fpr_mask[
+            serial_overscan.y0 : serial_overscan.y1,
+            serial_overscan.x0 : serial_overscan.x1,
+        ] = True
+
+        return fpr_mask
