### 2023-11-27 1edde8a
refactored plots so theyc an be made individually easier
