+++ autocti/clocker/two_d.py
@@ -1,10 +1,11 @@
 import numpy as np
-from typing import Optional
+from typing import List, Optional
 
 from arcticpy import add_cti
 from arcticpy import remove_cti
 
 from arcticpy import ROE
+from arcticpy.pixel_bounce import PixelBounce
 
 import autoarray as aa
 
@@ -39,7 +40,6 @@ class Clocker2D(AbstractClocker):
         serial_prune_frequency=0,
         serial_fast_mode: Optional[bool] = None,
         allow_negative_pixels=1,
-        pixel_bounce=None,
         verbosity: int = 0,
         poisson_seed: int = -1,
     ):
@@ -129,7 +129,6 @@ class Clocker2D(AbstractClocker):
         self.serial_fast_mode = serial_fast_mode
 
         self.allow_negative_pixels = allow_negative_pixels
-        self.pixel_bounce = pixel_bounce
 
         self.poisson_seed = poisson_seed
 
@@ -156,7 +155,7 @@ class Clocker2D(AbstractClocker):
         return trap_list, cti.serial_ccd
 
     def add_cti(
-        self, data: aa.Array2D, cti: CTI2D, preloads: Optional[Preloads] = Preloads()
+        self, data: aa.Array2D, cti: CTI2D, pixel_bounce_list: Optional[List[PixelBounce]] = None, preloads: Optional[Preloads] = Preloads()
     ) -> aa.Array2D:
         """
         Add CTI to a 2D dataset by passing it to the c++ arctic clocking algorithm.
@@ -232,7 +231,7 @@ class Clocker2D(AbstractClocker):
                 serial_prune_n_electrons=self.serial_prune_n_electrons,
                 serial_prune_frequency=self.serial_prune_frequency,
                 allow_negative_pixels=self.allow_negative_pixels,
-                pixel_bounce=self.pixel_bounce,
+                pixel_bounce_list=pixel_bounce_list,
                 verbosity=self.verbosity,
             )
         except TypeError:
@@ -260,7 +259,7 @@ class Clocker2D(AbstractClocker):
                 serial_time_stop=self.serial_time_stop,
                 serial_prune_n_electrons=self.serial_prune_n_electrons,
                 serial_prune_frequency=self.serial_prune_frequency,
-                pixel_bounce=self.pixel_bounce,
+                pixel_bounce_list=pixel_bounce_list,
                 verbosity=self.verbosity,
             )
 
@@ -271,7 +270,7 @@ class Clocker2D(AbstractClocker):
         except AttributeError:
             return image_post_cti
 
-    def add_cti_poisson_traps(self, data: aa.Array2D, cti: CTI2D) -> aa.Array2D:
+    def add_cti_poisson_traps(self, data: aa.Array2D, cti: CTI2D, pixel_bounce_list: Optional[List[PixelBounce]] = None) -> aa.Array2D:
         """
         Add CTI to a 2D dataset by passing it to the c++ arctic clocking algorithm.
 
@@ -337,7 +336,7 @@ class Clocker2D(AbstractClocker):
                     parallel_window_start=self.parallel_window_start,
                     parallel_window_stop=self.parallel_window_stop,
                     allow_negative_pixels=self.allow_negative_pixels,
-                    pixel_bounce=self.pixel_bounce,
+                    pixel_bounce_list=pixel_bounce_list,
                     verbosity=self.verbosity,
                 )[:, 0]
             except TypeError:
@@ -350,7 +349,7 @@ class Clocker2D(AbstractClocker):
                     parallel_window_offset=parallel_window_offset,
                     parallel_window_start=self.parallel_window_start,
                     parallel_window_stop=self.parallel_window_stop,
-                    pixel_bounce=self.pixel_bounce,
[…]
