+++ autocti/charge_injection/model/analysis.py
@@ -194,9 +194,19 @@ class AnalysisImagingCI(af.Analysis):
         if hyper_noise_scale and hasattr(instance, "hyper_noise"):
             hyper_noise_scalar_dict = instance.hyper_noise.as_dict
 
+        try:
+            self.clocker.pixel_bounce_list = instance.pixel_bounce_list
+        except AttributeError:
+            pass
+
         post_cti_data = self.clocker.add_cti(
             data=dataset.pre_cti_data, cti=instance.cti, preloads=self.preloads
         )
+
+        try:
+            self.clocker.pixel_bounce_list = None
+        except AttributeError:
+            pass
 
         return FitImagingCI(
             dataset=dataset,
