### 2023-12-06 5a65e2b
check_traps and check_ccd now also account ofr pixel bounce
