+++ autocti/clocker/abstract.py
@@ -2,6 +2,7 @@ from typing import List, Optional
 
 from arcticpy import CCD
 from arcticpy import CCDPhase
+from arcticpy import PixelBounce
 from arcticpy import TrapInstantCapture
 
 from autoconf.dictable import from_json, output_to_json
@@ -48,6 +49,7 @@ class AbstractClocker:
         self,
         trap_list_0: Optional[List[TrapInstantCapture]],
         trap_list_1: Optional[List[TrapInstantCapture]] = None,
+        pixel_bounce_list: Optional[List[PixelBounce]] = None,
     ):
         """
         Checks that there are trap species passed to the clocking algorithm and raises an exception if not.
@@ -57,12 +59,16 @@ class AbstractClocker:
         trap_list
             A list of the trap species on the CCD which capture and release electrons during clock to as to add CTI.
         """
+
+        if pixel_bounce_list is not None:
+            return
+
         if not any([trap_list_0, trap_list_1]):
             raise exc.ClockerException(
                 "No Trap species were passed to the add_cti method"
             )
 
-    def check_ccd(self, ccd_list: List[CCDPhase]):
+    def check_ccd(self, ccd_list: List[CCDPhase], pixel_bounce_list: Optional[List[PixelBounce]]):
         """
         Checks that there are trap species passed to the clocking algorithm and raises an exception if not.
 
@@ -72,6 +78,9 @@ class AbstractClocker:
             The ccd phase settings describing the volume-filling behaviour of the CCD which characterises the capture
             and release of electrons and therefore CTI.
         """
+        if pixel_bounce_list is not None:
+            return
+
         if not any(ccd_list):
             raise exc.ClockerException("No CCD object was passed to the add_cti method")
 
