### 2023-12-06 a6fb7ac
remove check_traps and check_ccd
