+++ autocti/clocker/one_d.py
@@ -102,9 +102,6 @@ class Clocker1D(AbstractClocker):
 
         trap_list, ccd = self._traps_ccd_from(cti=cti)
 
-        self.check_traps(trap_list_0=trap_list, pixel_bounce_list=pixel_bounce_list)
-        self.check_ccd(ccd_list=[ccd], pixel_bounce_list=pixel_bounce_list)
-
         image_pre_cti_2d = aa.Array2D.zeros(
             shape_native=(data.shape_native[0], 1), pixel_scales=data.pixel_scales
         ).native
@@ -176,9 +173,6 @@ class Clocker1D(AbstractClocker):
         """
 
         trap_list, ccd = self._traps_ccd_from(cti=cti)
-
-        self.check_traps(trap_list_0=trap_list, pixel_bounce_list=pixel_bounce_list)
-        self.check_ccd(ccd_list=[ccd], pixel_bounce_list=pixel_bounce_list)
 
         image_pre_cti_2d = aa.Array2D.zeros(
             shape_native=(data.shape_native[0], 1), pixel_scales=data.pixel_scales
