+++ autocti/clocker/abstract.py
@@ -45,45 +45,6 @@ class AbstractClocker:
         if ccd_phase is not None:
             return CCD(phases=[ccd_phase], fraction_of_traps_per_phase=[1.0])
 
-    def check_traps(
-        self,
-        trap_list_0: Optional[List[TrapInstantCapture]],
-        trap_list_1: Optional[List[TrapInstantCapture]] = None,
-        pixel_bounce_list: Optional[List[PixelBounce]] = None,
-    ):
-        """
-        Checks that there are trap species passed to the clocking algorithm and raises an exception if not.
-
-        Parameters
-        ----------
-        trap_list
-            A list of the trap species on the CCD which capture and release electrons during clock to as to add CTI.
-        """
-
-        if pixel_bounce_list is not None:
-            return
-
-        if not any([trap_list_0, trap_list_1]):
-            raise exc.ClockerException(
-                "No Trap species were passed to the add_cti method"
-            )
-
-    def check_ccd(self, ccd_list: List[CCDPhase], pixel_bounce_list: Optional[List[PixelBounce]]):
-        """
-        Checks that there are trap species passed to the clocking algorithm and raises an exception if not.
-
-        Parameters
-        ----------
-        ccd_list
-            The ccd phase settings describing the volume-filling behaviour of the CCD which characterises the capture
-            and release of electrons and therefore CTI.
-        """
-        if pixel_bounce_list is not None:
-            return
-
-        if not any(ccd_list):
-            raise exc.ClockerException("No CCD object was passed to the add_cti method")
-
     @classmethod
     def from_json(cls, file_path):
         return from_json(file_path=file_path)
