### 2023-12-11 eff53e6
remove ununsed functions in model_util
