### 2023-12-18 dbc3b35
calculation and output implemented
