+++ autocti/charge_injection/model/analysis.py
@@ -296,6 +296,48 @@ class AnalysisImagingCI(af.Analysis):
         if self.dataset_full is not None:
             output_dataset(dataset=self.dataset_full, prefix="dataset_full")
 
+    def save_results(self, paths: af.DirectoryPaths, result: ResultImagingCI):
+        """
+        At the end of a model-fit, this routine saves attributes of the `Analysis` object to the `files`
+        folder such that they can be loaded after the analysis using PyAutoFit's database and aggregator tools.
+
+        For this analysis it outputs the following:
+
+        - The Israel et al requirement on the spurious ellipticity based on the errors of the fit.
+
+        Parameters
+        ----------
+        paths
+            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
+            visualization and the pickled objects used by the aggregator output by this function.
+        result
+            The result of a model fit, including the non-linear search and samples.
+        """
+        delta_ellipticity_list = []
+
+        for sample in result.samples.sample_list:
+
+            instance = sample.instance_for_model(model=result.samples.model)
+
+            delta_ellipticity_list.append(instance.cti.delta_ellipticity)
+
+        (
+            median_delta_ellipticity,
+            upper_delta_ellipticity,
+            lower_delta_ellipticity,
+        ) = af.marginalize(
+            parameter_list=delta_ellipticity_list,
+            sigma=2.0,
+            weight_list=result.samples.weight_list,
+        )
+
+        delta_ellipticity = (upper_delta_ellipticity - lower_delta_ellipticity) / 2.0
+
+        output_to_json(
+            obj=delta_ellipticity,
+            file_path=paths._files_path / "delta_ellipticity.json",
+        )
+
     def in_ascending_fpr_order_from(self, quantity_list, fpr_value_list):
         if not conf.instance["visualize"]["general"]["general"][
             "subplot_ascending_fpr"
