### 2023-12-19 69713cf
refactor analysis classes to use AnalysisCTI
