+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-arcticpy==2.01.0
+arcticpy==2.6
 autoconf==2023.9.18.4
 numpydoc>=1.0.0
 pyprojroot==0.2.0
