### 2024-01-08 b0ccc1a
cast mask to Mask2D
