+++ autocti/extract/one_d/abstract.py
@@ -3,6 +3,7 @@ import numpy as np
 from typing import List, Optional, Tuple
 
 import autoarray as aa
+from autoarray import Mask1D
 
 from autocti.extract.settings import SettingsExtract
 
@@ -137,7 +138,7 @@ class Extract1D:
 
         return aa.Array1D(
             values=np.asarray(stacked_array_1d.data),
-            mask=sum(mask_1d_list) == len(mask_1d_list),
+            mask=mask_1d_list[0].with_new_array(sum(mask_1d_list) == len(mask_1d_list)),
         ).native
 
     def add_to_array(
