### 2024-01-08 80ccad5
cast back to mask after inversion
