### 2024-01-08 0f66602
remove bad usage of with_new_array
