+++ autocti/charge_injection/plot/fit_ci_plotters.py
@@ -419,14 +419,10 @@ class FitImagingCIPlotter(Plotter):
 
         if rows_fpr:
             data = copy.copy(self.dataset.data)
-            y = data.apply_mask(
-                mask=fpr_mask.with_new_array(np.invert(fpr_mask))
-            ).binned_across_rows
+            y = data.apply_mask(mask=fpr_mask.invert()).binned_across_rows
 
             model_data = copy.copy(self.fit.model_data)
-            y_extra = model_data.apply_mask(
-                mask=fpr_mask.with_new_array(np.invert(fpr_mask))
-            ).binned_across_rows
+            y_extra = model_data.apply_mask(mask=fpr_mask.invert()).binned_across_rows
 
             self.mat_plot_1d.plot_yx(
                 y=y,
@@ -465,13 +461,11 @@ class FitImagingCIPlotter(Plotter):
 
         if columns_fpr:
             data = copy.copy(self.dataset.data)
-            y = data.apply_mask(
-                mask=fpr_mask.with_new_array(np.invert(fpr_mask))
-            ).binned_across_columns
+            y = data.apply_mask(mask=fpr_mask.invert()).binned_across_columns
 
             model_data = copy.copy(self.fit.model_data)
             y_extra = model_data.apply_mask(
-                mask=fpr_mask.with_new_array(np.invert(fpr_mask))
+                mask=fpr_mask.invert()
             ).binned_across_columns
 
             self.mat_plot_1d.plot_yx(
