### 2024-01-08 4727b97
fix bad usage of with_new_array
