### 2024-01-26 c9835dc
implemened working function with unit test
