### 2024-01-26 b46867c
masking applied, uses sum of difference to give captured electrons
