+++ autocti/extract/two_d/parallel/fpr.py
@@ -98,21 +98,21 @@ class Extract2DParallelFPR(Extract2DParallel):
         """
         return extract_2d_util.binned_region_1d_fpr_from(pixels=settings.pixels)
 
-    def capture_estimate_from(self, array: aa.Array2D, pixels_from_start, pixels_from_end):
-
+    def capture_estimate_from(
+        self, array: aa.Array2D, pixels_from_start, pixels_from_end
+    ):
         capture_list = []
 
         for region in self.region_list:
-
             for x in range(region.x0, region.x1):
+                fpr = array.native[region.y0 : region.y1, x]
+                mask = np.invert(array.mask[region.y0 : region.y1, x])
 
-                fpr = array.native[region.y0:region.y1, x]
+                injection_estimate = np.median(fpr[-pixels_from_end:][mask[-pixels_from_end:]])
+                capture_estimate = np.sum(injection_estimate - fpr[0:pixels_from_start][mask[0:pixels_from_start]])
 
-                capture_list.append(np.median(fpr[-pixels_from_end:]) - np.median(fpr[0:pixels_from_start]))
-
-        print(capture_list)
+                capture_list.append(
+                    capture_estimate
+                )
 
         return np.median(capture_list)
-
-
-
