+++ autocti/extract/two_d/parallel/fpr.py
@@ -101,6 +101,36 @@ class Extract2DParallelFPR(Extract2DParallel):
     def capture_estimate_from(
         self, array: aa.Array2D, pixels_from_start, pixels_from_end
     ):
+        """
+        Estimates the total number of electrons captured in each column of the parallel FPRs.
+
+        This is performed by the following steps:
+
+        1) Go to each individual parallel FPR.
+        2) Estimate the injection level by taking the median of the final N pixels (N = `pixels_from_end`).
+        3) Subtracted this estimate from all N pixels at the start of the parallel FPR which will have had electrons
+           captured (N = `pixels_from_start`).
+        4) Sum the difference thereby estimating of the total number of electrons captured in the parallel FPR.
+        5) Return the median of all individual parallel FPR capture estimates.
+
+        This routine acts on every individual parallel FPR because charge injection non-uniformity means that using
+        averages over the full charge injection region will not be accurate.
+
+        The function also omits masked pixels, which are typically due to cosmic rays.
+
+        Parameters
+        ----------
+        array
+            The array contain the charge injection data whose parallel FPRs are used to estimate the capture.
+        pixels_from_start
+            The number of pixels at the start of the parallel FPR which are used to estimate the capture.
+        pixels_from_end
+            The number of pixels at the end of the parallel FPR which are used to estimate the injection level.
+
+        Returns
+        -------
+        An estimate of the total number of electrons captured in each column of the parallel FPRs.
+        """
         capture_list = []
 
         for region in self.region_list:
