### 2024-01-26 3bfcca9
test fully masked data does not break code
