### 2024-01-26 8b73f25
numba speedup
