### 2024-01-27 7a392eb
omit numba-scipy from build
