+++ .github/workflows/main.yml
@@ -66,7 +66,7 @@ jobs:
         pip3 install -r PyAutoConf/requirements.txt
         pip3 install -r PyAutoFit/requirements.txt
         pip3 install -r PyAutoArray/requirements.txt
-        pip3 install -r PyAutoArray/optional_requirements.txt
+        pip3 install -r PyAutoArray/optional_requirements.txt --no-deps numba-scipy
         pip3 install -r PyAutoCTI/requirements.txt
         pip3 install -r PyAutoCTI/optional_requirements.txt
     - name: Run tests
