### 2024-01-29 bd92881
feature/jax (#64)
Updated autoarray to support JAX
