### 2024-01-31 #4904 @Jammy2211
To my knowledge, the installation went fine in the end (https://github.com/Jammy2211/PyAutoCTI/issues/53).

Any updates would be good!
