### 2024-02-15 ecdc184
derived quantity on delta_Ellipticity
