### 2024-02-17 e50ba1e
fix delta ellipticity tests
