+++ autocti/fixtures.py
@@ -224,7 +224,7 @@ def make_samples_with_result():
     return MockSamples(
         model=model,
         max_log_likelihood_instance=instance,
-        gaussian_tuples=[(1.0, 2.0)] * model.prior_count,
+        prior_means=[1.0] * model.prior_count,
     )
 
 
