### 2024-02-19 85843e4
stray light simulated
