### 2024-02-20 #4904 @mwcraig
Apologies @Jammy2211 for the long delay.

I want to start by recognizing how excellent the example notebooks and documentation are for this project. I went through all of the overview notebooks, which are also the bulk of the narrative documentation. Although I found a number of minor typos/inconsistencies, none of them are substantive enough to hold up publication of this paper.

I very enthusiastically recommend that this paper be published in JOSS and apologize again for the long review delay.

I'll open up some issues in the `PyAutoCTI` repo regarding the notebook issues (but note that these do *not* need to be addressed before publication).
