### 2024-02-20 Jammy2211/PyAutoCTI#68 @mwcraig
Images are inverted in documentation on ReadTheDocs compared to the `overview` notebooks in the default workspace
I noticed while reviewing this for your JOSS submission that the orientation of the image plots in the documentation is inverted compared to those in the default workspace. It looks like in one case `origin=lower` in `imshow` and in the other that is not set. The orientation of the images in the documentation is incorrect I think.

For example, in the documentation page "Parallel and Serial" charge is supposed to be moving towards the top of the CCD but the figure in the the section [CTI Model (Parallel)](https://pyautocti.readthedocs.io/en/latest/overview/overview_2_parallel_and_serial.html#cti-model-parallel) the charge is moving towards the bottom of the plot.

That same inversion in all of the notebooks with 2D images.
