### 2024-02-20 Jammy2211/PyAutoCTI#69 @mwcraig
Suggestions for welcome.py script
I really like the `welcome.py` script! A couple of minor suggestions for it:

1. Add a prompt that you need to close the matplotlib window for the script to complete.
2. The numba warning ends with a line that starts "Install numba as described at the following webpage..." but there is no following web page.
