### 2024-02-20 Jammy2211/PyAutoCTI#70 @mwcraig
Minor issues in a few overview notebooks and/or corresponding documentation
There were a handful of fairly minor issues that you might want to clean up:

1. Consider numbering the folders in the `notebooks` folder of the default workspace. The `README` is clear that one should start with the overview, so this may not be necessary.
2. The CCD gif in the middle of `notebooks/overview_1_what_is_cti` doesn't load because the link is broken. The broken link is: https://github.com/Jammy2211/autocti_workspace/blob/main/dataset/ccd.gif
3. In `notebooks/overview_2_parallel_and_serial` there is a missing `'`  in item 3 of the section `__Two Dimensional CCD__`.
4. The "Wrap Up" section in `notebooks/overview_3_cti_features` is blank in the overview notebook and does not appear in the corresponding documentation page.
5. In `notebooks/overview_4_charge_injection_data` there is this line, with a question mark at the end that is meant to point somewhere instead: `**PyAutoCTI** has built in tools for both these tasks which are illustrated at ?.`
6. In `notebooks/overview_5_fitting` there is a plot at the very end of the section "Dataset (Charge Injection)". That plot does not  appear in the documentation at the end of that section: https://pyautocti.readthedocs.io/en/latest/overview/overview_5_fitting.html#dataset-charge-injection
7. In `notebooks/overview_5_fitting` there is text printed out in the cell that illustrates the "Good Fit" that should be in a markdown cell below the plots instead. This text begins "If we compare the `log_likelihood` to the value above...". In the documentation this has been fixed.
8. In `notebooks/overview_6_cti_calibration` the link to `PyAutoFit` doesn't work because it is written rST style instead of markdown style.
9. In notebooks 5 and 6 I get this warning in the first cell:

```python-traceback
Users/***/mambaforge/envs/autocti/lib/python3.12/site-packages/IPython/core/magics/osm.py:417: UserWarning: using dhist requires you to install the `pickleshare` library.
  self.shell.db['dhist'] = compress_dhist(dhist)[-100:]
```
