### 2024-02-20 Jammy2211/PyAutoCTI#71 @mwcraig
Suggestions for overview notebook 4
I have a few suggestions for `notebooks/overview_4_charge_injection_data` in the default workspace:

- Consider adding a banner at or near the top of the notebook that installing `numba` will make this one run much faster (at least 60x faster in my case, maybe more). I had ignored the numba warnings up to this point because it didn't seem to make much difference until I got to this notebook.
- Add a cell with the commands to run each of the three scripts that need to be run, perhaps:

```python
!python scripts/imaging_ci/simulators/overview/uniform.py
!python scripts/imaging_ci/simulators/overview/non_uniform_cosmic_rays.py
!python scripts/imaging_ci/simulators/overview/calibrate.py
```

(this is the last issue 😀 )
