### 2024-02-21 Jammy2211/PyAutoCTI#69 @Jammy2211
The welcome.py message has been updated:

https://github.com/Jammy2211/autocti_workspace/blob/main/welcome.py

```
    NOTE: If a matplotlib figure window appears, you may need to close it via the X button and then press
    enter to continue the script.
```

I have updated the exception on numba to point to this webpage: https://pyautolens.readthedocs.io/en/latest/installation/numba.html

Thank you!
