### 2024-02-21 Jammy2211/PyAutoCTI#68 @Jammy2211
All images have been updated to use the flipped y axis and other improvements to visuals :)
