### 2024-02-21 Jammy2211/PyAutoCTI#70 @mwcraig
> > Users/***/mambaforge/envs/autocti/lib/python3.12/site-packages/IPython/core/magics/osm.py:417: UserWarning: using dhist requires you to install the pickleshare library.
self.shell.db['dhist'] = compress_dhist(dhist)[-100:]

> Genuinely no idea whats going on here...

It isn't a big deal, the notebook still runs fine
