### 2024-02-21 #4904 @Jammy2211
Review done :):

https://github.com/Jammy2211/PyAutoCTI/issues/71
https://github.com/Jammy2211/PyAutoCTI/issues/53
https://github.com/Jammy2211/PyAutoCTI/issues/68
https://github.com/Jammy2211/PyAutoCTI/issues/69
https://github.com/Jammy2211/PyAutoCTI/issues/70
