### 2024-02-21 5a05700
flat field mode hack tested
