### 2024-02-23 3373644
flat field mode computes noise mao
