### 2024-02-23 a5bc1e4
improve layout exception error
