### 2024-03-03 22c261f
aggregator now uses autofit inheritance
