### 2024-03-04 #4904 @adonath
Thanks @Jammy2211, here are some final comments on the paper from my side:

- The paper reads well and is very informative on the software
- Merge the double citations into a single parentheses separated by a semicolon, like `[@Lomb1976; @Scargle1982]`
- Unify Fig and Figure (line 48/49)
- I would suggest to remove the list of software citations and add the citations to the "Software API and Features" and the citations of "Related Software" into the "Statement of Need". So put the citation, where it is mentioned in the text. If the software is not mentioned, just add a half sentence describing got what it is used.

Once those are addressed we can proceed with publishing.

Thanks @mwcraig and @jryon for contributing and finishing the reviews!
