+++ autocti/plot/__init__.py
@@ -1,10 +1,6 @@
-from autofit.plot.samples_plotters import SamplesPlotter
-from autofit.non_linear.search.nest.dynesty.plotter import DynestyPlotter
-from autofit.non_linear.search.nest.nautilus.plotter import NautilusPlotter
-from autofit.non_linear.search.nest.ultranest.plotter import UltraNestPlotter
-from autofit.non_linear.search.mcmc.emcee.plotter import EmceePlotter
-from autofit.non_linear.search.mcmc.zeus.plotter import ZeusPlotter
-from autofit.non_linear.search.optimize.pyswarms.plotter import PySwarmsPlotter
+from autofit.non_linear.plot.nest_plotters import NestPlotter
+from autofit.non_linear.plot.mcmc_plotters import MCMCPlotter
+from autofit.non_linear.plot.optimize_plotters import OptimizePlotter
 
 from autoarray.plot.wrap.base import Axis
 from autoarray.plot.wrap.base import Units
