+++ autocti/charge_injection/imaging/simulator.py
@@ -26,8 +26,8 @@ class SimulatorImagingCI(SimulatorImaging):
         non_uniform_norm_limit=None,
         read_noise: Optional[float] = None,
         charge_noise: Optional[float] = None,
-        stray_light : Optional[Tuple[float, float]] = None,
-        flat_field_mode : bool = False,
+        stray_light: Optional[Tuple[float, float]] = None,
+        flat_field_mode: bool = False,
         noise_if_add_noise_false: float = 0.1,
         noise_seed: int = -1,
         ci_seed: int = -1,
@@ -181,7 +181,6 @@ class SimulatorImagingCI(SimulatorImaging):
         """
 
         if self.flat_field_mode:
-
             return self.via_flat_field_mode(
                 layout=layout,
                 clocker=clocker,
@@ -216,17 +215,13 @@ class SimulatorImagingCI(SimulatorImaging):
             pre_cti_data += cosmic_ray_map.native
 
         if self.stray_light is not None:
-
             m, c = self.stray_light
 
             region = layout.extract.parallel_overscan.parallel_overscan
 
-            stray_light = np.array(
-                [m * i + c for i in range(0, region.y1)]
-            )
+            stray_light = np.array([m * i + c for i in range(0, region.y1)])
 
-            pre_cti_data[0: region.y1, region.x0: region.x1] += stray_light[:, None]
-
+            pre_cti_data[0 : region.y1, region.x0 : region.x1] += stray_light[:, None]
 
         if self.charge_noise is not None:
             pre_cti_data = layout.extract.parallel_fpr.add_gaussian_noise_to(
@@ -298,17 +293,14 @@ class SimulatorImagingCI(SimulatorImaging):
         )
 
     def via_flat_field_mode(
-            self,
-            layout: Layout2DCI,
-            clocker: Optional[Clocker2D],
-            cti: Optional[CTI2D],
-            cosmic_ray_map: Optional[aa.Array2D] = None,
+        self,
+        layout: Layout2DCI,
+        clocker: Optional[Clocker2D],
+        cti: Optional[CTI2D],
+        cosmic_ray_map: Optional[aa.Array2D] = None,
     ):
-
         pre_cti_data = np.zeros(layout.shape_2d)
-        pre_cti_data[
-            layout.region_list[0].slice
-        ] = self.norm
+        pre_cti_data[layout.region_list[0].slice] = self.norm
 
         pre_cti_data_poisson = np.random.poisson(pre_cti_data, pre_cti_data.shape)
 
@@ -336,20 +328,23 @@ class SimulatorImagingCI(SimulatorImaging):
             layout=layout,
             cosmic_ray_map=cosmic_ray_map,
         )
-        
+
         if self.read_noise is None:
             read_noise = 0.0
         else:
             read_noise = self.read_noise
 
         noise_map = (
-                read_noise
-                * aa.Array2D.ones(
-            shape_native=layout.shape_2d, pixel_scales=self.pixel_scales
-        ).native
+            read_noise
+            * aa.Array2D.ones(
+                shape_native=layout.shape_2d, pixel_scales=self.pixel_scales
+            ).native
         )
 
-        noise_map[layout.region_list[0].slice] = np.sqrt(dataset.data[layout.region_list[0].slice] + noise_map[layout.region_list[0].slice])
+        noise_map[layout.region_list[0].slice] = np.sqrt(
+            dataset.data[layout.region_list[0].slice]
+            + noise_map[layout.region_list[0].slice]
+        )
 
         dataset.noise_map = noise_map
 
