### 2024-03-20 1964384
remove derived quantity
