+++ autocti/model/model_util.py
@@ -1,7 +1,5 @@
 from typing import Optional, List
 
-import autofit as af
-
 from autoconf.dictable import from_json
 
 from arcticpy import CCDPhase
@@ -18,6 +16,7 @@ class AbstractCTI:
     def trap_all_list(self):
         raise NotImplementedError
 
+    @property
     def delta_ellipticity(self):
         return sum([trap.delta_ellipticity for trap in self.trap_all_list])
 
