### 2024-03-26 cb967cd
fix make_result
