+++ test_autocti/analysis/test_result.py
@@ -7,10 +7,10 @@ import pytest
 
 
 def test__result_contains_instance_with_cti_model(
-    analysis_imaging_ci_7x7, samples_with_result
+    analysis_imaging_ci_7x7, samples_summary_with_result
 ):
     result = res.Result(
-        samples=samples_with_result,
+        samples=samples_summary_with_result,
         analysis=analysis_imaging_ci_7x7,
     )
 
@@ -19,10 +19,10 @@ def test__result_contains_instance_with_cti_model(
 
 
 def test__clocker_passed_as_result_correctly(
-    analysis_imaging_ci_7x7, samples_with_result, parallel_clocker_2d
+    analysis_imaging_ci_7x7, samples_summary_with_result, parallel_clocker_2d
 ):
     result = res.Result(
-        samples=samples_with_result,
+        samples=samples_summary_with_result,
         analysis=analysis_imaging_ci_7x7,
     )
 
@@ -32,13 +32,13 @@ def test__clocker_passed_as_result_correctly(
 
 def test__masks_available_as_property(
     analysis_imaging_ci_7x7,
-    samples_with_result,
+    samples_summary_with_result,
     parallel_clocker_2d,
     traps_x1,
     ccd,
 ):
     result = res.ResultDataset(
-        samples=samples_with_result,
+        samples=samples_summary_with_result,
         analysis=analysis_imaging_ci_7x7,
     )
 
