### 2024-03-29 af5b566
made ReadoutPersistence class with functon to add
