+++ autocti/charge_injection/imaging/simulator.py
@@ -7,6 +7,7 @@ import autoarray as aa
 from autoarray.dataset.imaging.simulator import SimulatorImaging
 
 from autocti.charge_injection.imaging.imaging import ImagingCI
+from autocti.charge_injection.imaging.readout_persistence import ReadoutPersistence
 from autocti.charge_injection.layout import Layout2DCI
 from autocti.clocker.two_d import Clocker2D
 from autocti.extract.settings import SettingsExtract
@@ -27,6 +28,7 @@ class SimulatorImagingCI(SimulatorImaging):
         read_noise: Optional[float] = None,
         charge_noise: Optional[float] = None,
         stray_light: Optional[Tuple[float, float]] = None,
+        readout_persistance : Optional[ReadoutPersistence] = None,
         flat_field_mode: bool = False,
         noise_if_add_noise_false: float = 0.1,
         noise_seed: int = -1,
@@ -57,6 +59,7 @@ class SimulatorImagingCI(SimulatorImaging):
         self.read_noise = read_noise
         self.charge_noise = charge_noise
         self.stray_light = stray_light
+        self.readout_persistance =  readout_persistance
         self.flat_field_mode = flat_field_mode
 
         self.ci_seed = ci_seed
@@ -237,6 +240,12 @@ class SimulatorImagingCI(SimulatorImaging):
             post_cti_data = clocker.add_cti(data=pre_cti_data, cti=cti)
         else:
             post_cti_data = copy.copy(pre_cti_data)
+
+        if self.readout_persistance is not None:
+
+            post_cti_data = self.readout_persistance.data_with_readout_persistence_from(
+                data=post_cti_data
+            )
 
         if cosmic_ray_map is not None:
             pre_cti_data -= cosmic_ray_map.native
