+++ paper/paper.md
@@ -33,18 +33,18 @@ bibliography: paper.bib
 # Summary
 
 Over the past half a century, space telescopes have given us an extraordinary view of the Universe, for example detecting 
-the faint spectra of the first galaxies in the Universe [@Dunlop2013] [@Bouwens2015], measuring tiny distortions in
-galaxy shapes due to gravitational lensing [@Massey2007] [@Schrabback2010] and a high precision map of stars within
-the Milky Way [@Brown2018] [@Brown2020]. These observations require a deep understanding of the telescope's
+the faint spectra of the first galaxies in the Universe [@Dunlop2013; @Bouwens2015], measuring tiny distortions in
+galaxy shapes due to gravitational lensing [@Massey2007; @Schrabback2010] and a high precision map of stars within
+the Milky Way [@Brown2018; @Brown2020]. These observations require a deep understanding of the telescope's
 instrumental characteristics, including the calibration and correction of charge transfer inefficiency 
-(hereafter CTI) [@Massey2010d] [@Massey2014], a phenomena where radiation damage to the telescope's charge-coupled 
+(hereafter CTI) [@Massey2010d; @Massey2014], a phenomena where radiation damage to the telescope's charge-coupled 
 device (CCD) sensors leads to gradually increased smearing in acquired exposures over the telescope's lifetime.
 
-`PyAutoCTI` is an open-source Python 3.8 - 3.11 package for the calibration of CTI for space telescopes. By interfacing
+`PyAutoCTI` is an open-source Python 3.8 - 3.11 [@python] package for the calibration of CTI for space telescopes. By interfacing
 with `arCTIc` (`the algorithm for Charge Transfer Inefficiency correction`) the calibrated CTI models can straightforwardly
 be used to correct and remove CTI in every science image taken throughout the telescope's lifetime. Core features 
 include fully automated Bayesian model-fitting of CTI calibration data, support for different calibration
-datasets such as warm pixels used for Hubble Space Telescope calibration [@Massey2010d] [@Massey2010b] and a database 
+datasets such as warm pixels used for Hubble Space Telescope calibration [@Massey2010d; @Massey2010b] and a database 
 for building a temporal model of CTI over the telescope's lifetime. The software places a focus on big data analysis, 
 including support for graphical models that simultaneously fits large CTI calibration datasets and an SQLite3 
 database that allows extensive suites of calibration results to be loaded, queried and analysed. 
@@ -76,7 +76,7 @@ with the author's permission.](cti_image.png)
 Figure 1 shows an example of the CTI calibration data which `PyAutoCTI` was originally developed to fit -- warm 
 pixels in the Hubble Space Telescope [@Massey2010d]. Warm pixels are similar to short circuits in the CCD electrostatic 
 potentials used to collect charge, which continuously inject spurious charge into specific pixels. These should 
-appear as delta functions in the data, with CTI trails appearing in Fig 1 in the parallel clocking direction (upwards in 
+appear as delta functions in the data, with CTI trails appearing in Figure 1 in the parallel clocking direction (upwards in 
 Figure 1) next to each warm pixel. `PyAutoCTI` has dedicated functionality for fitting 1D calibration datasets 
 such as those extracted from warm pixels. `PyAutoCTI` also supports for 2D calibration datasets, for example 
 charge injection line imaging calibration data, which will be used to calibrate CTI for the European Space Agency's 
[…]
