### 2024-03-29 #4904 @adonath
@Jammy2211 I double checked the DOIs and it seems indeed there are no DOIs for the listed references. Can you confirm?
