+++ paper/paper.bib
@@ -35,6 +35,7 @@ abstract = {Linear operators and optimisation are at the core of many algorithms
 archivePrefix = {arXiv},
 arxivId = {1907.12349},
 author = {Ravasi, Matteo and Vasconcelos, Ivan},
+doi = {10.1016/j.softx.2019.100361},
 eprint = {1907.12349},
 file = {:home/jammy/Documents/Papers/Software/PyLops.pdf:pdf},
 title = {{PyLops -- A Linear-Operator Python Library for large scale optimization}},
@@ -376,18 +377,6 @@ url = {http://arxiv.org/abs/1801.01831},
 volume = {475},
 year = {2018}
 }
-@article{Hermans2019,
-abstract = {Posterior inference with an intractable likelihood is becoming an increasingly common task in scientific domains which rely on sophisticated computer simulations. Typically, these mechanistic models do not admit tractable densities forcing practitioners to rely on approximations during inference. This work proposes a novel approach to address the intractability of the likelihood and the marginal model. We achieve this by learning a flexible estimator which approximates the likelihood-to-evidence ratio. The resulting amortized ratio estimator is embedded in MCMC samplers such as Metropolis-Hastings and Hamiltonian Monte Carlo to approximate the likelihood-ratio between consecutive states in the Markov chain, allowing us to draw samples from the intractable posterior. Techniques are presented to improve the numerical stability. We demonstrate our approach on a variety of benchmarks and compare against well-established approximate inference techniques. Scientific applications in high energy and astrophysics with high-dimensional observations show its applicability.},
-archivePrefix = {arXiv},
-arxivId = {1903.04057},
-author = {Hermans, Joeri and Begy, Volodimir and Louppe, Gilles},
-eprint = {1903.04057},
-file = {:home/jammy/Documents/Papers/PyAutoLens/Hermanns2019LikelihoodFree.pdf:pdf},
-number = {i},
-title = {{Likelihood-free MCMC with Amortized Approximate Likelihood Ratios}},
-url = {http://arxiv.org/abs/1903.04057},
-year = {2019}
-}
 @article{Koopmans2009,
[…]
