### 2024-03-29 #4904 @Jammy2211
I have added a few DOIs, but the majority which are missing are because they dont exist (e.g. manuals).
