+++ autocti/__init__.py
@@ -54,7 +54,6 @@ from .extract.one_d.fpr import Extract1DFPR
 from .extract.one_d.eper import Extract1DEPER
 from .extract.one_d.master import Extract1DMaster
 from .layout.one_d import Layout1D
-from .dataset_1d.dataset_1d.settings import SettingsDataset1D
 from .dataset_1d.dataset_1d.dataset_1d import Dataset1D
 from .dataset_1d.dataset_1d.simulator import SimulatorDataset1D
 from .dataset_1d.fit import FitDataset1D
