+++ autocti/charge_injection/model/analysis.py
@@ -25,7 +25,6 @@ logger.setLevel(level="INFO")
 
 
 class AnalysisImagingCI(AnalysisCTI):
-
     Result = ResultImagingCI
 
     def __init__(
