+++ autocti/charge_injection/model/plotter_interface.py
@@ -1,9 +1,12 @@
 import logging
 
+import autoarray.plot as aplt
+
 import autocti.plot as aplt
 
-from autocti.model.visualizer import Visualizer
+from autocti.model.visualizer import PlotterInterface
 from autocti.model.visualizer import plot_setting
+
 from autocti import exc
 
 logger = logging.getLogger(__name__)
@@ -11,42 +14,42 @@ logger = logging.getLogger(__name__)
 logger.setLevel(level="INFO")
 
 
-class VisualizerDataset1D(Visualizer):
-    def visualize_dataset(self, dataset, folder_suffix=""):
+class PlotterInterfaceImagingCI(PlotterInterface):
+    def dataset(self, dataset, folder_suffix: str = ""):
         def should_plot(name):
             return plot_setting(section="dataset", name=name)
 
-        mat_plot_1d = self.mat_plot_1d_from(subfolders=f"dataset{folder_suffix}")
+        mat_plot = self.mat_plot_2d_from(subfolders=f"dataset{folder_suffix}")
 
-        dataset_plotter = aplt.Dataset1DPlotter(
-            dataset=dataset, mat_plot_1d=mat_plot_1d, include_1d=self.include_1d
+        dataset_plotter = aplt.ImagingCIPlotter(
+            dataset=dataset, mat_plot_2d=mat_plot, include_2d=self.include_2d
         )
 
         if should_plot("subplot_dataset"):
             dataset_plotter.subplot_dataset()
 
-        dataset_plotter.figures_1d(
+        dataset_plotter.figures_2d(
             data=should_plot("data"),
-            data_logy=should_plot("data_logy"),
             noise_map=should_plot("noise_map"),
             signal_to_noise_map=should_plot("signal_to_noise_map"),
             pre_cti_data=should_plot("pre_cti_data"),
+            cosmic_ray_map=should_plot("cosmic_ray_map"),
         )
 
-    def visualize_dataset_regions(self, dataset, region_list, folder_suffix=""):
+    def dataset_regions(self, dataset, region_list, folder_suffix: str = ""):
         def should_plot(name):
             return plot_setting(section="dataset", name=name)
 
-        mat_plot_1d = self.mat_plot_1d_from(subfolders=f"dataset{folder_suffix}")
+        mat_plot = self.mat_plot_1d_from(subfolders=f"dataset{folder_suffix}")
 
-        dataset_plotter = aplt.Dataset1DPlotter(
-            dataset=dataset, mat_plot_1d=mat_plot_1d, include_1d=self.include_1d
+        dataset_plotter = aplt.ImagingCIPlotter(
+            dataset=dataset, mat_plot_1d=mat_plot, include_2d=self.include_2d
         )
 
         for region in region_list:
             try:
                 if should_plot("subplot_dataset"):
-                    dataset_plotter.subplot_dataset(region=region)
+                    dataset_plotter.subplot_1d(region=region)
 
                 dataset_plotter.figures_1d(
                     region=region,
@@ -59,37 +62,74 @@ class VisualizerDataset1D(Visualizer):
 
             except (exc.RegionException, TypeError, ValueError):
                 logger.info(
-                    f"VISUALIZATION - Could not visualize the Dataset1D {region}"
+                    f"VISUALIZATION - Could not visualize the ImagingCI 1D {region}"
                 )
 
-    def visualize_dataset_combined(self, dataset_list, folder_suffix=""):
+    def dataset_combined(
+        self, dataset_list, folder_suffix: str = "", filename_suffix: str = ""
+    ):
         def should_plot(name):
             return plot_setting(section="dataset", name=name)
 
[…]

[…]
