+++ autocti/dataset_1d/model/plotter_interface.py
@@ -2,7 +2,7 @@ import logging
 
 import autocti.plot as aplt
 
-from autocti.model.visualizer import PlotterInterface
+from autocti.model.plotter_interface import PlotterInterface
 from autocti.model.visualizer import plot_setting
 from autocti import exc
 
