+++ autocti/dataset_1d/model/plotter_interface.py
@@ -3,7 +3,7 @@ import logging
 import autocti.plot as aplt
 
 from autocti.model.plotter_interface import PlotterInterface
-from autocti.model.visualizer import plot_setting
+from autocti.model.plotter_interface import plot_setting
 from autocti import exc
 
 logger = logging.getLogger(__name__)
