+++ autocti/charge_injection/model/plotter_interface.py
@@ -5,7 +5,7 @@ import autoarray.plot as aplt
 import autocti.plot as aplt
 
 from autocti.model.plotter_interface import PlotterInterface
-from autocti.model.visualizer import plot_setting
+from autocti.model.plotter_interface import plot_setting
 
 from autocti import exc
 
