+++ autocti/charge_injection/model/analysis.py
@@ -9,7 +9,7 @@ import autofit as af
 
 from autocti.charge_injection.imaging.imaging import ImagingCI
 from autocti.charge_injection.fit import FitImagingCI
-from autocti.charge_injection.model.plotter_interface import PlotterInterfaceImagingCI
+from autocti.charge_injection.model.visualizer import VisualizerImagingCI
 from autocti.charge_injection.model.result import ResultImagingCI
 from autocti.clocker.two_d import Clocker2D
 from autocti.charge_injection.hyper import HyperCINoiseCollection
@@ -26,6 +26,7 @@ logger.setLevel(level="INFO")
 
 class AnalysisImagingCI(AnalysisCTI):
     Result = ResultImagingCI
+    Visualizer = VisualizerImagingCI
 
     def __init__(
         self,
@@ -297,170 +298,3 @@ class AnalysisImagingCI(AnalysisCTI):
 
         if self.dataset_full is not None:
             output_dataset(dataset=self.dataset_full, prefix="dataset_full")
-
-    def visualize_before_fit(self, paths: af.DirectoryPaths, model: af.Collection):
-        if conf.instance["visualize"]["plots"]["combined_only"]:
-            return
-
-        visualizer = PlotterInterfaceImagingCI(image_path=paths.image_path)
-
-        region_list = self.region_list_from(model=model)
-
-        if conf.instance["visualize"]["plots"]["dataset"]["fpr_non_uniformity"]:
-            region_list += ["fpr_non_uniformity"]
-
-        visualizer.dataset(dataset=self.dataset)
-        visualizer.dataset_regions(
-            dataset=self.dataset, region_list=region_list
-        )
-
-        if self.dataset_full is not None:
-            visualizer.dataset(
-                dataset=self.dataset_full, folder_suffix="_full"
-            )
-            visualizer.dataset_regions(
-                dataset=self.dataset_full,
-                region_list=region_list,
-                folder_suffix="_full",
-            )
-
-    def visualize_before_fit_combined(
-        self, analyses, paths: af.DirectoryPaths, model: af.Collection
-    ):
-        if analyses is None:
-            return
-
-        visualizer = PlotterInterfaceImagingCI(image_path=paths.image_path)
-
-        region_list = self.region_list_from(model=model)
-
-        if conf.instance["visualize"]["plots"]["dataset"]["fpr_non_uniformity"]:
-            region_list += ["fpr_non_uniformity"]
-
-        dataset_list = [analysis.dataset for analysis in analyses]
-        fpr_value_list = [dataset.fpr_value for dataset in dataset_list]
-
-        dataset_list = self.in_ascending_fpr_order_from(
-            quantity_list=dataset_list,
-            fpr_value_list=fpr_value_list,
-        )
-
-        visualizer.dataset_combined(
-            dataset_list=dataset_list,
-        )
-
-        visualizer.dataset_regions_combined(
-            dataset_list=dataset_list,
-            region_list=region_list,
-        )
-
-        if self.dataset_full is not None:
-            dataset_full_list = [analysis.dataset_full for analysis in analyses]
-
-            dataset_full_list = self.in_ascending_fpr_order_from(
-                quantity_list=dataset_full_list,
-                fpr_value_list=fpr_value_list,
-            )
-
-            visualizer.dataset_combined(
-                dataset_list=dataset_full_list,
-                folder_suffix="_full",
-                filename_suffix="_full",
-            )
-            visualizer.dataset_regions_combined(
-                dataset_list=dataset_full_list,
-                region_list=region_list,
-                folder_suffix="_full",
-                filename_suffix="_full",
-            )
-
-    def visualize(
-        self,
-        paths: af.DirectoryPaths,
-        instance: af.ModelInstance,
-        during_analysis: bool,
-    ):
[…]
