+++ autocti/__init__.py
@@ -78,4 +78,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2024.1.27.4"
+__version__ = "2024.5.16.1"
