### 2024-05-16 #4904 @Jammy2211
Version: `2024.5.16.0`

Release: https://github.com/Jammy2211/PyAutoCTI/releases/tag/2024.5.16.0
Zenodo: https://zenodo.org/records/11203409
DOI: 10.5281/zenodo.11203409
