### 2024-05-17 #4904 @Jammy2211
HAha, proofs look good to me!
