### 2024-05-28 #4904 @Jammy2211
10.5281/zenodo.11367432

https://github.com/Jammy2211/PyAutoCTI/releases/tag/joss
